TPB_1SO

A simulated dataset based on the Theory of Planned Behaviour, where INT is a higher order construct of ATT, SN, and PBC.

Examples

tpb <- '
  # First order constructs
  ATT =~ att1 + att2 + att3
  SN  =~ sn1 + sn2 + sn3
  PBC =~ pbc1 + pbc2 + pbc3
  BEH =~ b1 + b2

  # Higher order constructs
  INT =~ ATT + PBC + SN

  # Higher order interaction
  INTxPBC =~ ATT:PBC + SN:PBC + PBC:PBC
  
  # Structural model
  BEH ~ PBC + INT + INTxPBC
'

# \dontrun{
est_ca <- modsem(tpb, data = TPB_1SO, method = "ca")
summary(est_ca)
#> Estimating baseline model (H0)
#> Warning: Unable to estimate baseline model for models with higher-order interaction terms!
#> modsem (version 1.0.11, approach = ca):
#> 
#> Interaction Model Fit Measures (H1):
#>   Loglikelihood                              -42117.62 
#>   Akaike (AIC)                                84339.24 
#>   Bayesian (BIC)                              84630.48 
#>   Chi-square                                   4246.90 
#>   Degrees of Freedom                               178 
#>   P-value (Chi-square)                           0.000 
#>   RMSEA                                          0.107 
#>   CFI                                            0.890 
#>   SRMR                                           0.135 
#> 
#> R-Squared Interaction Model (H1):
#>   BEH                                            0.232 
#> 
#> lavaan 0.6-19 ended normally after 445 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        73
#>   Row rank of the constraints matrix                21
#> 
#>   Number of observations                          2000
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                              4246.901
#>   Degrees of freedom                               178
#>   P-value (Chi-square)                           0.000
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   ATT =~                                              
#>     att1   (l_1_A)    1.000                           
#>     att2   (l_2_A)    0.904    0.010   89.441    0.000
#>     att3   (l_3_A)    0.801    0.009   85.442    0.000
#>   SN =~                                               
#>     sn1    (l_1_S)    1.000                           
#>     sn2    (l_2_S)    0.879    0.013   66.744    0.000
#>     sn3    (l_3_S)    0.780    0.012   63.639    0.000
#>   PBC =~                                              
#>     pbc1   (l_1_P)    1.000                           
#>     pbc2   (l_2_P)    0.900    0.007  135.630    0.000
#>     pbc3   (l_3_P)    0.776    0.006  125.110    0.000
#>   BEH =~                                              
#>     b1     (l_1_B)    1.000                           
#>     b2     (l_2_B)    0.863    0.033   26.043    0.000
#>   INT =~                                              
#>     ATT   (l_ATT_)    1.000                           
#>     PBC   (l_PBC_)    0.783    0.030   26.191    0.000
#>     SN     (l_SN_)    0.717    0.027   26.257    0.000
#>   INTxPBC =~                                          
#>     ATTPB (l_ATTP)    1.000                           
#>     SNPBC  (l_SNP)    0.735    0.020   35.914    0.000
#>     PBCPB (l_PBCP)    1.011    0.027   36.926    0.000
#>   ATTPBC =~                                           
#>     att11 (l_11_A)    1.000                           
#>     att22 (l_22_A)    0.813    0.009   87.006    0.000
#>     att33 (l_33_A)    0.621    0.008   82.373    0.000
#>   SNPBC =~                                            
#>     sn1p1 (l_11_S)    1.000                           
#>     sn2p2 (l_22_S)    0.792    0.012   68.052    0.000
#>     sn3p3 (l_33_S)    0.605    0.009   64.722    0.000
#>   PBCPBC =~                                           
#>     pbc11 (l_11_P)    1.000                           
#>     pbc22 (l_22_P)    0.810    0.012   67.815    0.000
#>     pbc33 (l_33_P)    0.602    0.010   62.555    0.000
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   BEH ~                                               
#>     PBC     (G_PB)    0.249    0.052    4.775    0.000
#>     INT   (G_INT_)    0.160    0.056    2.838    0.005
#>     INTPB (G_INTP)    0.221    0.016   13.492    0.000
#> 
#> Covariances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   INT ~~                                              
#>     INTxPBC (C_IN)   -0.019    0.025   -0.758    0.448
#>  .att1pbc1 ~~                                         
#>    .att2pb2           0.000                           
#>    .att3pb3           0.000                           
#>  .att2pbc2 ~~                                         
#>    .att3pb3           0.000                           
#>  .sn1pbc1 ~~                                          
#>    .sn2pbc2           0.000                           
#>    .sn3pbc3           0.000                           
#>  .sn2pbc2 ~~                                          
#>    .sn3pbc3           0.000                           
#>  .pbc1pbc1 ~~                                         
#>    .pbc2pb2           0.000                           
#>    .pbc3pb3           0.000                           
#>  .pbc2pbc2 ~~                                         
#>    .pbc3pb3           0.000                           
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .ATTPBC  (M_AT)    0.422    0.013   32.832    0.000
#>    .SNPBC   (M_SN)    0.302    0.010   30.547    0.000
#>    .PBCPBC  (M_PB)    0.575    0.011   51.528    0.000
#>    .att1              1.001    0.023   44.025    0.000
#>    .att2              1.008    0.021   47.861    0.000
#>    .att3              1.002    0.019   52.974    0.000
#>    .sn1               0.974    0.018   55.116    0.000
#>    .sn2               0.982    0.016   60.802    0.000
#>    .sn3               0.991    0.015   67.883    0.000
#>    .pbc1              0.983    0.021   47.193    0.000
#>    .pbc2              0.988    0.020   49.207    0.000
#>    .pbc3              0.998    0.018   54.376    0.000
#>    .b1                1.150    0.020   57.082    0.000
#>    .b2                1.132    0.018   61.428    0.000
#>    .att1pb1           0.391    0.038   10.185    0.000
#>    .att2pb2           0.330    0.032   10.249    0.000
#>    .att3pb3           0.256    0.025   10.137    0.000
#>    .sn1pbc1           0.262    0.029    9.112    0.000
#>    .sn2pbc2           0.226    0.024    9.516    0.000
#>    .sn3pbc3           0.177    0.019    9.454    0.000
#>    .pbc1pb1           0.553    0.038   14.525    0.000
#>    .pbc2pb2           0.501    0.032   15.560    0.000
#>    .pbc3pb3           0.421    0.025   16.843    0.000
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .ATT   (Vr_ATT)    0.346    0.023   15.014    0.000
#>    .SN     (Vr_SN)    0.179    0.013   13.986    0.000
#>    .PBC   (Vr_PBC)    0.245    0.016   15.052    0.000
#>    .BEH     (Z_BE)    0.513    0.029   18.005    0.000
#>     INT   (Vr_INT)    0.539    0.027   19.889    0.000
#>     INTPB (V_INTP)    1.535    0.076   20.148    0.000
#>    .ATTPB (V_ATTP)    0.687    0.022   30.946    0.000
#>    .SNPBC  (V_SNP)    0.354    0.013   27.980    0.000
#>    .PBCPB (V_PBCP)    0.662    0.026   25.764    0.000
#>    .att1   (Vr_t1)    0.150    0.008   18.624    0.000
#>    .att2   (Vr_t2)    0.165    0.007   22.170    0.000
#>    .att3   (Vr_t3)    0.147    0.006   23.582    0.000
#>    .sn1    (Vr_s1)    0.168    0.008   20.991    0.000
#>    .sn2    (Vr_s2)    0.168    0.007   24.105    0.000
#>    .sn3    (Vr_s3)    0.149    0.006   25.303    0.000
#>    .pbc1   (Vr_p1)    0.293    0.009   30.965    0.000
#>    .pbc2   (Vr_p2)    0.340    0.009   38.979    0.000
#>    .pbc3   (Vr_p3)    0.327    0.007   44.263    0.000
#>    .b1     (Vr_b1)    0.144    0.024    6.051    0.000
#>    .b2     (Vr_b2)    0.181    0.018    9.880    0.000
#>    .att11 (Vr_t11)    0.389    0.011   34.052    0.000
#>    .att22 (Vr_t22)    0.378    0.010   39.469    0.000
#>    .att33 (Vr_t33)    0.285    0.007   41.921    0.000
#>    .sn1p1 (Vr_s11)    0.279    0.008   35.023    0.000
#>    .sn2p2 (Vr_s22)    0.256    0.006   39.790    0.000
#>    .sn3p3 (Vr_s33)    0.191    0.005   41.982    0.000
#>    .pbc11 (Vr_p11)    0.423    0.015   28.080    0.000
#>    .pbc22 (Vr_p22)    0.432    0.013   33.193    0.000
#>    .pbc33 (Vr_p33)    0.334    0.009   35.936    0.000
#> 
#> Constraints:
#>                                                |Slack|
#>     V_ATTPBC-((V_ATT+V_INT*_ATT_INT^2)*(V_INT    0.000
#>     V_11-(_1_ATT^2*(V_ATT+V_INT*_ATT_INT^2)*V    0.000
#>     V_22-(_2_ATT^2*(V_ATT+V_INT*_ATT_INT^2)*V    0.000
#>     V_33-(_3_ATT^2*(V_ATT+V_INT*_ATT_INT^2)*V    0.000
#>     V_SNPBC-((V_INT*_SN_INT^2+V_SN)*(V_INT*_P    0.000
#>     V_11-(_1_SN^2*(V_INT*_SN_INT^2+V_SN)*V_1+    0.000
#>     V_22-(_2_SN^2*(V_INT*_SN_INT^2+V_SN)*V_2+    0.000
#>     V_33-(_3_SN^2*(V_INT*_SN_INT^2+V_SN)*V_3+    0.000
#>     V_PBCPBC-((V_INT*_PBC_INT^2+V_PBC)*(V_INT    0.000
#>     V_11-(_1_PBC^2*(V_INT*_PBC_INT^2+V_PBC)*V    0.000
#>     V_22-(_2_PBC^2*(V_INT*_PBC_INT^2+V_PBC)*V    0.000
#>     V_33-(_3_PBC^2*(V_INT*_PBC_INT^2+V_PBC)*V    0.000
#>     lmbd_tt1pbc1_ATTPBC-(lmbd_tt1_ATT*_1_PBC)    0.000
#>     lmbd_tt2pbc2_ATTPBC-(lmbd_tt2_ATT*_2_PBC)    0.000
#>     lmbd_tt3pbc3_ATTPBC-(lmbd_tt3_ATT*_3_PBC)    0.000
#>     lmbd_sn1pbc1_SNPBC-(lmbd_sn1_SN*lm_1_PBC)    0.000
#>     lmbd_sn2pbc2_SNPBC-(lmbd_sn2_SN*lm_2_PBC)    0.000
#>     lmbd_sn3pbc3_SNPBC-(lmbd_sn3_SN*lm_3_PBC)    0.000
#>     lmbd_pbc1pbc1_PBCPBC-(lmbd_p1_PBC*_1_PBC)    0.000
#>     lmbd_pbc2pbc2_PBCPBC-(lmbd_p2_PBC*_2_PBC)    0.000
#>     lmbd_pbc3pbc3_PBCPBC-(lmbd_p3_PBC*_3_PBC)    0.000
#>     Mn_ATTPBC-((Vr_INT*lmb_ATT_INT*_PBC_INT))    0.000
#>     Mn_SNPBC-((Vr_INT*lmbd_PBC_INT*l_SN_INT))    0.000
#>     Mn_PBCPBC-((Vr_INT*lmbd_PBC_INT^2+V_PBC))    0.000
#> 

est_dblcent  <- modsem(tpb, data = TPB_1SO, method = "dblcent")
summary(est_dblcent)
#> Estimating baseline model (H0)
#> Warning: Unable to estimate baseline model for models with higher-order interaction terms!
#> modsem (version 1.0.11, approach = dblcent):
#> 
#> Interaction Model Fit Measures (H1):
#>   Loglikelihood                              -45526.19 
#>   Akaike (AIC)                                91302.38 
#>   Bayesian (BIC)                              92002.49 
#>   Chi-square                                   6227.02 
#>   Degrees of Freedom                               505 
#>   P-value (Chi-square)                           0.000 
#>   RMSEA                                          0.075 
#>   CFI                                            0.954 
#>   SRMR                                           0.031 
#> 
#> R-Squared Interaction Model (H1):
#>   BEH                                            0.269 
#> 
#> lavaan 0.6-19 ended normally after 339 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                       125
#> 
#>   Number of observations                          2000
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                              6227.020
#>   Degrees of freedom                               505
#>   P-value (Chi-square)                           0.000
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   ATT =~                                              
#>     att1              1.000                           
#>     att2              0.909    0.011   82.767    0.000
#>     att3              0.802    0.010   80.426    0.000
#>   SN =~                                               
#>     sn1               1.000                           
#>     sn2               0.891    0.016   54.218    0.000
#>     sn3               0.790    0.015   52.448    0.000
#>   PBC =~                                              
#>     pbc1              1.000                           
#>     pbc2              0.909    0.014   66.751    0.000
#>     pbc3              0.793    0.013   63.368    0.000
#>   BEH =~                                              
#>     b1                1.000                           
#>     b2                0.864    0.029   30.237    0.000
#>   INT =~                                              
#>     ATT               1.000                           
#>     PBC               0.819    0.024   33.977    0.000
#>     SN                0.702    0.021   33.482    0.000
#>   INTxPBC =~                                          
#>     ATTPBC            1.000                           
#>     SNPBC             0.717    0.017   42.130    0.000
#>     PBCPBC            1.003    0.024   42.678    0.000
#>   ATTPBC =~                                           
#>     att1pbc1          1.000                           
#>     att2pbc1          0.896    0.010   94.261    0.000
#>     att3pbc1          0.808    0.009   92.565    0.000
#>     att1pbc2          0.899    0.011   79.087    0.000
#>     att2pbc2          0.804    0.013   61.655    0.000
#>     att3pbc2          0.727    0.012   61.499    0.000
#>     att1pbc3          0.752    0.011   70.229    0.000
#>     att2pbc3          0.675    0.012   55.856    0.000
#>     att3pbc3          0.614    0.011   56.951    0.000
#>   SNPBC =~                                            
#>     sn1pbc1           1.000                           
#>     sn2pbc1           0.884    0.013   65.689    0.000
#>     sn3pbc1           0.783    0.012   63.016    0.000
#>     sn1pbc2           0.893    0.012   76.169    0.000
#>     sn2pbc2           0.791    0.015   51.434    0.000
#>     sn3pbc2           0.693    0.014   49.947    0.000
#>     sn1pbc3           0.771    0.011   70.139    0.000
#>     sn2pbc3           0.673    0.014   48.082    0.000
#>     sn3pbc3           0.601    0.013   47.729    0.000
#>   PBCPBC =~                                           
#>     pbc1pbc1          1.000                           
#>     pbc2pbc1          0.902    0.010   86.187    0.000
#>     pbc3pbc1          0.764    0.010   79.520    0.000
#>     pbc2pbc2          0.812    0.018   44.917    0.000
#>     pbc3pbc2          0.689    0.013   51.106    0.000
#>     pbc3pbc3          0.590    0.015   40.172    0.000
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   BEH ~                                               
#>     PBC               0.189    0.043    4.359    0.000
#>     INT               0.187    0.045    4.140    0.000
#>     INTxPBC           0.203    0.015   13.754    0.000
#> 
#> Covariances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>  .att1pbc1 ~~                                         
#>    .att1pbc2          0.123    0.008   15.877    0.000
#>    .att1pbc3          0.108    0.007   15.606    0.000
#>    .att2pbc1          0.188    0.011   16.599    0.000
#>    .att3pbc1          0.166    0.010   16.500    0.000
#>  .att1pbc2 ~~                                         
#>    .att1pbc3          0.098    0.006   15.298    0.000
#>    .att2pbc2          0.211    0.011   19.889    0.000
#>    .att3pbc2          0.194    0.010   20.282    0.000
#>  .att1pbc3 ~~                                         
#>    .att2pbc3          0.237    0.010   23.982    0.000
#>    .att3pbc3          0.204    0.009   23.565    0.000
#>  .att2pbc1 ~~                                         
#>    .att2pbc2          0.156    0.008   20.453    0.000
#>    .att2pbc3          0.144    0.007   20.721    0.000
#>    .att3pbc1          0.153    0.009   16.556    0.000
#>  .att2pbc2 ~~                                         
#>    .att2pbc3          0.130    0.006   20.338    0.000
#>    .att3pbc2          0.174    0.009   19.943    0.000
#>  .att2pbc3 ~~                                         
#>    .att3pbc3          0.185    0.008   23.276    0.000
#>  .att3pbc1 ~~                                         
#>    .att3pbc2          0.133    0.006   20.972    0.000
#>    .att3pbc3          0.119    0.006   20.773    0.000
#>  .att3pbc2 ~~                                         
#>    .att3pbc3          0.107    0.005   20.310    0.000
#>  .sn1pbc1 ~~                                          
#>    .sn1pbc2           0.163    0.008   19.281    0.000
#>    .sn1pbc3           0.140    0.007   18.855    0.000
#>    .sn2pbc1           0.092    0.006   15.534    0.000
#>    .sn3pbc1           0.082    0.005   15.167    0.000
#>  .sn1pbc2 ~~                                          
#>    .sn1pbc3           0.128    0.007   18.607    0.000
#>    .sn2pbc2           0.103    0.006   18.569    0.000
#>    .sn3pbc2           0.097    0.005   19.080    0.000
#>  .sn1pbc3 ~~                                          
#>    .sn2pbc3           0.110    0.005   21.785    0.000
#>    .sn3pbc3           0.099    0.005   21.584    0.000
#>  .sn2pbc1 ~~                                          
#>    .sn2pbc2           0.137    0.007   19.800    0.000
#>    .sn2pbc3           0.122    0.006   19.734    0.000
#>    .sn3pbc1           0.074    0.005   15.077    0.000
#>  .sn2pbc2 ~~                                          
#>    .sn2pbc3           0.110    0.006   19.347    0.000
#>    .sn3pbc2           0.084    0.005   18.558    0.000
#>  .sn2pbc3 ~~                                          
#>    .sn3pbc3           0.087    0.004   21.102    0.000
#>  .sn3pbc1 ~~                                          
#>    .sn3pbc2           0.121    0.006   20.829    0.000
#>    .sn3pbc3           0.106    0.005   20.481    0.000
#>  .sn3pbc2 ~~                                          
#>    .sn3pbc3           0.094    0.005   20.111    0.000
#>  .pbc1pbc1 ~~                                         
#>    .pbc2pbc1          0.268    0.014   19.031    0.000
#>    .pbc3pbc1          0.240    0.012   19.187    0.000
#>  .pbc2pbc1 ~~                                         
#>    .pbc2pbc2          0.302    0.014   22.286    0.000
#>    .pbc3pbc1          0.106    0.006   17.710    0.000
#>    .pbc3pbc2          0.132    0.006   20.914    0.000
#>  .pbc2pbc2 ~~                                         
#>    .pbc3pbc2          0.244    0.011   21.890    0.000
#>  .pbc3pbc1 ~~                                         
#>    .pbc3pbc2          0.152    0.006   23.891    0.000
#>    .pbc3pbc3          0.272    0.011   24.933    0.000
#>  .pbc3pbc2 ~~                                         
#>    .pbc3pbc3          0.243    0.010   24.501    0.000
#>   INT ~~                                              
#>     INTxPBC          -0.038    0.033   -1.150    0.250
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .att1              0.152    0.008   18.217    0.000
#>    .att2              0.168    0.008   21.469    0.000
#>    .att3              0.147    0.006   22.691    0.000
#>    .sn1               0.171    0.009   19.842    0.000
#>    .sn2               0.165    0.008   21.850    0.000
#>    .sn3               0.151    0.006   23.313    0.000
#>    .pbc1              0.162    0.009   18.804    0.000
#>    .pbc2              0.167    0.008   21.285    0.000
#>    .pbc3              0.159    0.007   23.501    0.000
#>    .b1                0.144    0.021    6.755    0.000
#>    .b2                0.181    0.017   10.910    0.000
#>    .att1pbc1          0.363    0.016   23.326    0.000
#>    .att2pbc1          0.370    0.014   26.472    0.000
#>    .att3pbc1          0.305    0.011   26.846    0.000
#>    .att1pbc2          0.378    0.014   26.203    0.000
#>    .att2pbc2          0.366    0.013   28.665    0.000
#>    .att3pbc2          0.303    0.010   29.105    0.000
#>    .att1pbc3          0.369    0.013   29.040    0.000
#>    .att2pbc3          0.361    0.011   31.610    0.000
#>    .att3pbc3          0.281    0.009   31.219    0.000
#>    .sn1pbc1           0.306    0.012   25.602    0.000
#>    .sn2pbc1           0.256    0.010   26.184    0.000
#>    .sn3pbc1           0.228    0.008   27.105    0.000
#>    .sn1pbc2           0.300    0.011   27.855    0.000
#>    .sn2pbc2           0.243    0.009   27.943    0.000
#>    .sn3pbc2           0.209    0.007   29.007    0.000
#>    .sn1pbc3           0.260    0.009   29.414    0.000
#>    .sn2pbc3           0.225    0.007   30.173    0.000
#>    .sn3pbc3           0.185    0.006   30.379    0.000
#>    .pbc1pbc1          0.646    0.030   21.568    0.000
#>    .pbc2pbc1          0.306    0.011   27.014    0.000
#>    .pbc3pbc1          0.290    0.010   29.693    0.000
#>    .pbc2pbc2          0.618    0.025   24.778    0.000
#>    .pbc3pbc2          0.271    0.009   30.899    0.000
#>    .pbc3pbc3          0.482    0.018   27.155    0.000
#>    .ATT               0.389    0.025   15.717    0.000
#>    .SN                0.199    0.014   14.732    0.000
#>    .PBC               0.300    0.018   16.350    0.000
#>    .BEH               0.521    0.027   19.503    0.000
#>     INT               0.995    0.049   20.233    0.000
#>     INTxPBC           1.701    0.074   22.836    0.000
#>    .ATTPBC            0.309    0.025   12.116    0.000
#>    .SNPBC             0.193    0.015   12.939    0.000
#>    .PBCPBC            0.206    0.027    7.713    0.000
#> 
# }