TPB_UK

A dataset based on the Theory of Planned Behaviour from a UK sample. 4 variables with high communality were selected for each latent variable (ATT, SN, PBC, INT, BEH), from two time points (t1 and t2).

Source

Gathered from a replication study by Hagger et al. (2023).

Obtained from doi:10.23668/psycharchives.12187

Examples

tpb_uk <- "
# Outer Model (Based on Hagger et al., 2007)
 ATT =~ att3 + att2 + att1 + att4
 SN =~ sn4 + sn2 + sn3 + sn1
 PBC =~ pbc2 + pbc1 + pbc3 + pbc4
 INT =~ int2 + int1 + int3 + int4
 BEH =~ beh3 + beh2 + beh1 + beh4

# Inner Model (Based on Steinmetz et al., 2011)
 # Causal Relationsships
 INT ~ ATT + SN + PBC
 BEH ~ INT + PBC
 BEH ~ INT:PBC
"

est <- modsem(tpb_uk, data = TPB_UK)
summary(est)
#> Estimating baseline model (H0)
#> modsem (version 1.0.15, approach = dblcent):
#> 
#> Interaction Model Fit Measures (H1):
#>   Loglikelihood                              -67028.76 
#>   Akaike (AIC)                               134321.51 
#>   Bayesian (BIC)                             134989.95 
#>   Chi-square                                   7911.47 
#>   Degrees of Freedom                               534 
#>   P-value (Chi-square)                           0.000 
#>   RMSEA                                          0.109 
#>   CFI                                            0.908 
#>   SRMR                                           0.069 
#> 
#> Fit Measures for Baseline Model (H0):
#>   Loglikelihood                              -67128.02 
#>   Akaike (AIC)                               134518.04 
#>   Bayesian (BIC)                             135181.41 
#>   Chi-square                                   8110.00 
#>   Degrees of Freedom                               535 
#>   P-value (Chi-square)                           0.000 
#>   RMSEA                                          0.110 
#>   CFI                                            0.906 
#>   SRMR                                           0.086 
#> 
#> Comparative Fit to H0 (LRT test):
#>   Chi-square diff                              198.526 
#>   Degrees of freedom diff                            1 
#>   P-value (LRT)                                  0.000 
#> 
#> R-Squared Interaction Model (H1):
#>   INT                                            0.898 
#>   BEH                                            0.901 
#> R-Squared Baseline Model (H0):
#>   INT                                            0.898 
#>   BEH                                            0.868 
#> R-Squared Change (H1 - H0):
#>   INT                                           -0.000 
#>   BEH                                            0.033 
#> 
#> lavaan 0.6-20 ended normally after 273 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                       132
#> 
#>   Number of observations                          1169
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                              7911.470
#>   Degrees of freedom                               534
#>   P-value (Chi-square)                           0.000
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   ATT =~                                              
#>     att3              1.000                           
#>     att2              0.965    0.011   84.977    0.000
#>     att1              0.812    0.017   47.096    0.000
#>     att4              0.870    0.019   45.663    0.000
#>   SN =~                                               
#>     sn4               1.000                           
#>     sn2               1.313    0.039   33.539    0.000
#>     sn3               1.350    0.040   34.169    0.000
#>     sn1               1.001    0.039   25.926    0.000
#>   PBC =~                                              
#>     pbc2              1.000                           
#>     pbc1              0.862    0.020   42.710    0.000
#>     pbc3              0.936    0.017   54.206    0.000
#>     pbc4              0.822    0.020   41.241    0.000
#>   INT =~                                              
#>     int2              1.000                           
#>     int1              0.970    0.011   91.925    0.000
#>     int3              0.985    0.010   98.760    0.000
#>     int4              0.992    0.009  104.902    0.000
#>   BEH =~                                              
#>     beh3              1.000                           
#>     beh2              0.986    0.013   75.930    0.000
#>     beh1              0.815    0.019   43.156    0.000
#>     beh4              0.805    0.019   41.913    0.000
#>   INTPBC =~                                           
#>     int2pbc2          1.000                           
#>     int1pbc2          0.984    0.017   57.274    0.000
#>     int3pbc2          1.014    0.020   50.865    0.000
#>     int4pbc2          1.013    0.017   60.855    0.000
#>     int2pbc1          1.283    0.042   30.600    0.000
#>     int1pbc1          1.239    0.043   28.553    0.000
#>     int3pbc1          1.280    0.045   28.502    0.000
#>     int4pbc1          1.300    0.045   29.093    0.000
#>     int2pbc3          1.023    0.039   25.974    0.000
#>     int1pbc3          1.004    0.042   23.860    0.000
#>     int3pbc3          1.031    0.041   25.075    0.000
#>     int4pbc3          1.029    0.042   24.692    0.000
#>     int2pbc4          1.291    0.042   30.637    0.000
#>     int1pbc4          1.251    0.044   28.483    0.000
#>     int3pbc4          1.285    0.045   28.663    0.000
#>     int4pbc4          1.292    0.044   29.062    0.000
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   INT ~                                               
#>     ATT              -0.068    0.028   -2.437    0.015
#>     SN                0.043    0.033    1.289    0.197
#>     PBC               1.049    0.036   29.185    0.000
#>   BEH ~                                               
#>     INT               0.638    0.047   13.427    0.000
#>     PBC               0.350    0.051    6.840    0.000
#>     INTPBC            0.150    0.011   13.910    0.000
#> 
#> Covariances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>  .int1pbc2 ~~                                         
#>    .int1pbc1          0.875    0.065   13.560    0.000
#>  .int1pbc1 ~~                                         
#>    .int1pbc3          0.906    0.061   14.724    0.000
#>    .int1pbc4          1.283    0.072   17.773    0.000
#>  .int2pbc1 ~~                                         
#>    .int1pbc1          3.090    0.222   13.941    0.000
#>  .int1pbc1 ~~                                         
#>    .int3pbc1          2.960    0.217   13.663    0.000
#>    .int4pbc1          2.964    0.219   13.504    0.000
#>  .int1pbc2 ~~                                         
#>    .int1pbc3          1.176    0.069   17.141    0.000
#>    .int1pbc4          0.952    0.065   14.621    0.000
#>  .int2pbc2 ~~                                         
#>    .int1pbc2          5.915    0.290   20.426    0.000
#>  .int1pbc2 ~~                                         
#>    .int3pbc2          5.292    0.268   19.742    0.000
#>    .int4pbc2          5.766    0.287   20.079    0.000
#>  .int1pbc3 ~~                                         
#>    .int1pbc4          1.022    0.063   16.151    0.000
#>  .int2pbc3 ~~                                         
#>    .int1pbc3          6.073    0.298   20.410    0.000
#>  .int1pbc3 ~~                                         
#>    .int3pbc3          5.471    0.275   19.928    0.000
#>    .int4pbc3          6.077    0.295   20.571    0.000
#>  .int2pbc4 ~~                                         
#>    .int1pbc4          2.984    0.220   13.550    0.000
#>  .int1pbc4 ~~                                         
#>    .int3pbc4          2.950    0.217   13.571    0.000
#>    .int4pbc4          3.034    0.220   13.774    0.000
#>  .int2pbc2 ~~                                         
#>    .int2pbc1          0.752    0.053   14.131    0.000
#>  .int2pbc1 ~~                                         
#>    .int2pbc3          0.453    0.044   10.295    0.000
#>    .int2pbc4          0.792    0.054   14.614    0.000
#>    .int3pbc1          3.234    0.227   14.221    0.000
#>    .int4pbc1          3.244    0.230   14.080    0.000
#>  .int2pbc2 ~~                                         
#>    .int2pbc3          0.669    0.050   13.460    0.000
#>    .int2pbc4          0.556    0.047   11.750    0.000
#>    .int3pbc2          5.346    0.269   19.843    0.000
#>    .int4pbc2          5.869    0.287   20.417    0.000
#>  .int2pbc3 ~~                                         
#>    .int2pbc4          0.368    0.041    9.088    0.000
#>    .int3pbc3          5.519    0.275   20.058    0.000
#>    .int4pbc3          5.983    0.293   20.389    0.000
#>  .int2pbc4 ~~                                         
#>    .int3pbc4          3.010    0.223   13.483    0.000
#>    .int4pbc4          3.086    0.226   13.662    0.000
#>  .int3pbc2 ~~                                         
#>    .int3pbc1          0.673    0.071    9.523    0.000
#>  .int3pbc1 ~~                                         
#>    .int3pbc3          0.410    0.058    7.121    0.000
#>    .int3pbc4          1.423    0.076   18.659    0.000
#>    .int4pbc1          3.124    0.226   13.826    0.000
#>  .int3pbc2 ~~                                         
#>    .int3pbc3          1.347    0.077   17.406    0.000
#>    .int3pbc4          0.542    0.068    8.020    0.000
#>    .int4pbc2          5.302    0.269   19.709    0.000
#>  .int3pbc3 ~~                                         
#>    .int3pbc4          0.419    0.056    7.440    0.000
#>    .int4pbc3          5.500    0.273   20.118    0.000
#>  .int3pbc4 ~~                                         
#>    .int4pbc4          3.069    0.224   13.713    0.000
#>  .int4pbc2 ~~                                         
#>    .int4pbc1          0.404    0.052    7.849    0.000
#>  .int4pbc1 ~~                                         
#>    .int4pbc3          0.355    0.044    8.109    0.000
#>    .int4pbc4          0.831    0.055   15.012    0.000
#>  .int4pbc2 ~~                                         
#>    .int4pbc3          0.882    0.056   15.805    0.000
#>    .int4pbc4          0.323    0.049    6.640    0.000
#>  .int4pbc3 ~~                                         
#>    .int4pbc4          0.349    0.042    8.228    0.000
#>   ATT ~~                                              
#>     SN                1.681    0.108   15.630    0.000
#>     PBC               3.683    0.177   20.760    0.000
#>     INTPBC            0.177    0.186    0.951    0.342
#>   SN ~~                                               
#>     PBC               1.939    0.115   16.822    0.000
#>     INTPBC           -0.092    0.117   -0.791    0.429
#>   PBC ~~                                              
#>     INTPBC            0.390    0.185    2.104    0.035
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .att3              0.296    0.024   12.124    0.000
#>    .att2              0.305    0.023   12.999    0.000
#>    .att1              1.286    0.057   22.563    0.000
#>    .att4              1.585    0.070   22.678    0.000
#>    .sn4               1.362    0.061   22.411    0.000
#>    .sn2               0.494    0.034   14.316    0.000
#>    .sn3               0.375    0.033   11.410    0.000
#>    .sn1               1.444    0.064   22.512    0.000
#>    .pbc2              0.716    0.039   18.338    0.000
#>    .pbc1              1.442    0.065   22.140    0.000
#>    .pbc3              0.803    0.041   19.708    0.000
#>    .pbc4              1.442    0.065   22.331    0.000
#>    .int2              0.237    0.014   17.100    0.000
#>    .int1              0.406    0.020   20.342    0.000
#>    .int3              0.331    0.017   19.292    0.000
#>    .int4              0.270    0.015   18.063    0.000
#>    .beh3              0.459    0.030   15.127    0.000
#>    .beh2              0.520    0.032   16.379    0.000
#>    .beh1              1.826    0.080   22.746    0.000
#>    .beh4              1.905    0.083   22.842    0.000
#>    .int2pbc2          6.989    0.301   23.186    0.000
#>    .int1pbc2          7.325    0.307   23.830    0.000
#>    .int3pbc2          6.953    0.284   24.503    0.000
#>    .int4pbc2          7.007    0.300   23.324    0.000
#>    .int2pbc1          4.440    0.247   17.991    0.000
#>    .int1pbc1          4.356    0.234   18.652    0.000
#>    .int3pbc1          4.730    0.245   19.336    0.000
#>    .int4pbc1          4.155    0.243   17.098    0.000
#>    .int2pbc3          6.804    0.306   22.238    0.000
#>    .int1pbc3          7.543    0.315   23.952    0.000
#>    .int3pbc3          6.483    0.276   23.521    0.000
#>    .int4pbc3          6.879    0.301   22.874    0.000
#>    .int2pbc4          3.897    0.238   16.403    0.000
#>    .int1pbc4          4.508    0.237   19.008    0.000
#>    .int3pbc4          4.545    0.242   18.812    0.000
#>    .int4pbc4          4.116    0.240   17.151    0.000
#>     ATT               4.453    0.197   22.581    0.000
#>     SN                1.718    0.117   14.744    0.000
#>     PBC               4.346    0.209   20.813    0.000
#>    .INT               0.505    0.039   13.093    0.000
#>    .BEH               0.526    0.035   14.904    0.000
#>     INTPBC            7.949    0.567   14.020    0.000
#>