Bootstrap a modsem Model

A generic interface for parametric and non‑parametric bootstrap procedures for structural equation models estimated with the modsem ecosystem. The function dispatches on the class of model; currently dedicated methods exist for modsem_pi (product‑indicator approach) and modsem_da (distributional‑analytic approach).

Usage

bootstrap_modsem(model, FUN, ...)

# S3 method for class 'modsem_pi'
bootstrap_modsem(model, FUN, ...)

# S3 method for class 'modsem_da'
bootstrap_modsem(
  model,
  FUN = "coef",
  R = 1000L,
  P.max = 1e+05,
  type = c("nonparametric", "parameteric"),
  verbose = interactive(),
  calc.se = FALSE,
  optimize = FALSE,
  ...
)

Arguments

model

A fitted modsem object.

FUN

A function that returns the statistic of interest when applied to a fitted model. The function must accept a single argument, the model object, and should ideally return a numeric vector; see Value.

...

Additional arguments forwarded to lavaan::bootstrapLavaan for modsem_pi objects, or modsem_da for modsem_da objects.

R

number of bootstrap replicates.

P.max

ceiling for the simulated population size.

type

Bootstrap flavour, see Details.

verbose

Should progress information be printed to the console?

calc.se

Should standard errors for each replicate. Defaults to FALSE.

optimize

Should starting values be re-optimized for each replicate. Defaults to FALSE.

Value

Depending on the return type of FUN either

numeric

A matrix with R rows (bootstrap replicates) and as many columns as length(FUN(model)).

other

A list of length R; each element is the raw output of FUN. NOTE: Only applies for modsem_da objects

Details

A thin wrapper around lavaan::bootstrapLavaan() that performs the necessary book‑keeping so that FUN receives a fully‑featured modsem_pi object—rather than a bare lavaan fit—at every iteration.

The function internally resamples the observed data (non‑parametric case) or simulates from the estimated parameter table (parametric case), feeds the sample to modsem_da, evaluates FUN on the refitted object and finally collates the results.

Methods (by class)

bootstrap_modsem(modsem_pi): Bootstrap a modsem_pi model by delegating to bootstrapLavaan.

bootstrap_modsem(modsem_da): Parametric or non‑parametric bootstrap for modsem_da models.

See also

Examples

m1 <- '
  X =~ x1 + x2
  Z =~ z1 + z2
  Y =~ y1 + y2

  Y ~ X + Z + X:Z
'

fit_pi <- modsem(m1, oneInt)
bootstrap_modsem(fit_pi, FUN = coef, R = 10L)
#>           X=~x2     Z=~z2     Y=~y2       Y~X       Y~Z      Y~XZ  XZ=~x2z1
#>  [1,] 0.8884237 0.8262879 0.7797705 0.7407114 0.5918870 0.6707317 0.8246330
#>  [2,] 0.8347994 0.8233353 0.8025760 0.6824978 0.6016801 0.6865373 0.8116811
#>  [3,] 0.8206546 0.8518766 0.8007877 0.6971329 0.5362172 0.6820370 0.8238181
#>  [4,] 0.8189006 0.8113157 0.8325491 0.6869278 0.5782950 0.7256083 0.7789803
#>  [5,] 0.7947128 0.8558326 0.7924540 0.7333650 0.6106507 0.7282757 0.8221329
#>  [6,] 0.8254009 0.8749817 0.7929519 0.6579637 0.6131392 0.7361599 0.8396432
#>  [7,] 0.8332576 0.7692559 0.8131625 0.7067837 0.5708261 0.7091723 0.8355242
#>  [8,] 0.8619048 0.8171167 0.7882294 0.7320317 0.6299300 0.7734579 0.8497304
#>  [9,] 0.8272198 0.8581544 0.8065377 0.7420436 0.5475846 0.7715898 0.7902094
#> [10,] 0.8325728 0.8515355 0.8051629 0.6431409 0.6255022 0.7367125 0.8150739
#>        XZ=~x1z2  XZ=~x2z2 x1z1~~x1z2 x1z1~~x2z1 x1z2~~x2z2 x2z1~~x2z2    x1~~x1
#>  [1,] 0.7706057 0.6378158  0.1321030 0.12470048 0.13859185  0.1236854 0.2695500
#>  [2,] 0.7620211 0.6213905  0.1510716 0.08864615 0.15350198  0.1165963 0.2064331
#>  [3,] 0.8140617 0.6704431  0.1392288 0.12953526 0.12479679  0.1391153 0.2037414
#>  [4,] 0.7923717 0.6249382  0.1028956 0.14624920 0.13305408  0.1534371 0.2024947
#>  [5,] 0.8138680 0.6766186  0.1016855 0.17156133 0.13094839  0.1542808 0.1466010
#>  [6,] 0.7600386 0.6436328  0.1071415 0.07537641 0.17072580  0.1236519 0.1713316
#>  [7,] 0.7848949 0.6607577  0.1366991 0.11834861 0.13430508  0.1109281 0.2181188
#>  [8,] 0.8331755 0.7173464  0.1227065 0.15731268 0.08557116  0.1282096 0.2078369
#>  [9,] 0.8248556 0.6578495  0.1125763 0.16228584 0.10511858  0.1216584 0.2034808
#> [10,] 0.8145436 0.6751385  0.1228619 0.15560824 0.12230915  0.1129503 0.1708143
#>          x2~~x2    z1~~z1    z2~~z2    y1~~y1    y2~~y2 x1z1~~x1z1 x2z1~~x2z1
#>  [1,] 0.1068363 0.2120969 0.1364207 0.1522135 0.1715433  0.3467576  0.2808781
#>  [2,] 0.1326718 0.1849395 0.1475439 0.1745294 0.1382708  0.3035680  0.2380796
#>  [3,] 0.1548883 0.1971554 0.1424606 0.1466531 0.1517053  0.3601218  0.2922566
#>  [4,] 0.1458213 0.1939336 0.1492024 0.1856174 0.1279031  0.3284856  0.3442310
#>  [5,] 0.1768187 0.2159352 0.1325914 0.1723244 0.1427089  0.3648349  0.3440650
#>  [6,] 0.1504944 0.2341332 0.1190925 0.1613587 0.1484900  0.2417272  0.2421190
#>  [7,] 0.1401831 0.1358201 0.1775705 0.1744732 0.1359542  0.3335131  0.2663161
#>  [8,] 0.1369614 0.2094457 0.1333073 0.1566526 0.1512443  0.3620863  0.3218547
#>  [9,] 0.1455015 0.2406770 0.1118286 0.2188779 0.1184675  0.3601850  0.3186712
#> [10,] 0.1532201 0.2130209 0.1350956 0.1946312 0.1367605  0.3495799  0.2831961
#>       x1z2~~x1z2 x2z2~~x2z2      X~~X      Z~~Z      Y~~Y    XZ~~XZ      X~~Z
#>  [1,]  0.3058370  0.2433937 0.9250376 0.9509946 1.0404153 0.9839416 0.1510574
#>  [2,]  0.3521438  0.2414933 0.9325914 1.0209942 0.9759887 1.1139648 0.1684377
#>  [3,]  0.3048880  0.2515079 0.9511693 0.9631768 0.9344947 1.1739524 0.1752234
#>  [4,]  0.2790108  0.2581793 0.9726406 1.0246144 0.9579877 1.1140874 0.1988152
#>  [5,]  0.2749447  0.2623786 1.0739408 0.9043633 0.9671002 0.9805299 0.1828608
#>  [6,]  0.3181851  0.2738636 0.9577705 0.9431554 0.9825834 1.0354563 0.1617813
#>  [7,]  0.3089064  0.2212350 0.8914256 1.0704986 0.9203757 0.9886425 0.1810837
#>  [8,]  0.2373762  0.1982495 0.8862861 1.0080460 0.9482141 0.8543592 0.2248769
#>  [9,]  0.2559218  0.2049166 0.9444809 0.9177360 0.9206756 0.9224049 0.1949300
#> [10,]  0.2906092  0.2260151 0.9269864 0.9611296 0.9645808 0.8736839 0.1970902
#>               X~~XZ      Z~~XZ
#>  [1,]  0.0045452923 0.06431493
#>  [2,] -0.0094172460 0.08608002
#>  [3,]  0.0412521613 0.13538289
#>  [4,] -0.0004327796 0.06455298
#>  [5,] -0.0391503620 0.04776190
#>  [6,]  0.0258802984 0.03401701
#>  [7,]  0.0028456548 0.05461465
#>  [8,]  0.0450855234 0.02570986
#>  [9,] -0.0242799280 0.04269811
#> [10,]  0.0395353338 0.04479171
#> attr(,"error.idx")
#> integer(0)
#> attr(,"nonadmissible")
#> integer(0)
#> attr(,"seed")
#> [1] 80750137

m1 <- '
  X =~ x1 + x2
  Z =~ z1 + z2
  Y =~ y1 + y2

  Y ~ X + Z + X:Z
'

# \dontrun{
fit_lms <- modsem(m1, oneInt, method = "lms")
bootstrap_modsem(fit_lms, FUN = coef, R = 10L)
#>           X=~x2     Z=~z2     Y=~y2      x1~1     x2~1      z1~1     z2~1
#>  [1,] 0.7894381 0.8165751 0.8105481 0.9973693 1.224880 0.9994796 1.200980
#>  [2,] 0.8331841 0.8051841 0.7962777 0.9936042 1.207708 1.0157326 1.195425
#>  [3,] 0.8193427 0.8242812 0.8142980 0.9927981 1.199501 0.9931045 1.210192
#>  [4,] 0.8365117 0.8298861 0.8068577 1.0097207 1.203789 0.9954156 1.188908
#>  [5,] 0.8158523 0.8152020 0.7877853 0.9691707 1.176828 0.9692409 1.189185
#>  [6,] 0.7946593 0.8005864 0.7990006 1.0311934 1.223692 1.0057577 1.203710
#>  [7,] 0.8305283 0.8338572 0.8264926 1.0042485 1.207220 1.0229617 1.204045
#>  [8,] 0.8177754 0.8469796 0.8096941 1.0175175 1.213969 1.0116689 1.203822
#>  [9,] 0.8245500 0.8104430 0.8073422 1.0071998 1.203752 0.9786610 1.180682
#> [10,] 0.7902249 0.7935164 0.8140481 1.0516057 1.228572 0.9929604 1.209066
#>            y1~1     y2~1    x1~~x1    x2~~x2    z1~~z1    z2~~z2    y1~~y1
#>  [1,] 1.0340181 1.219180 0.1600891 0.1599330 0.1685458 0.1547462 0.1955621
#>  [2,] 1.0192376 1.192515 0.1809772 0.1461584 0.1771669 0.1456546 0.1735967
#>  [3,] 1.0517702 1.193204 0.1662756 0.1712980 0.1724747 0.1502523 0.1845224
#>  [4,] 1.0612830 1.233944 0.1671169 0.1479685 0.1867670 0.1504038 0.1977463
#>  [5,] 0.9633551 1.159458 0.1677580 0.1611803 0.1715718 0.1632913 0.1735572
#>  [6,] 1.0192256 1.198203 0.1609212 0.1726697 0.1656097 0.1578540 0.1759655
#>  [7,] 1.0096373 1.209528 0.1699689 0.1659470 0.1801441 0.1473709 0.2190238
#>  [8,] 1.0569811 1.225090 0.1678311 0.1584622 0.1994462 0.1418989 0.1944516
#>  [9,] 0.9598368 1.172779 0.1801700 0.1472442 0.1798854 0.1410010 0.1774392
#> [10,] 1.0595982 1.241959 0.1336812 0.1779883 0.1600576 0.1574076 0.1856022
#>          y2~~y2      X~~X      X~~Z     Z~~Z      Y~~Y       Y~X       Y~Z
#>  [1,] 0.1205706 0.9771458 0.2424757 1.049472 0.9203301 0.6566347 0.6126260
#>  [2,] 0.1460726 0.9272634 0.2131866 1.136236 1.0027116 0.6425338 0.5359367
#>  [3,] 0.1288665 0.9732755 0.1881165 1.036654 0.9902156 0.6526408 0.5753070
#>  [4,] 0.1428636 0.9203692 0.1796760 1.048697 0.9582821 0.6931793 0.5987856
#>  [5,] 0.1473339 0.9786099 0.2249203 1.023476 0.9423357 0.5972354 0.5342007
#>  [6,] 0.1403258 1.0243164 0.2258463 1.132225 0.9525672 0.6840771 0.5965143
#>  [7,] 0.1059683 0.8955007 0.1707920 1.011368 0.8728637 0.6745512 0.5923576
#>  [8,] 0.1435447 0.9519428 0.1906275 1.012764 0.8889545 0.6745981 0.6552853
#>  [9,] 0.1288651 0.9342486 0.2201249 1.141007 0.9294426 0.6865500 0.5723825
#> [10,] 0.1358333 0.9939645 0.2721215 1.121061 0.9332573 0.6047255 0.5769146
#>           Y~X:Z
#>  [1,] 0.7291547
#>  [2,] 0.7241083
#>  [3,] 0.7123860
#>  [4,] 0.7770419
#>  [5,] 0.7323234
#>  [6,] 0.7244791
#>  [7,] 0.7547535
#>  [8,] 0.7509751
#>  [9,] 0.7436686
#> [10,] 0.7102271
# }