Estimate baseline model for modsem models

Estimates a baseline model (H0) from a given model (H1) and compares the fit of both models. The baseline model is estimated by removing all interaction terms from the model.

Usage

estimate_h0(object, warn_no_interaction = TRUE, ...)

Arguments

object

An object of class modsem_da or modsem_pi.

warn_no_interaction

Logical. If `TRUE`, a warning is issued if no interaction terms are found in the model.

...

Additional arguments passed to the `modsem_da` function, overriding the arguments in the original model.

Examples

if (FALSE) { # \dontrun{
m1 <- "
 # Outer Model
 X =~ x1 + x2 + x3
 Y =~ y1 + y2 + y3
 Z =~ z1 + z2 + z3

 # Inner model
 Y ~ X + Z + X:Z
"

# LMS approach
est_h1 <- modsem(m1, oneInt, "lms")
est_h0 <- estimate_h0(est_h1, calc.se=FALSE) # std.errors are not needed
compare_fit(est_h1 = est_h1, est_h0 = est_h0)

# Double centering approach
est_h1 <- modsem(m1, oneInt, method = "dblcent")
est_h0 <- estimate_h0(est_h1, oneInt)

compare_fit(est_h1 = est_h1, est_h0 = est_h0)

# Constrained approach
est_h1 <- modsem(m1, oneInt, method = "ca")
est_h0 <- estimate_h0(est_h1, oneInt)

compare_fit(est_h1 = est_h1, est_h0 = est_h0)
} # }