extract lavaan object from modsem object estimated using product indicators

extract lavaan object from modsem object estimated using product indicators

Usage

extract_lavaan(object)

Arguments

object

modsem object

Value

lavaan object

Examples

library(modsem)
m1 <- '
  # Outer Model
  X =~ x1 + x2 + x3
  Y =~ y1 + y2 + y3
  Z =~ z1 + z2 + z3

  # Inner model
  Y ~ X + Z + X:Z
'
est <- modsem_pi(m1, oneInt)
lav_est <- extract_lavaan(est)