Jordan subset of PISA 2006 data

The data stem from the large-scale assessment study PISA 2006 (Organisation for Economic Co-Operation and Development, 2009) where competencies of 15-year-old students in reading, mathematics, and science are assessed using nationally representative samples in 3-year cycles. In this example, data from the student background questionnaire from the Jordan sample of PISA 2006 were used. Only data of students with complete responses to all 15 items (N = 6,038) were considered.

Format

A data frame of fifteen variables and 6,038 observations:

enjoy1

indicator for enjoyment of science, item ST16Q01: I generally have fun when I am learning <broad science> topics.

enjoy2

indicator for enjoyment of science, item ST16Q02: I like reading about <broad science>.

enjoy3

indicator for enjoyment of science, item ST16Q03: I am happy doing <broad science> problems.

enjoy4

indicator for enjoyment of science, item ST16Q04: I enjoy acquiring new knowledge in <broad science>.

enjoy5

indicator for enjoyment of science, item ST16Q05: I am interested in learning about <broad science>.

academic1

indicator for academic self-concept in science, item ST37Q01: I can easily understand new ideas in <school science>.

academic2

indicator for academic self-concept in science, item ST37Q02: Learning advanced <school science> topics would be easy for me.

academic3

indicator for academic self-concept in science, item ST37Q03: I can usually give good answers to <test questions> on <school science> topics.

academic4

indicator for academic self-concept in science, item ST37Q04: I learn <school science> topics quickly.

academic5

indicator for academic self-concept in science, item ST37Q05: <School science> topics are easy for me.

academic6

indicator for academic self-concept in science, item ST37Q06: When I am being taught <school science>, I can understand the concepts very well.

career1

indicator for career aspirations in science, item ST29Q01: I would like to work in a career involving <broad science>.

career2

indicator for career aspirations in science, item ST29Q02: I would like to study <broad science> after <secondary school>.

career3

indicator for career aspirations in science, item ST29Q03: I would like to spend my life doing advanced <broad science>.

career4

indicator for career aspirations in science, item ST29Q04: I would like to work on <broad science> projects as an adult.

Source

This version of the dataset, as well as the description was gathered from the documentation of the 'nlsem' package (https://cran.r-project.org/package=nlsem), where the only difference is that the names of the variables were changed

Originally the dataset was gathered by the Organisation for Economic Co-Operation and Development (2009). Pisa 2006: Science competencies for tomorrow's world (Tech. Rep.). Paris, France. Obtained from: https://www.oecd.org/pisa/pisaproducts/database-pisa2006.htm

Examples

# \dontrun{
m1 <- "
  ENJ =~ enjoy1 + enjoy2 + enjoy3 + enjoy4 + enjoy5
  CAREER =~ career1 + career2 + career3 + career4
  SC =~ academic1 + academic2 + academic3 + academic4 + academic5 + academic6
  CAREER ~ ENJ + SC + ENJ:ENJ + SC:SC + ENJ:SC
"

est <- modsem(m1, data = jordan, method = "qml")
#> Warning: The variance-covariance matrix of the estimated parameters
#> (vcov) does not appear to be positive definite! The smallest
#> eigenvalue (= -3.811339e-05) is smaller than zero. This may
#> be a symptom that the model is not identified.
summary(est)
#> 
#> modsem (1.0.14) ended normally after 61 iterations
#> 
#>   Estimator                                        QML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        51
#> 
#>   Number of observations                          6038
#> 
#> Loglikelihood and Information Criteria:
#>   Loglikelihood                             -110519.99
#>   Akaike (AIC)                               221141.99
#>   Bayesian (BIC)                             221483.98
#>  
#> Fit Measures for Baseline Model (H0):
#>                                               Standard
#>   Chi-square                                   1016.34
#>   Degrees of Freedom (Chi-square)                   87
#>   P-value (Chi-square)                           0.000
#>   RMSEA                                          0.042
#>                                                       
#>   Loglikelihood                             -110521.29
#>   Akaike (AIC)                               221138.58
#>   Bayesian (BIC)                             221460.46
#>  
#> Comparative Fit to H0 (LRT test):
#>   Loglikelihood change                            1.30
#>   Difference test (D)                             2.59
#>   Degrees of freedom (D)                             3
#>   P-value (D)                                    0.459
#>  
#> R-Squared Interaction Model (H1):
#>   CAREER                                         0.513
#> R-Squared Baseline Model (H0):
#>   CAREER                                         0.510
#> R-Squared Change (H1 - H0):
#>   CAREER                                         0.003
#> 
#> Parameter Estimates:
#>   Coefficients                          unstandardized
#>   Information                                 observed
#>   Standard errors                             standard
#>  
#> Latent Variables:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>   ENJ =~        
#>     enjoy1          1.000                             
#>     enjoy2          1.002      0.020   50.577    0.000
#>     enjoy3          0.894      0.020   43.667    0.000
#>     enjoy4          0.999      0.021   48.222    0.000
#>     enjoy5          1.047      0.021   50.394    0.000
#>   SC =~         
#>     academic1       1.000                             
#>     academic2       1.104      0.028   38.946    0.000
#>     academic3       1.235      0.030   41.721    0.000
#>     academic4       1.254      0.030   41.830    0.000
#>     academic5       1.113      0.029   38.649    0.000
#>     academic6       1.198      0.030   40.358    0.000
#>   CAREER =~     
#>     career1         1.000                             
#>     career2         1.040      0.016   65.181    0.000
#>     career3         0.952      0.016   57.840    0.000
#>     career4         0.818      0.017   48.360    0.000
#> 
#> Regressions:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>   CAREER ~      
#>     ENJ             0.526      0.020   26.308    0.000
#>     SC              0.465      0.023   20.113    0.000
#>     ENJ:ENJ         0.029      0.021    1.373    0.170
#>     ENJ:SC         -0.046      0.043   -1.083    0.279
#>     SC:SC           0.002      0.033    0.049    0.961
#> 
#> Intercepts:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>    .enjoy1          0.000      0.010   -0.022    0.983
#>    .enjoy2          0.000      0.001    0.043    0.966
#>    .enjoy3          0.000      0.011   -0.037    0.970
#>    .enjoy4          0.000      0.001   -0.063    0.950
#>    .enjoy5          0.000      0.010    0.021    0.984
#>    .academic1       0.000      0.012   -0.018    0.986
#>    .academic2       0.000      0.012   -0.019    0.985
#>    .academic3      -0.001      0.012   -0.042    0.966
#>    .academic4       0.000      0.012   -0.029    0.977
#>    .academic5      -0.001      0.012   -0.060    0.952
#>    .academic6       0.001      0.012    0.045    0.964
#>    .career1        -0.005      0.016   -0.307    0.759
#>    .career2        -0.006      0.016   -0.357    0.721
#>    .career3        -0.005      0.016   -0.312    0.755
#>    .career4        -0.005      0.015   -0.321    0.748
#> 
#> Covariances:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>   ENJ ~~        
#>     SC              0.218      0.009   25.478    0.000
#> 
#> Variances:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>    .enjoy1          0.487      0.011   44.350    0.000
#>    .enjoy2          0.489      0.011   44.420    0.000
#>    .enjoy3          0.596      0.012   48.234    0.000
#>    .enjoy4          0.488      0.011   44.569    0.000
#>    .enjoy5          0.442      0.010   42.478    0.000
#>    .academic1       0.644      0.013   49.812    0.000
#>    .academic2       0.566      0.012   47.864    0.000
#>    .academic3       0.473      0.011   44.319    0.000
#>    .academic4       0.455      0.010   43.581    0.000
#>    .academic5       0.565      0.012   47.687    0.000
#>    .academic6       0.502      0.011   45.439    0.000
#>    .career1         0.373      0.009   40.396    0.000
#>    .career2         0.328      0.009   37.016    0.000
#>    .career3         0.436      0.010   43.273    0.000
#>    .career4         0.576      0.012   48.375    0.000
#>     ENJ             0.500      0.017   29.547    0.000
#>     SC              0.338      0.015   23.194    0.000
#>    .CAREER          0.302      0.010   29.725    0.000
#> 
# }