Get the simple slopes of a SEM model

Get the simple slopes of a SEM model

Usage

simple_slopes(
  x,
  z,
  y,
  xz = NULL,
  model,
  vals_x = -3:3,
  vals_z = -1:1,
  rescale = TRUE,
  ci_width = 1.96,
  relative_h0 = TRUE,
  ...
)

Arguments

x

The name of the variable on the x-axis

z

The name of the moderator variable

y

The name of the outcome variable

xz

The name of the interaction term. If the interaction term is not specified, it

model

An object of class modsem_pi, modsem_da, or modsem_mplus will be created using x and z.

vals_x

The values of the x variable to plot, the more values the smoother the std.error-area will be. NOTE: vals_x are measured relative to the mean and standard deviation of x (overridden by rescale=FALSE). The correct values will show up in the table.

vals_z

The values of the moderator variable to plot. A separate regression NOTE: vals_z are measured relative to the mean and standard deviation of z (overridden by rescale=FALSE). The correct values will show up in the table. line (y ~ x | z) will be plotted for each value of the moderator variable

rescale

Logical. If TRUE (default), the values of x and z will be rescaled relative to their means and standard deviations.

ci_width

The width of the confidence interval (default is 1.96, corresponding to a 95% confidence interval)

relative_h0

Logical. If TRUE (default), the null hypothesis is that the mean of the outcome variable is equal to the predicted value of the outcome variable. If FALSE, the null hypothesis is that the outcome variable is equal to zero.

...

Additional arguments passed to other functions

Value

A ggplot object

Examples

library(modsem)
if (FALSE) { # \dontrun{
m1 <- "
# Outer Model
  X =~ x1
  X =~ x2 + x3
  Z =~ z1 + z2 + z3
  Y =~ y1 + y2 + y3

# Inner model
  Y ~ X + Z + X:Z
"
est1 <- modsem(m1, data = oneInt)
simple_slopes(x = "X", z = "Z", y = "Y", model = est1)
} # }