Get standardized estimates

Get standardized estimates

Usage

standardized_estimates(object, ...)

Arguments

object

An object of class modsem_da, modsem_mplus, or a parTable of class data.frame

...

Additional arguments passed to other functions

Value

A data.frame with the standardized estimates of the model parameters, in the est column.

Details

For modsem_da, and modsem_mplus objects, the interaction term is not standardized such that var(xz) = 1. The interaction term is not an actual variable in the model, meaning that it does not have a variance. It must therefore be calculated from the other parameters in the model.

Assuming normality and zero-means, the variance is calculated as var(xz) = var(x) * var(z) + cov(x, z)^2. Thus setting the variance of the interaction term to 1 would only be 'correct' if the correlation between x and z is zero. This means that the standardized estimates for the interaction term will be different from those using lavaan, since there the interaction term is an actual latent variable in the model, with a standardized variance of 1.

In modsem_pi the interaction term is standardized such that var(xz) = 1. It is possible to apply a correction to the standardized estimates from modsem_pi, by passing correction = TRUE.

NOTE that the standardized coefficent will be placed in the est column (not est.std) for all models, inluding those from modsem_pi. This is different from the results from lavaan::standardizedSolution where the standardized estimates are placed in the est.std column.

Examples

m1 <- '
  # Outer Model
  X =~ x1 + x2 + x3
  Z =~ z1 + z2 + z3
  Y =~ y1 + y2 + y3

  # Inner Model
  Y ~ X + Z + X:Z
'
# Double centering approach
est_dca <- modsem(m1, oneInt)

standardized_estimates(est_dca) # no correction
#>     lhs op  rhs   est    se       z pvalue ci.lower ci.upper
#> 1     X =~   x1 0.927 0.005 197.805  0.000    0.918    0.936
#> 2     X =~   x2 0.892 0.006 156.888  0.000    0.881    0.904
#> 3     X =~   x3 0.914 0.005 180.631  0.000    0.904    0.924
#> 4     Z =~   z1 0.926 0.005 197.375  0.000    0.917    0.936
#> 5     Z =~   z2 0.899 0.005 164.838  0.000    0.889    0.910
#> 6     Z =~   z3 0.913 0.005 180.409  0.000    0.903    0.923
#> 7     Y =~   y1 0.969 0.002 493.285  0.000    0.966    0.973
#> 8     Y =~   y2 0.955 0.002 390.119  0.000    0.950    0.959
#> 9     Y =~   y3 0.962 0.002 435.067  0.000    0.957    0.966
#> 10    Y  ~    X 0.424 0.016  26.787  0.000    0.393    0.455
#> 11    Y  ~    Z 0.358 0.016  22.372  0.000    0.327    0.390
#> 12    Y  ~   XZ 0.455 0.016  28.971  0.000    0.424    0.486
#> 13   XZ =~ x1z1 0.878 0.007 129.129  0.000    0.865    0.891
#> 14   XZ =~ x2z1 0.836 0.008 105.776  0.000    0.820    0.851
#> 15   XZ =~ x3z1 0.843 0.008 108.821  0.000    0.828    0.858
#> 16   XZ =~ x1z2 0.833 0.008 102.749  0.000    0.817    0.849
#> 17   XZ =~ x2z2 0.804 0.009  90.666  0.000    0.787    0.821
#> 18   XZ =~ x3z2 0.812 0.009  93.853  0.000    0.795    0.829
#> 19   XZ =~ x1z3 0.867 0.007 121.949  0.000    0.853    0.881
#> 20   XZ =~ x2z3 0.829 0.008 102.106  0.000    0.813    0.845
#> 21   XZ =~ x3z3 0.826 0.008 100.718  0.000    0.810    0.842
#> 22 x1z1 ~~ x2z2 0.000 0.000      NA     NA    0.000    0.000
#> 23 x1z1 ~~ x2z3 0.000 0.000      NA     NA    0.000    0.000
#> 24 x1z1 ~~ x3z2 0.000 0.000      NA     NA    0.000    0.000
#> 25 x1z1 ~~ x3z3 0.000 0.000      NA     NA    0.000    0.000
#> 26 x2z1 ~~ x1z2 0.000 0.000      NA     NA    0.000    0.000
#> 27 x1z2 ~~ x2z3 0.000 0.000      NA     NA    0.000    0.000
#> 28 x3z1 ~~ x1z2 0.000 0.000      NA     NA    0.000    0.000
#> 29 x1z2 ~~ x3z3 0.000 0.000      NA     NA    0.000    0.000
#> 30 x2z1 ~~ x1z3 0.000 0.000      NA     NA    0.000    0.000
#> 31 x2z2 ~~ x1z3 0.000 0.000      NA     NA    0.000    0.000
#> 32 x3z1 ~~ x1z3 0.000 0.000      NA     NA    0.000    0.000
#> 33 x3z2 ~~ x1z3 0.000 0.000      NA     NA    0.000    0.000
#> 34 x2z1 ~~ x3z2 0.000 0.000      NA     NA    0.000    0.000
#> 35 x2z1 ~~ x3z3 0.000 0.000      NA     NA    0.000    0.000
#> 36 x3z1 ~~ x2z2 0.000 0.000      NA     NA    0.000    0.000
#> 37 x2z2 ~~ x3z3 0.000 0.000      NA     NA    0.000    0.000
#> 38 x3z1 ~~ x2z3 0.000 0.000      NA     NA    0.000    0.000
#> 39 x3z2 ~~ x2z3 0.000 0.000      NA     NA    0.000    0.000
#> 40 x1z1 ~~ x1z2 0.384 0.018  21.659  0.000    0.349    0.419
#> 41 x1z1 ~~ x1z3 0.393 0.019  20.928  0.000    0.356    0.429
#> 42 x1z1 ~~ x2z1 0.415 0.017  24.499  0.000    0.382    0.448
#> 43 x1z1 ~~ x3z1 0.440 0.017  25.573  0.000    0.406    0.474
#> 44 x1z2 ~~ x1z3 0.367 0.018  20.929  0.000    0.333    0.401
#> 45 x1z2 ~~ x2z2 0.486 0.014  33.698  0.000    0.458    0.515
#> 46 x1z2 ~~ x3z2 0.520 0.014  36.545  0.000    0.492    0.548
#> 47 x1z3 ~~ x2z3 0.450 0.016  28.177  0.000    0.419    0.481
#> 48 x1z3 ~~ x3z3 0.474 0.016  29.999  0.000    0.443    0.505
#> 49 x2z1 ~~ x2z2 0.510 0.014  35.404  0.000    0.481    0.538
#> 50 x2z1 ~~ x2z3 0.542 0.014  38.238  0.000    0.514    0.570
#> 51 x2z1 ~~ x3z1 0.370 0.016  22.694  0.000    0.338    0.402
#> 52 x2z2 ~~ x2z3 0.486 0.015  33.189  0.000    0.457    0.515
#> 53 x2z2 ~~ x3z2 0.456 0.014  31.605  0.000    0.428    0.484
#> 54 x2z3 ~~ x3z3 0.404 0.015  26.242  0.000    0.374    0.434
#> 55 x3z1 ~~ x3z2 0.464 0.015  30.577  0.000    0.434    0.494
#> 56 x3z1 ~~ x3z3 0.507 0.015  34.430  0.000    0.478    0.536
#> 57 x3z2 ~~ x3z3 0.464 0.015  31.361  0.000    0.435    0.493
#> 58   x1 ~~   x1 0.140 0.009  16.115  0.000    0.123    0.157
#> 59   x2 ~~   x2 0.204 0.010  20.047  0.000    0.184    0.223
#> 60   x3 ~~   x3 0.165 0.009  17.875  0.000    0.147    0.183
#> 61   z1 ~~   z1 0.142 0.009  16.321  0.000    0.125    0.159
#> 62   z2 ~~   z2 0.191 0.010  19.481  0.000    0.172    0.210
#> 63   z3 ~~   z3 0.167 0.009  18.042  0.000    0.149    0.185
#> 64   y1 ~~   y1 0.060 0.004  15.807  0.000    0.053    0.068
#> 65   y2 ~~   y2 0.089 0.005  18.988  0.000    0.080    0.098
#> 66   y3 ~~   y3 0.075 0.004  17.723  0.000    0.067    0.084
#> 67 x1z1 ~~ x1z1 0.229 0.012  19.190  0.000    0.206    0.253
#> 68 x2z1 ~~ x2z1 0.301 0.013  22.788  0.000    0.275    0.327
#> 69 x3z1 ~~ x3z1 0.289 0.013  22.109  0.000    0.263    0.315
#> 70 x1z2 ~~ x1z2 0.306 0.014  22.696  0.000    0.280    0.333
#> 71 x2z2 ~~ x2z2 0.353 0.014  24.786  0.000    0.326    0.381
#> 72 x3z2 ~~ x3z2 0.341 0.014  24.264  0.000    0.313    0.368
#> 73 x1z3 ~~ x1z3 0.249 0.012  20.178  0.000    0.224    0.273
#> 74 x2z3 ~~ x2z3 0.313 0.013  23.291  0.000    0.287    0.340
#> 75 x3z3 ~~ x3z3 0.317 0.014  23.421  0.000    0.291    0.344
#> 76    X ~~    X 1.000 0.000      NA     NA    1.000    1.000
#> 77    Z ~~    Z 1.000 0.000      NA     NA    1.000    1.000
#> 78    Y ~~    Y 0.398 0.016  25.401  0.000    0.367    0.429
#> 79   XZ ~~   XZ 1.000 0.000      NA     NA    1.000    1.000
#> 80    X ~~    Z 0.201 0.023   8.786  0.000    0.157    0.246
#> 81    X ~~   XZ 0.015 0.024   0.628  0.530   -0.033    0.063
#> 82    Z ~~   XZ 0.060 0.024   2.462  0.014    0.012    0.108
# standardized_estimates(est_dca, correction = TRUE) # apply correction

if (FALSE) { # \dontrun{
est_lms <- modsem(m1, oneInt, method = "lms")
standardized_estimates(est_lms) # correction not relevant for lms
} # }