Summarize a parameter table from a modsem model.

Summarize a parameter table from a modsem model.

Usage

summarize_partable(
  parTable,
  scientific = FALSE,
  ci = FALSE,
  digits = 3,
  loadings = TRUE,
  regressions = TRUE,
  covariances = TRUE,
  intercepts = TRUE,
  variances = TRUE
)

Arguments

parTable

A parameter table, typically obtained from a modsem model using parameter_estimates or standardized_estimates.

scientific

Logical, whether to print p-values in scientific notation.

ci

Logical, whether to include confidence intervals in the output.

digits

Integer, number of digits to round the estimates to (default is 3).

loadings

Logical, whether to include factor loadings in the output.

regressions

Logical, whether to include regression coefficients in the output.

covariances

Logical, whether to include covariance estimates in the output.

intercepts

Logical, whether to include intercepts in the output.

variances

Logical, whether to include variance estimates in the output.

Value

A summary object containing the parameter table and additional information.

Examples

m1 <- '
  # Outer Model
  X =~ x1 + x2 + x3
  Z =~ z1 + z2 + z3
  Y =~ y1 + y2 + y3

  # Inner Model
  Y ~ X + Z + X:Z
'
# Double centering approach
est_dca <- modsem(m1, oneInt)

std <- standardized_estimates(est_dca, correction = TRUE)
summarize_partable(std)
#> Error in labels[labels != ""] <- paste0("(", labels[labels != ""], ")"): NAs are not allowed in subscripted assignments