summary for modsem objects

summary for modsem objects

summary for modsem objects

summary for modsem objects

Usage

# S3 method for class 'modsem_da'
summary(
  object,
  H0 = is_interaction_model(object),
  verbose = interactive(),
  r.squared = TRUE,
  fit = FALSE,
  adjusted.stat = FALSE,
  digits = 3,
  scientific = FALSE,
  ci = FALSE,
  standardized = FALSE,
  centered = FALSE,
  monte.carlo = FALSE,
  mc.reps = 10000,
  loadings = TRUE,
  regressions = TRUE,
  covariances = TRUE,
  intercepts = TRUE,
  variances = TRUE,
  var.interaction = FALSE,
  ...
)

# S3 method for class 'modsem_mplus'
summary(
  object,
  scientific = FALSE,
  standardized = FALSE,
  ci = FALSE,
  digits = 3,
  loadings = TRUE,
  regressions = TRUE,
  covariances = TRUE,
  intercepts = TRUE,
  variances = TRUE,
  ...
)

# S3 method for class 'modsem_pi'
summary(
  object,
  H0 = is_interaction_model(object),
  r.squared = TRUE,
  adjusted.stat = FALSE,
  digits = 3,
  scientific = FALSE,
  verbose = TRUE,
  ...
)

Arguments

object

modsem object to summarized

H0

Should the baseline model be estimated, and used to produce comparative fit?

verbose

Should messages be printed?

r.squared

Calculate R-squared.

fit

Print additional fit measures.

adjusted.stat

Should sample size corrected/adjustes AIC and BIC be reported?

digits

Number of digits for printed numerical values

scientific

Should scientific format be used for p-values?

ci

print confidence intervals

standardized

standardize estimates

centered

Print mean centered estimates.

monte.carlo

Should Monte Carlo bootstrapped standard errors be used? Only relevant if standardized = TRUE. If FALSE delta method is used instead.

mc.reps

Number of Monte Carlo repetitions. Only relevant if monte.carlo = TRUE, and standardized = TRUE.

loadings

print loadings

regressions

print regressions

covariances

print covariances

intercepts

print intercepts

variances

print variances

var.interaction

If FALSE variances for interaction terms will be removed from the output.

...

arguments passed to lavaan::summary()

Examples

# \dontrun{
m1 <- "
 # Outer Model
 X =~ x1 + x2 + x3
 Y =~ y1 + y2 + y3
 Z =~ z1 + z2 + z3

 # Inner model
 Y ~ X + Z + X:Z
"

est1 <- modsem(m1, oneInt, "qml")
summary(est1, ci = TRUE, scientific = TRUE)
#> 
#> modsem (1.0.17) ended normally after 36 iterations
#> 
#>   Estimator                                           QML
#>   Optimization method                              NLMINB
#>   Number of model parameters                           31
#> 
#>   Number of observations                             2000
#> 
#> Loglikelihood and Information Criteria:
#>   Loglikelihood                                  -1.7e+04
#>   Akaike (AIC)                                    3.5e+04
#>   Bayesian (BIC)                                  3.5e+04
#>  
#> Fit Measures for Baseline Model (H0):
#>                                                  Standard
#>   Chi-square                                      1.8e+01
#>   Degrees of Freedom (Chi-square)                      24
#>   P-value (Chi-square)                            0.82554
#>   RMSEA                                             0e+00
#>                                                          
#>   Loglikelihood                                  -1.8e+04
#>   Akaike (AIC)                                    3.6e+04
#>   Bayesian (BIC)                                  3.6e+04
#>  
#> Comparative Fit to H0 (LRT test):
#>   Loglikelihood change                            3.4e+02
#>   Difference test (D)                             6.8e+02
#>   Degrees of freedom (D)                                1
#>   P-value (D)                                  < 2.22e-16
#>  
#> R-Squared Interaction Model (H1):
#>   Y                                              5.99e-01
#> R-Squared Baseline Model (H0):
#>   Y                                              3.95e-01
#> R-Squared Change (H1 - H0):
#>   Y                                              2.04e-01
#> 
#> Parameter Estimates:
#>   Coefficients                             unstandardized
#>   Information                                    observed
#>   Standard errors                                standard
#>  
#> Latent Variables:
#>                  Estimate  Std.Error  z.value     P(>|z|)  ci.lower  ci.upper
#>   X =~          
#>     x1              1.000                                                    
#>     x2              0.804      0.013   63.895  < 2.22e-16     0.779     0.828
#>     x3              0.914      0.013   67.709  < 2.22e-16     0.888     0.940
#>   Z =~          
#>     z1              1.000                                                    
#>     z2              0.810      0.012   65.078  < 2.22e-16     0.786     0.835
#>     z3              0.881      0.013   67.606  < 2.22e-16     0.856     0.907
#>   Y =~          
#>     y1              1.000                                                    
#>     y2              0.798      0.007  107.529  < 2.22e-16     0.784     0.813
#>     y3              0.899      0.008  112.559  < 2.22e-16     0.884     0.915
#> 
#> Regressions:
#>                  Estimate  Std.Error  z.value     P(>|z|)  ci.lower  ci.upper
#>   Y ~           
#>     X               0.673      0.031   21.672  < 2.22e-16     0.612     0.734
#>     Z               0.569      0.030   18.717  < 2.22e-16     0.509     0.628
#>     X:Z             0.719      0.028   25.831  < 2.22e-16     0.664     0.773
#> 
#> Intercepts:
#>                  Estimate  Std.Error  z.value     P(>|z|)  ci.lower  ci.upper
#>    .x1              1.023      0.024   42.824  < 2.22e-16     0.976     1.070
#>    .x2              1.215      0.020   60.904  < 2.22e-16     1.176     1.255
#>    .x3              0.919      0.022   41.418  < 2.22e-16     0.876     0.963
#>    .z1              1.011      0.024   41.560  < 2.22e-16     0.964     1.059
#>    .z2              1.206      0.020   59.254  < 2.22e-16     1.166     1.245
#>    .z3              0.916      0.022   42.048  < 2.22e-16     0.873     0.958
#>    .y1              1.036      0.033   31.393  < 2.22e-16     0.972     1.101
#>    .y2              1.220      0.027   45.418  < 2.22e-16     1.167     1.272
#>    .y3              0.953      0.030   31.797  < 2.22e-16     0.895     1.012
#> 
#> Covariances:
#>                  Estimate  Std.Error  z.value     P(>|z|)  ci.lower  ci.upper
#>   X ~~          
#>     Z               0.200      0.024    8.242  < 2.22e-16     0.153     0.248
#> 
#> Variances:
#>                  Estimate  Std.Error  z.value     P(>|z|)  ci.lower  ci.upper
#>    .x1              0.158      0.009   18.145  < 2.22e-16     0.141     0.175
#>    .x2              0.162      0.007   23.153  < 2.22e-16     0.148     0.176
#>    .x3              0.165      0.008   20.766  < 2.22e-16     0.149     0.180
#>    .z1              0.167      0.009   18.505  < 2.22e-16     0.149     0.184
#>    .z2              0.160      0.007   22.675  < 2.22e-16     0.146     0.173
#>    .z3              0.158      0.008   20.771  < 2.22e-16     0.143     0.173
#>    .y1              0.160      0.009   18.009  < 2.22e-16     0.142     0.177
#>    .y2              0.154      0.007   22.679  < 2.22e-16     0.141     0.168
#>    .y3              0.164      0.008   20.678  < 2.22e-16     0.148     0.179
#>     X               0.983      0.036   26.975  < 2.22e-16     0.911     1.054
#>     Z               1.018      0.038   26.928  < 2.22e-16     0.943     1.092
#>    .Y               0.980      0.038   25.905  < 2.22e-16     0.906     1.054
#> 
# }