TPB_1SO

A simulated dataset based on the Theory of Planned Behaviour, where INT is a higher order construct of ATT, SN, and PBC.

Examples

tpb <- '
  # First order constructs
  ATT =~ att1 + att2 + att3
  SN  =~ sn1 + sn2 + sn3
  PBC =~ pbc1 + pbc2 + pbc3
  BEH =~ b1 + b2

  # Higher order constructs
  INT =~ ATT + PBC + SN

  # Higher order interaction
  INTxPBC =~ ATT:PBC + SN:PBC + PBC:PBC

  # Structural model
  BEH ~ PBC + INT + INTxPBC
'

# \dontrun{
est_ca <- modsem(tpb, data = TPB_1SO, method = "ca")
summary(est_ca)
#> Estimating baseline model (H0)
#> Warning: Unable to estimate baseline model for models with higher-order interaction terms!
#> modsem (version 1.0.14, approach = ca):
#> 
#> Interaction Model Fit Measures (H1):
#>   Loglikelihood                              -42117.62 
#>   Akaike (AIC)                                84339.24 
#>   Bayesian (BIC)                              84630.48 
#>   Chi-square                                   4246.90 
#>   Degrees of Freedom                               178 
#>   P-value (Chi-square)                           0.000 
#>   RMSEA                                          0.107 
#>   CFI                                            0.890 
#>   SRMR                                           0.135 
#> 
#> R-Squared Interaction Model (H1):
#>   ATT                                            0.609 
#>   PBC                                            0.574 
#>   SN                                             0.607 
#>   ATTPBC                                         0.691 
#>   SNPBC                                          0.701 
#>   PBCPBC                                         0.703 
#>   BEH                                            0.232 
#> 
#> lavaan 0.6-20 ended normally after 450 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        73
#>   Row rank of the constraints matrix                21
#> 
#>   Number of observations                          2000
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                              4246.901
#>   Degrees of freedom                               178
#>   P-value (Chi-square)                           0.000
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   ATT =~                                              
#>     att1   (l_1_A)    1.000                           
#>     att2   (l_2_A)    0.904    0.010   89.441    0.000
#>     att3   (l_3_A)    0.801    0.009   85.442    0.000
#>   SN =~                                               
#>     sn1    (l_1_S)    1.000                           
#>     sn2    (l_2_S)    0.879    0.013   66.744    0.000
#>     sn3    (l_3_S)    0.780    0.012   63.639    0.000
#>   PBC =~                                              
#>     pbc1   (l_1_P)    1.000                           
#>     pbc2   (l_2_P)    0.900    0.007  135.630    0.000
#>     pbc3   (l_3_P)    0.776    0.006  125.110    0.000
#>   BEH =~                                              
#>     b1     (l_1_B)    1.000                           
#>     b2     (l_2_B)    0.863    0.033   26.043    0.000
#>   INT =~                                              
#>     ATT   (l_ATT_)    1.000                           
#>     PBC   (l_PBC_)    0.783    0.030   26.191    0.000
#>     SN     (l_SN_)    0.717    0.027   26.257    0.000
#>   INTxPBC =~                                          
#>     ATTPB (l_ATTP)    1.000                           
#>     SNPBC  (l_SNP)    0.735    0.020   35.914    0.000
#>     PBCPB (l_PBCP)    1.011    0.027   36.926    0.000
#>   ATTPBC =~                                           
#>     att11 (l_11_A)    1.000                           
#>     att22 (l_22_A)    0.813    0.009   87.005    0.000
#>     att33 (l_33_A)    0.621    0.008   82.373    0.000
#>   SNPBC =~                                            
#>     sn1p1 (l_11_S)    1.000                           
#>     sn2p2 (l_22_S)    0.792    0.012   68.052    0.000
#>     sn3p3 (l_33_S)    0.605    0.009   64.723    0.000
#>   PBCPBC =~                                           
#>     pbc11 (l_11_P)    1.000                           
#>     pbc22 (l_22_P)    0.810    0.012   67.815    0.000
#>     pbc33 (l_33_P)    0.602    0.010   62.555    0.000
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   BEH ~                                               
#>     PBC     (G_PB)    0.249    0.052    4.775    0.000
#>     INT   (G_INT_)    0.160    0.056    2.838    0.005
#>     INTPB (G_INTP)    0.221    0.016   13.492    0.000
#> 
#> Covariances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   INT ~~                                              
#>     INTxPBC (C_IN)   -0.019    0.025   -0.759    0.448
#>  .att1pbc1 ~~                                         
#>    .att2pb2           0.000                           
#>    .att3pb3           0.000                           
#>  .att2pbc2 ~~                                         
#>    .att3pb3           0.000                           
#>  .sn1pbc1 ~~                                          
#>    .sn2pbc2           0.000                           
#>    .sn3pbc3           0.000                           
#>  .sn2pbc2 ~~                                          
#>    .sn3pbc3           0.000                           
#>  .pbc1pbc1 ~~                                         
#>    .pbc2pb2           0.000                           
#>    .pbc3pb3           0.000                           
#>  .pbc2pbc2 ~~                                         
#>    .pbc3pb3           0.000                           
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .ATTPBC  (M_AT)    0.422    0.013   32.832    0.000
#>    .SNPBC   (M_SN)    0.302    0.010   30.547    0.000
#>    .PBCPBC  (M_PB)    0.575    0.011   51.527    0.000
#>    .att1              1.001    0.023   44.025    0.000
#>    .att2              1.008    0.021   47.861    0.000
#>    .att3              1.002    0.019   52.974    0.000
#>    .sn1               0.974    0.018   55.116    0.000
#>    .sn2               0.982    0.016   60.802    0.000
#>    .sn3               0.991    0.015   67.883    0.000
#>    .pbc1              0.983    0.021   47.193    0.000
#>    .pbc2              0.988    0.020   49.207    0.000
#>    .pbc3              0.998    0.018   54.376    0.000
#>    .b1                1.150    0.020   57.082    0.000
#>    .b2                1.132    0.018   61.428    0.000
#>    .att1pb1           0.391    0.038   10.185    0.000
#>    .att2pb2           0.330    0.032   10.249    0.000
#>    .att3pb3           0.256    0.025   10.137    0.000
#>    .sn1pbc1           0.262    0.029    9.112    0.000
#>    .sn2pbc2           0.226    0.024    9.516    0.000
#>    .sn3pbc3           0.177    0.019    9.454    0.000
#>    .pbc1pb1           0.553    0.038   14.525    0.000
#>    .pbc2pb2           0.501    0.032   15.560    0.000
#>    .pbc3pb3           0.421    0.025   16.843    0.000
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .ATT   (Vr_ATT)    0.346    0.023   15.014    0.000
#>    .SN     (Vr_SN)    0.179    0.013   13.986    0.000
#>    .PBC   (Vr_PBC)    0.245    0.016   15.052    0.000
#>    .BEH     (Z_BE)    0.513    0.029   18.005    0.000
#>     INT   (Vr_INT)    0.539    0.027   19.889    0.000
#>     INTPB (V_INTP)    1.535    0.076   20.148    0.000
#>    .ATTPB (V_ATTP)    0.687    0.022   30.946    0.000
#>    .SNPBC  (V_SNP)    0.354    0.013   27.980    0.000
#>    .PBCPB (V_PBCP)    0.662    0.026   25.764    0.000
#>    .att1   (Vr_t1)    0.150    0.008   18.624    0.000
#>    .att2   (Vr_t2)    0.165    0.007   22.170    0.000
#>    .att3   (Vr_t3)    0.147    0.006   23.582    0.000
#>    .sn1    (Vr_s1)    0.168    0.008   20.991    0.000
#>    .sn2    (Vr_s2)    0.168    0.007   24.105    0.000
#>    .sn3    (Vr_s3)    0.149    0.006   25.303    0.000
#>    .pbc1   (Vr_p1)    0.293    0.009   30.965    0.000
#>    .pbc2   (Vr_p2)    0.340    0.009   38.979    0.000
#>    .pbc3   (Vr_p3)    0.327    0.007   44.263    0.000
#>    .b1     (Vr_b1)    0.144    0.024    6.051    0.000
#>    .b2     (Vr_b2)    0.181    0.018    9.880    0.000
#>    .att11 (Vr_t11)    0.389    0.011   34.052    0.000
#>    .att22 (Vr_t22)    0.378    0.010   39.469    0.000
#>    .att33 (Vr_t33)    0.285    0.007   41.921    0.000
#>    .sn1p1 (Vr_s11)    0.279    0.008   35.024    0.000
#>    .sn2p2 (Vr_s22)    0.256    0.006   39.790    0.000
#>    .sn3p3 (Vr_s33)    0.191    0.005   41.982    0.000
#>    .pbc11 (Vr_p11)    0.423    0.015   28.081    0.000
#>    .pbc22 (Vr_p22)    0.432    0.013   33.193    0.000
#>    .pbc33 (Vr_p33)    0.334    0.009   35.936    0.000
#> 

est_dblcent  <- modsem(tpb, data = TPB_1SO, method = "dblcent")
summary(est_dblcent)
#> Estimating baseline model (H0)
#> Warning: Unable to estimate baseline model for models with higher-order interaction terms!
#> modsem (version 1.0.14, approach = dblcent):
#> 
#> Interaction Model Fit Measures (H1):
#>   Loglikelihood                              -42695.54 
#>   Akaike (AIC)                                85803.08 
#>   Bayesian (BIC)                              86956.87 
#>   Chi-square                                    565.72 
#>   Degrees of Freedom                               424 
#>   P-value (Chi-square)                           0.000 
#>   RMSEA                                          0.013 
#>   CFI                                            0.999 
#>   SRMR                                           0.018 
#> 
#> R-Squared Interaction Model (H1):
#>   ATT                                            0.719 
#>   PBC                                            0.690 
#>   SN                                             0.712 
#>   ATTPBC                                         0.811 
#>   SNPBC                                          0.786 
#>   PBCPBC                                         0.812 
#>   BEH                                            0.276 
#> 
#> lavaan 0.6-20 ended normally after 543 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                       206
#> 
#>   Number of observations                          2000
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                               565.724
#>   Degrees of freedom                               424
#>   P-value (Chi-square)                           0.000
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   ATT =~                                              
#>     att1              1.000                           
#>     att2              0.909    0.011   82.768    0.000
#>     att3              0.802    0.010   80.425    0.000
#>   SN =~                                               
#>     sn1               1.000                           
#>     sn2               0.891    0.016   54.218    0.000
#>     sn3               0.790    0.015   52.448    0.000
#>   PBC =~                                              
#>     pbc1              1.000                           
#>     pbc2              0.909    0.014   66.750    0.000
#>     pbc3              0.793    0.013   63.370    0.000
#>   BEH =~                                              
#>     b1                1.000                           
#>     b2                0.865    0.029   30.297    0.000
#>   INT =~                                              
#>     ATT               1.000                           
#>     PBC               0.819    0.024   33.976    0.000
#>     SN                0.702    0.021   33.484    0.000
#>   INTxPBC =~                                          
#>     ATTPBC            1.000                           
#>     SNPBC             0.717    0.017   41.167    0.000
#>     PBCPBC            0.971    0.022   43.199    0.000
#>   ATTPBC =~                                           
#>     att1pbc1          1.000                           
#>     att2pbc1          0.897    0.010   92.034    0.000
#>     att3pbc1          0.807    0.009   90.390    0.000
#>     att1pbc2          0.908    0.010   87.991    0.000
#>     att2pbc2          0.813    0.012   65.243    0.000
#>     att3pbc2          0.733    0.011   65.091    0.000
#>     att1pbc3          0.774    0.010   81.181    0.000
#>     att2pbc3          0.694    0.011   60.977    0.000
#>     att3pbc3          0.630    0.010   61.880    0.000
#>   SNPBC =~                                            
#>     sn1pbc1           1.000                           
#>     sn2pbc1           0.884    0.014   63.726    0.000
#>     sn3pbc1           0.782    0.013   61.274    0.000
#>     sn1pbc2           0.907    0.011   84.140    0.000
#>     sn2pbc2           0.802    0.015   52.856    0.000
#>     sn3pbc2           0.702    0.014   51.227    0.000
#>     sn1pbc3           0.787    0.010   79.471    0.000
#>     sn2pbc3           0.686    0.014   50.281    0.000
#>     sn3pbc3           0.613    0.012   50.032    0.000
#>   PBCPBC =~                                           
#>     pbc1pbc1          1.000                           
#>     pbc2pbc1          0.909    0.010   89.205    0.000
#>     pbc3pbc1          0.782    0.009   85.688    0.000
#>     pbc2pbc2          0.827    0.017   47.996    0.000
#>     pbc3pbc2          0.710    0.013   54.545    0.000
#>     pbc3pbc3          0.616    0.014   45.151    0.000
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   BEH ~                                               
#>     PBC               0.188    0.043    4.337    0.000
#>     INT               0.188    0.045    4.173    0.000
#>     INTxPBC           0.217    0.016   13.789    0.000
#> 
#> Covariances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>  .att1pbc1 ~~                                         
#>    .att1pbc2          0.122    0.008   15.676    0.000
#>    .att1pbc3          0.108    0.007   15.545    0.000
#>    .att2pbc1          0.201    0.013   16.047    0.000
#>    .att3pbc1          0.178    0.011   16.033    0.000
#>    .pbc1pbc1          0.256    0.018   13.950    0.000
#>    .pbc2pbc1          0.118    0.009   12.940    0.000
#>    .pbc3pbc1          0.098    0.008   12.236    0.000
#>    .sn1pbc1           0.116    0.009   13.217    0.000
#>    .sn2pbc1           0.100    0.008   12.831    0.000
#>    .sn3pbc1           0.097    0.007   13.331    0.000
#>  .att1pbc2 ~~                                         
#>    .att1pbc3          0.099    0.006   15.365    0.000
#>    .att2pbc2          0.205    0.011   18.285    0.000
#>    .att3pbc2          0.189    0.010   18.718    0.000
#>    .pbc2pbc1          0.136    0.009   15.063    0.000
#>    .pbc2pbc2          0.241    0.015   15.591    0.000
#>    .pbc3pbc2          0.106    0.007   14.278    0.000
#>    .sn1pbc2           0.123    0.008   15.415    0.000
#>    .sn2pbc2           0.098    0.007   13.955    0.000
#>    .sn3pbc2           0.098    0.007   15.102    0.000
#>  .att1pbc3 ~~                                         
#>    .att2pbc3          0.232    0.010   22.926    0.000
#>    .att3pbc3          0.199    0.009   22.457    0.000
#>    .pbc3pbc1          0.144    0.008   18.050    0.000
#>    .pbc3pbc2          0.130    0.007   17.741    0.000
#>    .pbc3pbc3          0.232    0.012   18.949    0.000
#>    .sn1pbc3           0.125    0.007   17.887    0.000
#>    .sn2pbc3           0.107    0.006   17.035    0.000
#>    .sn3pbc3           0.103    0.006   17.869    0.000
#>  .att2pbc1 ~~                                         
#>    .att2pbc2          0.156    0.008   20.447    0.000
#>    .att2pbc3          0.144    0.007   20.765    0.000
#>    .att3pbc1          0.163    0.010   16.025    0.000
#>    .pbc1pbc1          0.238    0.017   14.066    0.000
#>    .pbc2pbc1          0.105    0.008   12.572    0.000
#>    .pbc3pbc1          0.088    0.007   11.960    0.000
#>    .sn1pbc1           0.111    0.008   13.687    0.000
#>    .sn2pbc1           0.093    0.007   12.940    0.000
#>    .sn3pbc1           0.092    0.007   13.718    0.000
#>  .att2pbc2 ~~                                         
#>    .att2pbc3          0.130    0.006   20.314    0.000
#>    .att3pbc2          0.169    0.009   18.418    0.000
#>    .pbc2pbc1          0.121    0.008   14.699    0.000
#>    .pbc2pbc2          0.223    0.014   15.709    0.000
#>    .pbc3pbc2          0.097    0.007   14.256    0.000
#>    .sn1pbc2           0.108    0.007   14.776    0.000
#>    .sn2pbc2           0.090    0.006   14.023    0.000
#>    .sn3pbc2           0.085    0.006   14.369    0.000
#>  .att2pbc3 ~~                                         
#>    .att3pbc3          0.182    0.008   22.286    0.000
#>    .pbc3pbc1          0.129    0.007   17.566    0.000
#>    .pbc3pbc2          0.116    0.007   17.202    0.000
#>    .pbc3pbc3          0.214    0.011   18.903    0.000
#>    .sn1pbc3           0.116    0.006   17.956    0.000
#>    .sn2pbc3           0.100    0.006   17.266    0.000
#>    .sn3pbc3           0.093    0.005   17.549    0.000
#>  .att3pbc1 ~~                                         
#>    .att3pbc2          0.132    0.006   20.977    0.000
#>    .att3pbc3          0.118    0.006   20.694    0.000
#>    .pbc1pbc1          0.205    0.015   13.672    0.000
#>    .pbc2pbc1          0.093    0.007   12.446    0.000
#>    .pbc3pbc1          0.078    0.007   11.904    0.000
#>    .sn1pbc1           0.098    0.007   13.636    0.000
#>    .sn2pbc1           0.084    0.006   13.195    0.000
#>    .sn3pbc1           0.082    0.006   13.824    0.000
#>  .att3pbc2 ~~                                         
#>    .att3pbc3          0.105    0.005   20.231    0.000
#>    .pbc2pbc1          0.115    0.007   15.475    0.000
#>    .pbc2pbc2          0.216    0.013   16.714    0.000
#>    .pbc3pbc2          0.091    0.006   14.907    0.000
#>    .sn1pbc2           0.096    0.007   14.798    0.000
#>    .sn2pbc2           0.080    0.006   13.856    0.000
#>    .sn3pbc2           0.077    0.005   14.399    0.000
#>  .att3pbc3 ~~                                         
#>    .pbc3pbc1          0.116    0.007   17.724    0.000
#>    .pbc3pbc2          0.106    0.006   17.648    0.000
#>    .pbc3pbc3          0.186    0.010   18.584    0.000
#>    .sn1pbc3           0.098    0.006   17.232    0.000
#>    .sn2pbc3           0.087    0.005   16.906    0.000
#>    .sn3pbc3           0.080    0.005   17.060    0.000
#>  .pbc1pbc1 ~~                                         
#>    .pbc2pbc1          0.283    0.017   16.582    0.000
#>    .pbc3pbc1          0.251    0.015   16.550    0.000
#>  .sn1pbc1 ~~                                          
#>    .pbc1pbc1          0.201    0.014   14.829    0.000
#>  .sn2pbc1 ~~                                          
#>    .pbc1pbc1          0.171    0.012   14.066    0.000
#>  .sn3pbc1 ~~                                          
#>    .pbc1pbc1          0.163    0.011   14.496    0.000
#>  .pbc2pbc1 ~~                                         
#>    .pbc2pbc2          0.299    0.015   19.338    0.000
#>    .pbc3pbc1          0.111    0.007   15.091    0.000
#>    .pbc3pbc2          0.130    0.007   17.842    0.000
#>  .sn1pbc1 ~~                                          
#>    .pbc2pbc1          0.089    0.007   13.195    0.000
#>  .sn1pbc2 ~~                                          
#>    .pbc2pbc1          0.088    0.007   13.471    0.000
#>  .sn2pbc1 ~~                                          
#>    .pbc2pbc1          0.079    0.006   13.005    0.000
#>  .sn2pbc2 ~~                                          
#>    .pbc2pbc1          0.080    0.006   13.607    0.000
#>  .sn3pbc1 ~~                                          
#>    .pbc2pbc1          0.071    0.006   12.854    0.000
#>  .sn3pbc2 ~~                                          
#>    .pbc2pbc1          0.077    0.005   14.169    0.000
#>  .pbc2pbc2 ~~                                         
#>    .pbc3pbc2          0.241    0.013   18.860    0.000
#>  .sn1pbc2 ~~                                          
#>    .pbc2pbc2          0.166    0.011   14.639    0.000
#>  .sn2pbc2 ~~                                          
#>    .pbc2pbc2          0.152    0.010   14.913    0.000
#>  .sn3pbc2 ~~                                          
#>    .pbc2pbc2          0.137    0.009   14.766    0.000
#>  .pbc3pbc1 ~~                                         
#>    .pbc3pbc2          0.146    0.007   21.068    0.000
#>    .pbc3pbc3          0.263    0.012   22.316    0.000
#>  .sn1pbc1 ~~                                          
#>    .pbc3pbc1          0.078    0.006   12.972    0.000
#>  .sn1pbc3 ~~                                          
#>    .pbc3pbc1          0.091    0.006   15.897    0.000
#>  .sn2pbc1 ~~                                          
#>    .pbc3pbc1          0.067    0.005   12.407    0.000
#>  .sn2pbc3 ~~                                          
#>    .pbc3pbc1          0.083    0.005   16.014    0.000
#>  .sn3pbc1 ~~                                          
#>    .pbc3pbc1          0.062    0.005   12.611    0.000
#>  .sn3pbc3 ~~                                          
#>    .pbc3pbc1          0.079    0.005   16.645    0.000
#>  .pbc3pbc2 ~~                                         
#>    .pbc3pbc3          0.236    0.011   21.902    0.000
#>  .sn1pbc2 ~~                                          
#>    .pbc3pbc2          0.075    0.005   13.720    0.000
#>  .sn1pbc3 ~~                                          
#>    .pbc3pbc2          0.083    0.005   15.837    0.000
#>  .sn2pbc2 ~~                                          
#>    .pbc3pbc2          0.067    0.005   13.755    0.000
#>  .sn2pbc3 ~~                                          
#>    .pbc3pbc2          0.077    0.005   16.119    0.000
#>  .sn3pbc2 ~~                                          
#>    .pbc3pbc2          0.062    0.004   13.867    0.000
#>  .sn3pbc3 ~~                                          
#>    .pbc3pbc2          0.071    0.004   16.237    0.000
#>  .sn1pbc3 ~~                                          
#>    .pbc3pbc3          0.153    0.009   17.396    0.000
#>  .sn2pbc3 ~~                                          
#>    .pbc3pbc3          0.142    0.008   17.662    0.000
#>  .sn3pbc3 ~~                                          
#>    .pbc3pbc3          0.129    0.007   17.642    0.000
#>  .sn1pbc1 ~~                                          
#>    .sn1pbc2           0.163    0.009   19.134    0.000
#>    .sn1pbc3           0.139    0.007   18.626    0.000
#>    .sn2pbc1           0.099    0.006   15.359    0.000
#>    .sn3pbc1           0.089    0.006   15.027    0.000
#>  .sn1pbc2 ~~                                          
#>    .sn1pbc3           0.126    0.007   18.289    0.000
#>    .sn2pbc2           0.101    0.006   17.396    0.000
#>    .sn3pbc2           0.096    0.005   18.005    0.000
#>  .sn1pbc3 ~~                                          
#>    .sn2pbc3           0.106    0.005   20.768    0.000
#>    .sn3pbc3           0.095    0.005   20.530    0.000
#>  .sn2pbc1 ~~                                          
#>    .sn2pbc2           0.136    0.007   19.542    0.000
#>    .sn2pbc3           0.122    0.006   19.630    0.000
#>    .sn3pbc1           0.079    0.005   14.861    0.000
#>  .sn2pbc2 ~~                                          
#>    .sn2pbc3           0.109    0.006   19.185    0.000
#>    .sn3pbc2           0.083    0.005   17.507    0.000
#>  .sn2pbc3 ~~                                          
#>    .sn3pbc3           0.084    0.004   20.140    0.000
#>  .sn3pbc1 ~~                                          
#>    .sn3pbc2           0.122    0.006   20.972    0.000
#>    .sn3pbc3           0.107    0.005   20.599    0.000
#>  .sn3pbc2 ~~                                          
#>    .sn3pbc3           0.096    0.005   20.403    0.000
#>   INT ~~                                              
#>     INTxPBC          -0.038    0.033   -1.147    0.252
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .att1              0.152    0.008   18.217    0.000
#>    .att2              0.168    0.008   21.468    0.000
#>    .att3              0.147    0.006   22.692    0.000
#>    .sn1               0.171    0.009   19.842    0.000
#>    .sn2               0.165    0.008   21.851    0.000
#>    .sn3               0.151    0.006   23.314    0.000
#>    .pbc1              0.162    0.009   18.801    0.000
#>    .pbc2              0.167    0.008   21.287    0.000
#>    .pbc3              0.159    0.007   23.501    0.000
#>    .b1                0.145    0.021    6.812    0.000
#>    .b2                0.180    0.017   10.895    0.000
#>    .att1pbc1          0.377    0.017   22.325    0.000
#>    .att2pbc1          0.381    0.015   25.514    0.000
#>    .att3pbc1          0.314    0.012   25.965    0.000
#>    .att1pbc2          0.371    0.015   24.510    0.000
#>    .att2pbc2          0.360    0.013   27.082    0.000
#>    .att3pbc2          0.298    0.011   27.636    0.000
#>    .att1pbc3          0.364    0.013   27.887    0.000
#>    .att2pbc3          0.357    0.012   30.624    0.000
#>    .att3pbc3          0.277    0.009   30.093    0.000
#>    .sn1pbc1           0.315    0.013   25.133    0.000
#>    .sn2pbc1           0.261    0.010   25.532    0.000
#>    .sn3pbc1           0.232    0.009   26.660    0.000
#>    .sn1pbc2           0.297    0.011   26.676    0.000
#>    .sn2pbc2           0.240    0.009   26.744    0.000
#>    .sn3pbc2           0.211    0.007   28.273    0.000
#>    .sn1pbc3           0.253    0.009   28.195    0.000
#>    .sn2pbc3           0.222    0.008   29.257    0.000
#>    .sn3pbc3           0.183    0.006   29.644    0.000
#>    .pbc1pbc1          0.680    0.036   19.050    0.000
#>    .pbc2pbc1          0.310    0.013   23.088    0.000
#>    .pbc3pbc1          0.287    0.011   25.366    0.000
#>    .pbc2pbc2          0.611    0.028   21.890    0.000
#>    .pbc3pbc2          0.264    0.010   26.146    0.000
#>    .pbc3pbc3          0.470    0.019   24.836    0.000
#>    .ATT               0.389    0.025   15.723    0.000
#>    .SN                0.199    0.014   14.729    0.000
#>    .PBC               0.300    0.018   16.350    0.000
#>    .BEH               0.516    0.027   19.381    0.000
#>     INT               0.994    0.049   20.231    0.000
#>     INTxPBC           1.597    0.073   21.909    0.000
#>    .ATTPBC            0.373    0.025   14.700    0.000
#>    .SNPBC             0.224    0.015   15.118    0.000
#>    .PBCPBC            0.349    0.025   13.718    0.000
#> 
# }