Bootstrap a modsem Model

A generic interface for parametric and non-parametric bootstrap procedures for structural equation models estimated with the modsem ecosystem. The function dispatches on the class of model; currently dedicated methods exist for modsem_pi (product‑indicator approach) and modsem_da (distributional‑analytic approach).

Usage

bootstrap_modsem(model = modsem, FUN, ...)

# S3 method for class 'modsem_pi'
bootstrap_modsem(model, FUN, ...)

# S3 method for class 'modsem_da'
bootstrap_modsem(
  model,
  FUN = "coef",
  R = 1000L,
  P.max = 1e+05,
  type = c("nonparametric", "parameteric"),
  verbose = interactive(),
  calc.se = FALSE,
  optimize = FALSE,
  ...
)

# S3 method for class '`function`'
bootstrap_modsem(
  model = modsem,
  FUN = "coef",
  data,
  R = 1000L,
  verbose = interactive(),
  FUN.args = list(),
  ...
)

Arguments

model

A fitted modsem object, or a function to be bootstrapped (e.g., modsem, modsem_da and modsem_pi)

FUN

A function that returns the statistic of interest when applied to a fitted model. The function must accept a single argument, the model object, and should ideally return a numeric vector; see Value.

...

Additional arguments forwarded to lavaan::bootstrapLavaan for modsem_pi objects, or modsem_da for modsem_da objects.

R

number of bootstrap replicates.

P.max

ceiling for the simulated population size.

type

Bootstrap flavour, see Details.

verbose

Should progress information be printed to the console?

calc.se

Should standard errors for each replicate. Defaults to FALSE.

optimize

Should starting values be re-optimized for each replicate. Defaults to FALSE.

data

Dataset to be resampled.

FUN.args

Arguments passed to FUN

Value

Depending on the return type of FUN either

numeric

A matrix with R rows (bootstrap replicates) and as many columns as length(FUN(model)).

other

A list of length R; each element is the raw output of FUN. NOTE: Only applies for modsem_da objects

Details

A thin wrapper around lavaan::bootstrapLavaan() that performs the necessary book‑keeping so that FUN receives a fully‑featured modsem_pi object—rather than a bare lavaan fit—at every iteration.

The function internally resamples the observed data (non-parametric case) or simulates from the estimated parameter table (parametric case), feeds the sample to modsem_da, evaluates FUN on the refitted object and finally collates the results.

This is a more general version of boostrap_modsem for bootstrapping modsem functions, not modsem objects. model is now a function to be boostrapped, and ... are now passed to the function (model), not FUN. To pass arguments to FUN use FUN.args.

Methods (by class)

bootstrap_modsem(modsem_pi): Bootstrap a modsem_pi model by delegating to bootstrapLavaan.

bootstrap_modsem(modsem_da): Parametric or non-parametric bootstrap for modsem_da models.

bootstrap_modsem(`function`): Non-parametric bootstrap of modsem functions

See also

Examples

m1 <- '
  X =~ x1 + x2
  Z =~ z1 + z2
  Y =~ y1 + y2

  Y ~ X + Z + X:Z
'

fit_pi <- modsem(m1, oneInt)
bootstrap_modsem(fit_pi, FUN = coef, R = 10L)
#>           X=~x2     Z=~z2     Y=~y2       Y~X       Y~Z      Y~XZ  XZ=~x2z1
#>  [1,] 0.8884237 0.8262879 0.7797705 0.7407114 0.5918870 0.6707317 0.8246330
#>  [2,] 0.8347994 0.8233353 0.8025760 0.6824978 0.6016801 0.6865373 0.8116811
#>  [3,] 0.8206546 0.8518766 0.8007877 0.6971329 0.5362172 0.6820370 0.8238181
#>  [4,] 0.8189006 0.8113157 0.8325491 0.6869278 0.5782950 0.7256083 0.7789803
#>  [5,] 0.7947128 0.8558326 0.7924540 0.7333650 0.6106507 0.7282757 0.8221329
#>  [6,] 0.8254009 0.8749817 0.7929519 0.6579637 0.6131392 0.7361599 0.8396432
#>  [7,] 0.8332576 0.7692559 0.8131625 0.7067837 0.5708261 0.7091723 0.8355242
#>  [8,] 0.8619048 0.8171167 0.7882294 0.7320317 0.6299300 0.7734579 0.8497304
#>  [9,] 0.8272198 0.8581544 0.8065377 0.7420436 0.5475846 0.7715898 0.7902094
#> [10,] 0.8325728 0.8515355 0.8051629 0.6431409 0.6255022 0.7367125 0.8150739
#>        XZ=~x1z2  XZ=~x2z2 x1z1~~x1z2 x1z1~~x2z1 x1z2~~x2z2 x2z1~~x2z2    x1~~x1
#>  [1,] 0.7706057 0.6378158  0.1321030 0.12470048 0.13859185  0.1236854 0.2695500
#>  [2,] 0.7620211 0.6213905  0.1510716 0.08864615 0.15350198  0.1165963 0.2064331
#>  [3,] 0.8140617 0.6704431  0.1392288 0.12953526 0.12479679  0.1391153 0.2037414
#>  [4,] 0.7923717 0.6249382  0.1028956 0.14624920 0.13305408  0.1534371 0.2024947
#>  [5,] 0.8138680 0.6766186  0.1016855 0.17156133 0.13094839  0.1542808 0.1466010
#>  [6,] 0.7600386 0.6436328  0.1071415 0.07537641 0.17072580  0.1236519 0.1713316
#>  [7,] 0.7848949 0.6607577  0.1366991 0.11834861 0.13430508  0.1109281 0.2181188
#>  [8,] 0.8331755 0.7173464  0.1227065 0.15731268 0.08557116  0.1282096 0.2078369
#>  [9,] 0.8248556 0.6578495  0.1125763 0.16228584 0.10511858  0.1216584 0.2034808
#> [10,] 0.8145436 0.6751385  0.1228619 0.15560824 0.12230915  0.1129503 0.1708143
#>          x2~~x2    z1~~z1    z2~~z2    y1~~y1    y2~~y2 x1z1~~x1z1 x2z1~~x2z1
#>  [1,] 0.1068363 0.2120969 0.1364207 0.1522135 0.1715433  0.3467576  0.2808781
#>  [2,] 0.1326718 0.1849395 0.1475439 0.1745294 0.1382708  0.3035680  0.2380796
#>  [3,] 0.1548883 0.1971554 0.1424606 0.1466531 0.1517053  0.3601218  0.2922566
#>  [4,] 0.1458213 0.1939336 0.1492024 0.1856174 0.1279031  0.3284856  0.3442310
#>  [5,] 0.1768187 0.2159352 0.1325914 0.1723244 0.1427089  0.3648349  0.3440650
#>  [6,] 0.1504944 0.2341332 0.1190925 0.1613587 0.1484900  0.2417272  0.2421190
#>  [7,] 0.1401831 0.1358201 0.1775705 0.1744732 0.1359542  0.3335131  0.2663161
#>  [8,] 0.1369614 0.2094457 0.1333073 0.1566526 0.1512443  0.3620863  0.3218547
#>  [9,] 0.1455015 0.2406770 0.1118286 0.2188779 0.1184675  0.3601850  0.3186712
#> [10,] 0.1532201 0.2130209 0.1350956 0.1946312 0.1367605  0.3495799  0.2831961
#>       x1z2~~x1z2 x2z2~~x2z2      X~~X      Z~~Z      Y~~Y    XZ~~XZ      X~~Z
#>  [1,]  0.3058370  0.2433937 0.9250376 0.9509946 1.0404153 0.9839416 0.1510574
#>  [2,]  0.3521438  0.2414933 0.9325914 1.0209942 0.9759887 1.1139648 0.1684377
#>  [3,]  0.3048880  0.2515079 0.9511693 0.9631768 0.9344947 1.1739524 0.1752234
#>  [4,]  0.2790108  0.2581793 0.9726406 1.0246144 0.9579877 1.1140874 0.1988152
#>  [5,]  0.2749447  0.2623786 1.0739408 0.9043633 0.9671002 0.9805299 0.1828608
#>  [6,]  0.3181851  0.2738636 0.9577705 0.9431554 0.9825834 1.0354563 0.1617813
#>  [7,]  0.3089064  0.2212350 0.8914256 1.0704986 0.9203757 0.9886425 0.1810837
#>  [8,]  0.2373762  0.1982495 0.8862861 1.0080460 0.9482141 0.8543592 0.2248769
#>  [9,]  0.2559218  0.2049166 0.9444809 0.9177360 0.9206756 0.9224049 0.1949300
#> [10,]  0.2906092  0.2260151 0.9269864 0.9611296 0.9645808 0.8736839 0.1970902
#>               X~~XZ      Z~~XZ
#>  [1,]  0.0045452923 0.06431493
#>  [2,] -0.0094172460 0.08608002
#>  [3,]  0.0412521613 0.13538289
#>  [4,] -0.0004327796 0.06455298
#>  [5,] -0.0391503620 0.04776190
#>  [6,]  0.0258802984 0.03401701
#>  [7,]  0.0028456548 0.05461465
#>  [8,]  0.0450855234 0.02570986
#>  [9,] -0.0242799280 0.04269811
#> [10,]  0.0395353338 0.04479171
#> attr(,"error.idx")
#> integer(0)
#> attr(,"nonadmissible")
#> integer(0)
#> attr(,"seed")
#> [1] 80750137

m1 <- '
  X =~ x1 + x2
  Z =~ z1 + z2
  Y =~ y1 + y2

  Y ~ X + Z + X:Z
'

# \dontrun{
fit_lms <- modsem(m1, oneInt, method = "lms")
bootstrap_modsem(fit_lms, FUN = coef, R = 10L)
#>           X=~x2     Z=~z2     Y=~y2     x1~1     x2~1      z1~1     z2~1
#>  [1,] 0.8040139 0.7974103 0.8112692 0.995468 1.200687 0.9580069 1.166683
#>  [2,] 0.8333570 0.8105357 0.8214405 1.080927 1.281127 1.0254874 1.201534
#>  [3,] 0.8382835 0.7968939 0.8100909 1.008874 1.203067 1.0059120 1.195529
#>  [4,] 0.8141571 0.8216341 0.8042756 1.024376 1.225203 1.0294250 1.222931
#>  [5,] 0.8004355 0.8260245 0.8099914 1.016241 1.211890 0.9839395 1.173593
#>  [6,] 0.8350461 0.8227698 0.8016336 1.005548 1.207825 1.0225603 1.218247
#>  [7,] 0.8062273 0.7993492 0.8142134 1.033722 1.227224 1.0021301 1.198290
#>  [8,] 0.7979379 0.7972658 0.8112360 1.047589 1.211627 0.9941529 1.212785
#>  [9,] 0.8069447 0.8450427 0.8017229 1.027089 1.225065 1.0123343 1.214047
#> [10,] 0.8053960 0.7867124 0.8110293 1.017793 1.213455 1.0102403 1.215485
#>            y1~1     y2~1    x1~~x1    x2~~x2    z1~~z1    z2~~z2    y1~~y1
#>  [1,] 0.9587829 1.153548 0.1669537 0.1485408 0.1665049 0.1557491 0.1991528
#>  [2,] 1.0706562 1.265233 0.1777581 0.1379251 0.1803637 0.1429009 0.2018495
#>  [3,] 1.0459771 1.224976 0.1922475 0.1330876 0.1557783 0.1582530 0.1867995
#>  [4,] 1.0768132 1.254750 0.1607736 0.1455273 0.1654686 0.1507847 0.1878990
#>  [5,] 1.0124864 1.202901 0.1607397 0.1540014 0.1800556 0.1537090 0.1908705
#>  [6,] 0.9949927 1.189992 0.1975315 0.1492749 0.1701868 0.1438058 0.1911888
#>  [7,] 1.0230628 1.200277 0.1588626 0.1576367 0.1564142 0.1639070 0.1796361
#>  [8,] 1.0646520 1.238937 0.1873205 0.1404794 0.1657468 0.1680423 0.2040134
#>  [9,] 1.0504857 1.233826 0.1787924 0.1553563 0.1792628 0.1436033 0.1807134
#> [10,] 1.0165742 1.197971 0.1564873 0.1643630 0.1537288 0.1540249 0.2131539
#>          y2~~y2      X~~X      X~~Z      Z~~Z      Y~~Y       Y~X       Y~Z
#>  [1,] 0.1439571 0.9935415 0.2208777 1.0539382 0.9911098 0.6347926 0.5326612
#>  [2,] 0.1327457 0.9052853 0.1796990 1.0740032 0.9081954 0.7409488 0.5697885
#>  [3,] 0.1301512 0.9162585 0.1941018 1.0875749 0.9557482 0.6609089 0.5578724
#>  [4,] 0.1339169 0.9766979 0.1978818 0.9788427 0.9571253 0.6586689 0.6294172
#>  [5,] 0.1337863 0.9460811 0.2116461 1.0418668 0.9445864 0.6178305 0.5801367
#>  [6,] 0.1413081 0.9116695 0.1620240 0.9899584 0.9252543 0.6736824 0.5779796
#>  [7,] 0.1347092 0.9735630 0.2198199 1.0348972 0.9295714 0.6358003 0.5713421
#>  [8,] 0.1222921 0.9646231 0.2369304 1.0431547 0.9949853 0.6726214 0.5782274
#>  [9,] 0.1445127 0.9444648 0.2401907 1.0404718 0.9129557 0.6242576 0.6073143
#> [10,] 0.1172496 1.0086696 0.2218496 1.0513544 0.9818826 0.6404802 0.5864549
#>           Y~X:Z
#>  [1,] 0.7397298
#>  [2,] 0.7321311
#>  [3,] 0.7435334
#>  [4,] 0.7575347
#>  [5,] 0.7684894
#>  [6,] 0.7041944
#>  [7,] 0.7226417
#>  [8,] 0.7139892
#>  [9,] 0.7215723
#> [10,] 0.6557154
# }

tpb <- "
# Outer Model (Based on Hagger et al., 2007)
  ATT =~ att1 + att2 + att3 + att4 + att5
  SN =~ sn1 + sn2
  PBC =~ pbc1 + pbc2 + pbc3
  INT =~ int1 + int2 + int3
  BEH =~ b1 + b2

# Inner Model (Based on Steinmetz et al., 2011)
  INT ~ ATT + SN + PBC
  BEH ~ INT + PBC + INT:PBC
"

# \dontrun{
boot <- bootstrap_modsem(model = modsem, 
                         model.syntax = tpb, data = TPB,
                         method = "dblcent", rcs = TRUE, 
                         rcs.scale.corrected = TRUE,
                         FUN = "coef")
coef <- apply(boot, MARGIN = 2, FUN = mean, na.rm = TRUE)
se   <- apply(boot, MARGIN = 2, FUN = sd, na.rm = TRUE)

cat("Parameter Estimates:\n")
#> Parameter Estimates:
print(coef)
#>        INT~ATT         INT~SN        INT~PBC        BEH~INT        BEH~PBC 
#>     0.21355519     0.17625531     0.21710382     0.18943361     0.22898572 
#>     BEH~INTPBC       ATT~~ATT         SN~~SN       PBC~~PBC       INT~~INT 
#>     0.20336938     0.99754989     0.98684728     0.96127940     0.48902043 
#>       BEH~~BEH INTPBC~~INTPBC        ATT~~SN       ATT~~PBC    ATT~~INTPBC 
#>     0.45025841     1.00136848     0.62757842     0.67713341     0.08505076 
#>        SN~~PBC     SN~~INTPBC    PBC~~INTPBC 
#>     0.67743622     0.05675458     0.08635358 

cat("Standard Errors: \n")
#> Standard Errors: 
print(se)
#>        INT~ATT         INT~SN        INT~PBC        BEH~INT        BEH~PBC 
#>     0.02616082     0.02604345     0.02929455     0.02303566     0.02135587 
#>     BEH~INTPBC       ATT~~ATT         SN~~SN       PBC~~PBC       INT~~INT 
#>     0.01892063     0.03815545     0.03785563     0.03639057     0.02012892 
#>       BEH~~BEH INTPBC~~INTPBC        ATT~~SN       ATT~~PBC    ATT~~INTPBC 
#>     0.02454723     0.08776389     0.02783643     0.02944580     0.03248502 
#>        SN~~PBC     SN~~INTPBC    PBC~~INTPBC 
#>     0.02847771     0.03066317     0.03650240 

# }