Get lavaan syntax for product indicator approaches

get_pi_syntax() is a function for creating the lavaan syntax used for estimating latent interaction models using one of the product indicators in lavaan.

Usage

get_pi_syntax(model.syntax, method = "dblcent", match = FALSE, ...)

Arguments

model.syntax

lavaan syntax

method

method to use: "rca" = residual centering approach, "uca" = unconstrained approach, "dblcent" = double centering approach, "pind" = prod ind approach, with no constraints or centering, "custom" = use parameters specified in the function call

match

should the product indicators be created by using the match-strategy

...

arguments passed to other functions (e.g., modsem_pi)

Value

character vector

Examples

library(modsem)
library(lavaan)
m1 <- '
  # Outer Model
  X =~ x1 + x2 + x3
  Y =~ y1 + y2 + y3
  Z =~ z1 + z2 + z3

  # Inner model
  Y ~ X + Z + X:Z
'
syntax <- get_pi_syntax(m1)
data <- get_pi_data(m1, oneInt)
est <- sem(syntax, data)