Define or disable the color theme used by modsem

All arguments are optional; omitted ones fall back to the defaults below. Pass active = FALSE to turn highlighting off (and reset the palette).

Usage

set_modsem_colors(
  positive = "green3",
  negative = positive,
  true = "green3",
  false = "red",
  nan = "purple",
  na = "purple",
  inf = "purple",
  string = "cyan",
  active = TRUE
)

Arguments

positive

color of positive numbers.

negative

color of negative numbers.

true

color of TRUE.

false

color of FALSE.

nan

color of NaN.

na

color of NA.

inf

color of -Inf and Inf.

string

color of quoted strings.

active

Should color-theme be activated/deactived?

Value

TRUE if colors are active afterwards, otherwise FALSE.

Examples

set_modsem_colors(positive = "red3", 
                  negative = "red3",
                  true = "darkgreen", 
                  false = "red3", 
                  na = "purple",
                  string = "darkgreen")
#> [1] TRUE

m1 <- "
# Outer Model
  X =~ x1 + x2 + x3
  Z =~ z1 + z2 + z3
  Y =~ y1 + y2 + y3
# Inner Model
  Y ~ X + Z + X:Z
"

est <- modsem(m1, data = oneInt)
colorize_output(summary(est))
#> Estimating baseline model (H0)
#> modsem (version 1.0.11, approach = dblcent):
#> 
#> Interaction Model Fit Measures (H1):
#>   Loglikelihood                              -26807.61 
#>   Akaike (AIC)                                53735.22 
#>   Bayesian (BIC)                              54071.28 
#>   Chi-square                                    122.92 
#>   Degrees of Freedom                               111 
#>   P-value (Chi-square)                           0.207 
#>   RMSEA                                          0.007 
#>   CFI                                            1.000 
#>   SRMR                                           0.008 
#> 
#> Fit Measures for Baseline Model (H0):
#>   Loglikelihood                              -27137.74 
#>   Akaike (AIC)                                54393.48 
#>   Bayesian (BIC)                              54723.93 
#>   Chi-square                                    783.18 
#>   Degrees of Freedom                               112 
#>   P-value (Chi-square)                           0.000 
#>   RMSEA                                          0.055 
#>   CFI                                            0.987 
#>   SRMR                                           0.141 
#> 
#> Comparative Fit to H0 (LRT test):
#>   Chi-square diff                              660.257 
#>   Degrees of freedom diff                            1 
#>   P-value (LRT)                                  0.000 
#> 
#> R-Squared Interaction Model (H1):
#>   Y                                              0.602 
#> R-Squared Baseline Model (H0):
#>   Y                                              0.397 
#> R-Squared Change (H1 - H0):
#>   Y                                              0.204 
#> 
#> lavaan 0.6-19 ended normally after 159 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        60
#> 
#>   Number of observations                          2000
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                               122.924
#>   Degrees of freedom                               111
#>   P-value (Chi-square)                           0.207
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   X =~                                                
#>     x1                1.000                           
#>     x2                0.804    0.013   63.612    0.000
#>     x3                0.916    0.014   67.144    0.000
#>   Z =~                                                
#>     z1                1.000                           
#>     z2                0.812    0.013   64.763    0.000
#>     z3                0.882    0.013   67.014    0.000
#>   Y =~                                                
#>     y1                1.000                           
#>     y2                0.798    0.007  107.428    0.000
#>     y3                0.899    0.008  112.453    0.000
#>   XZ =~                                               
#>     x1z1              1.000                           
#>     x2z1              0.805    0.013   60.636    0.000
#>     x3z1              0.877    0.014   62.680    0.000
#>     x1z2              0.793    0.013   59.343    0.000
#>     x2z2              0.646    0.015   43.672    0.000
#>     x3z2              0.706    0.016   44.292    0.000
#>     x1z3              0.887    0.014   63.700    0.000
#>     x2z3              0.716    0.016   45.645    0.000
#>     x3z3              0.781    0.017   45.339    0.000
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   Y ~                                                 
#>     X                 0.675    0.027   25.379    0.000
#>     Z                 0.561    0.026   21.606    0.000
#>     XZ                0.702    0.027   26.360    0.000
#> 
#> Covariances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>  .x1z1 ~~                                             
#>    .x1z2              0.115    0.008   14.802    0.000
#>    .x1z3              0.114    0.008   13.947    0.000
#>    .x2z1              0.125    0.008   16.095    0.000
#>    .x3z1              0.140    0.009   16.135    0.000
#>  .x1z2 ~~                                             
#>    .x1z3              0.103    0.007   14.675    0.000
#>    .x2z2              0.128    0.006   20.850    0.000
#>    .x3z2              0.146    0.007   21.243    0.000
#>  .x1z3 ~~                                             
#>    .x2z3              0.116    0.007   17.818    0.000
#>    .x3z3              0.135    0.007   18.335    0.000
#>  .x2z1 ~~                                             
#>    .x2z2              0.135    0.006   20.905    0.000
#>    .x2z3              0.145    0.007   21.145    0.000
#>    .x3z1              0.114    0.007   16.058    0.000
#>  .x2z2 ~~                                             
#>    .x2z3              0.117    0.006   20.419    0.000
#>    .x3z2              0.116    0.006   20.586    0.000
#>  .x2z3 ~~                                             
#>    .x3z3              0.109    0.006   18.059    0.000
#>  .x3z1 ~~                                             
#>    .x3z2              0.138    0.007   19.331    0.000
#>    .x3z3              0.158    0.008   20.269    0.000
#>  .x3z2 ~~                                             
#>    .x3z3              0.131    0.007   19.958    0.000
#>   X ~~                                                
#>     Z                 0.201    0.024    8.271    0.000
#>     XZ                0.016    0.025    0.628    0.530
#>   Z ~~                                                
#>     XZ                0.062    0.025    2.449    0.014
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .x1                0.160    0.009   17.871    0.000
#>    .x2                0.162    0.007   22.969    0.000
#>    .x3                0.163    0.008   20.161    0.000
#>    .z1                0.168    0.009   18.143    0.000
#>    .z2                0.158    0.007   22.264    0.000
#>    .z3                0.158    0.008   20.389    0.000
#>    .y1                0.159    0.009   17.896    0.000
#>    .y2                0.154    0.007   22.640    0.000
#>    .y3                0.164    0.008   20.698    0.000
#>    .x1z1              0.311    0.014   22.227    0.000
#>    .x2z1              0.292    0.011   27.287    0.000
#>    .x3z1              0.327    0.012   26.275    0.000
#>    .x1z2              0.290    0.011   26.910    0.000
#>    .x2z2              0.239    0.008   29.770    0.000
#>    .x3z2              0.270    0.009   29.117    0.000
#>    .x1z3              0.272    0.012   23.586    0.000
#>    .x2z3              0.245    0.009   27.979    0.000
#>    .x3z3              0.297    0.011   28.154    0.000
#>     X                 0.981    0.036   26.895    0.000
#>     Z                 1.016    0.038   26.856    0.000
#>    .Y                 0.990    0.038   25.926    0.000
#>     XZ                1.045    0.044   24.004    0.000
#> 
#>  
colorize_output(est) # same as colorize_output(print(est))
#> modsem (version 1.0.11, approach = dblcent):
#> lavaan 0.6-19 ended normally after 159 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        60
#> 
#>   Number of observations                          2000
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                               122.924
#>   Degrees of freedom                               111
#>   P-value (Chi-square)                           0.207
#>  
colorize_output(modsem_inspect(est, what = "coef"))
#> $lambda
#>          X     Z     Y    XZ
#> x1   1.000 0.000 0.000 0.000
#> x2   0.804 0.000 0.000 0.000
#> x3   0.916 0.000 0.000 0.000
#> z1   0.000 1.000 0.000 0.000
#> z2   0.000 0.812 0.000 0.000
#> z3   0.000 0.882 0.000 0.000
#> y1   0.000 0.000 1.000 0.000
#> y2   0.000 0.000 0.798 0.000
#> y3   0.000 0.000 0.899 0.000
#> x1z1 0.000 0.000 0.000 1.000
#> x2z1 0.000 0.000 0.000 0.805
#> x3z1 0.000 0.000 0.000 0.877
#> x1z2 0.000 0.000 0.000 0.793
#> x2z2 0.000 0.000 0.000 0.646
#> x3z2 0.000 0.000 0.000 0.706
#> x1z3 0.000 0.000 0.000 0.887
#> x2z3 0.000 0.000 0.000 0.716
#> x3z3 0.000 0.000 0.000 0.781
#> 
#> $theta
#>         x1    x2    x3    z1    z2    z3    y1    y2    y3  x1z1  x2z1  x3z1
#> x1   0.160                                                                  
#> x2   0.000 0.162                                                            
#> x3   0.000 0.000 0.163                                                      
#> z1   0.000 0.000 0.000 0.168                                                
#> z2   0.000 0.000 0.000 0.000 0.158                                          
#> z3   0.000 0.000 0.000 0.000 0.000 0.158                                    
#> y1   0.000 0.000 0.000 0.000 0.000 0.000 0.159                              
#> y2   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.154                        
#> y3   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.164                  
#> x1z1 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.311            
#> x2z1 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.125 0.292      
#> x3z1 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.140 0.114 0.327
#> x1z2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.115 0.000 0.000
#> x2z2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.135 0.000
#> x3z2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.138
#> x1z3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.114 0.000 0.000
#> x2z3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.145 0.000
#> x3z3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.158
#>       x1z2  x2z2  x3z2  x1z3  x2z3  x3z3
#> x1                                      
#> x2                                      
#> x3                                      
#> z1                                      
#> z2                                      
#> z3                                      
#> y1                                      
#> y2                                      
#> y3                                      
#> x1z1                                    
#> x2z1                                    
#> x3z1                                    
#> x1z2 0.290                              
#> x2z2 0.128 0.239                        
#> x3z2 0.146 0.116 0.270                  
#> x1z3 0.103 0.000 0.000 0.272            
#> x2z3 0.000 0.117 0.000 0.116 0.245      
#> x3z3 0.000 0.000 0.131 0.135 0.109 0.297
#> 
#> $psi
#>        X     Z     Y    XZ
#> X  0.981                  
#> Z  0.201 1.016            
#> Y  0.000 0.000 0.990      
#> XZ 0.016 0.062 0.000 1.045
#> 
#> $beta
#>        X     Z Y    XZ
#> X  0.000 0.000 0 0.000
#> Z  0.000 0.000 0 0.000
#> Y  0.675 0.561 0 0.702
#> XZ 0.000 0.000 0 0.000
#> 
#>  

# \dontrun{
colorize_output(split = TRUE, {
  # Get live (uncolored) output
  # And print colored output at the end of execution

  est_lms <- modsem(m1, data = oneInt, method = "lms")
  summary(est_lms)
}) 
#> 
#> modsem (version 1.0.11):
#> 
#>   Estimator                                         LMS
#>   Optimization method                        EMA-NLMINB
#>   Number of observations                           2000
#>   Number of iterations                               45
#>   Loglikelihood                               -17494.06
#>   Akaike (AIC)                                 35050.12
#>   Bayesian (BIC)                               35223.74
#>  
#> Numerical Integration:
#>   Points of integration (per dim)                    24
#>   Dimensions                                          1
#>   Total points of integration                        24
#>  
#> Fit Measures for Baseline Model (H0):
#>   Loglikelihood                               -17831.87
#>   Akaike (AIC)                                 35723.75
#>   Bayesian (BIC)                               35891.78
#>   Chi-square                                      17.52
#>   Degrees of Freedom (Chi-square)                    24
#>   P-value (Chi-square)                            0.826
#>   RMSEA                                           0.000
#>  
#> Comparative Fit to H0 (LRT test):
#>   Loglikelihood change                           337.82
#>   Difference test (D)                            675.63
#>   Degrees of freedom (D)                              1
#>   P-value (D)                                     0.000
#>  
#> R-Squared Interaction Model (H1):
#>   Y                                               0.595
#> R-Squared Baseline Model (H0):
#>   Y                                               0.395
#> R-Squared Change (H1 - H0):
#>   Y                                               0.199
#> 
#> Parameter Estimates:
#>   Coefficients                           unstandardized
#>   Information                                  observed
#>   Standard errors                              standard
#>  
#> Latent Variables:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>   X =~          
#>     x1              1.000                             
#>     x2              0.804      0.012    64.31    0.000
#>     x3              0.915      0.013    67.94    0.000
#>   Z =~          
#>     z1              1.000                             
#>     z2              0.810      0.012    65.09    0.000
#>     z3              0.881      0.013    67.61    0.000
#>   Y =~          
#>     y1              1.000                             
#>     y2              0.798      0.007   107.55    0.000
#>     y3              0.899      0.008   112.58    0.000
#> 
#> Regressions:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>   Y ~           
#>     X               0.672      0.031    21.71    0.000
#>     Z               0.569      0.030    19.05    0.000
#>     X:Z             0.713      0.028    25.72    0.000
#> 
#> Intercepts:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>    .x1              1.022      0.021    47.66    0.000
#>    .x2              1.215      0.018    67.19    0.000
#>    .x3              0.919      0.020    45.92    0.000
#>    .z1              1.011      0.024    41.69    0.000
#>    .z2              1.205      0.020    59.45    0.000
#>    .z3              0.915      0.022    42.18    0.000
#>    .y1              1.036      0.032    32.28    0.000
#>    .y2              1.220      0.026    46.66    0.000
#>    .y3              0.953      0.029    32.69    0.000
#> 
#> Covariances:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>   X ~~          
#>     Z               0.199      0.024     8.32    0.000
#> 
#> Variances:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>    .x1              0.160      0.008    19.30    0.000
#>    .x2              0.163      0.007    23.88    0.000
#>    .x3              0.165      0.008    21.23    0.000
#>    .z1              0.166      0.009    18.47    0.000
#>    .z2              0.160      0.007    22.67    0.000
#>    .z3              0.158      0.008    20.80    0.000
#>    .y1              0.160      0.009    18.01    0.000
#>    .y2              0.154      0.007    22.68    0.000
#>    .y3              0.164      0.008    20.68    0.000
#>     X               0.972      0.033    29.85    0.000
#>     Z               1.017      0.038    26.95    0.000
#>    .Y               0.984      0.038    26.00    0.000
#> 
#> 
#> modsem (version 1.0.11):
#> 
#>   Estimator                                         LMS
#>   Optimization method                        EMA-NLMINB
#>   Number of observations                           2000
#>   Number of iterations                               45
#>   Loglikelihood                               -17494.06
#>   Akaike (AIC)                                 35050.12
#>   Bayesian (BIC)                               35223.74
#>  
#> Numerical Integration:
#>   Points of integration (per dim)                    24
#>   Dimensions                                          1
#>   Total points of integration                        24
#>  
#> Fit Measures for Baseline Model (H0):
#>   Loglikelihood                               -17831.87
#>   Akaike (AIC)                                 35723.75
#>   Bayesian (BIC)                               35891.78
#>   Chi-square                                      17.52
#>   Degrees of Freedom (Chi-square)                    24
#>   P-value (Chi-square)                            0.826
#>   RMSEA                                           0.000
#>  
#> Comparative Fit to H0 (LRT test):
#>   Loglikelihood change                           337.82
#>   Difference test (D)                            675.63
#>   Degrees of freedom (D)                              1
#>   P-value (D)                                     0.000
#>  
#> R-Squared Interaction Model (H1):
#>   Y                                               0.595
#> R-Squared Baseline Model (H0):
#>   Y                                               0.395
#> R-Squared Change (H1 - H0):
#>   Y                                               0.199
#> 
#> Parameter Estimates:
#>   Coefficients                           unstandardized
#>   Information                                  observed
#>   Standard errors                              standard
#>  
#> Latent Variables:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>   X =~          
#>     x1              1.000                             
#>     x2              0.804      0.012    64.31    0.000
#>     x3              0.915      0.013    67.94    0.000
#>   Z =~          
#>     z1              1.000                             
#>     z2              0.810      0.012    65.09    0.000
#>     z3              0.881      0.013    67.61    0.000
#>   Y =~          
#>     y1              1.000                             
#>     y2              0.798      0.007   107.55    0.000
#>     y3              0.899      0.008   112.58    0.000
#> 
#> Regressions:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>   Y ~           
#>     X               0.672      0.031    21.71    0.000
#>     Z               0.569      0.030    19.05    0.000
#>     X:Z             0.713      0.028    25.72    0.000
#> 
#> Intercepts:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>    .x1              1.022      0.021    47.66    0.000
#>    .x2              1.215      0.018    67.19    0.000
#>    .x3              0.919      0.020    45.92    0.000
#>    .z1              1.011      0.024    41.69    0.000
#>    .z2              1.205      0.020    59.45    0.000
#>    .z3              0.915      0.022    42.18    0.000
#>    .y1              1.036      0.032    32.28    0.000
#>    .y2              1.220      0.026    46.66    0.000
#>    .y3              0.953      0.029    32.69    0.000
#> 
#> Covariances:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>   X ~~          
#>     Z               0.199      0.024     8.32    0.000
#> 
#> Variances:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>    .x1              0.160      0.008    19.30    0.000
#>    .x2              0.163      0.007    23.88    0.000
#>    .x3              0.165      0.008    21.23    0.000
#>    .z1              0.166      0.009    18.47    0.000
#>    .z2              0.160      0.007    22.67    0.000
#>    .z3              0.158      0.008    20.80    0.000
#>    .y1              0.160      0.009    18.01    0.000
#>    .y2              0.154      0.007    22.68    0.000
#>    .y3              0.164      0.008    20.68    0.000
#>     X               0.972      0.033    29.85    0.000
#>     Z               1.017      0.038    26.95    0.000
#>    .Y               0.984      0.038    26.00    0.000
#> 
#>  
                
colorize_output(modsem_inspect(est_lms))
#> $N
#> [1] 2000
#> 
#> $vcov.all
#>        X=~x2 X=~x3 Z=~z2 Z=~z3 Y=~y2 Y=~y3  x1~1  x2~1  x3~1  z1~1  z2~1  z3~1
#> X=~x2  0.000                                                                  
#> X=~x3  0.000 0.000                                                            
#> Z=~z2  0.000 0.000 0.000                                                      
#> Z=~z3  0.000 0.000 0.000 0.000                                                
#> Y=~y2  0.000 0.000 0.000 0.000 0.000                                          
#> Y=~y3  0.000 0.000 0.000 0.000 0.000 0.000                                    
#> x1~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000                              
#> x2~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000                        
#> x3~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000                  
#> z1~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.001            
#> z2~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000      
#> z3~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y1~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y2~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y3~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> x1~~x1 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> x2~~x2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> x3~~x3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> z1~~z1 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> z2~~z2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> z3~~z3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y1~~y1 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y2~~y2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y3~~y3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> X~~X   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> X~~Z   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Z~~Z   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Y~~Y   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Y~X    0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Y~Z    0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Y~X:Z  0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#>         y1~1  y2~1  y3~1 x1~~x1 x2~~x2 x3~~x3 z1~~z1 z2~~z2 z3~~z3 y1~~y1
#> X=~x2                                                                    
#> X=~x3                                                                    
#> Z=~z2                                                                    
#> Z=~z3                                                                    
#> Y=~y2                                                                    
#> Y=~y3                                                                    
#> x1~1                                                                     
#> x2~1                                                                     
#> x3~1                                                                     
#> z1~1                                                                     
#> z2~1                                                                     
#> z3~1                                                                     
#> y1~1   0.001                                                             
#> y2~1   0.001 0.001                                                       
#> y3~1   0.001 0.001 0.001                                                 
#> x1~~x1 0.000 0.000 0.000  0.000                                          
#> x2~~x2 0.000 0.000 0.000  0.000  0.000                                   
#> x3~~x3 0.000 0.000 0.000  0.000  0.000  0.000                            
#> z1~~z1 0.000 0.000 0.000  0.000  0.000  0.000  0.000                     
#> z2~~z2 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000              
#> z3~~z3 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000       
#> y1~~y1 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> y2~~y2 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> y3~~y3 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> X~~X   0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> X~~Z   0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Z~~Z   0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Y~~Y   0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Y~X    0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Y~Z    0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Y~X:Z  0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#>        y2~~y2 y3~~y3  X~~X  X~~Z  Z~~Z  Y~~Y   Y~X   Y~Z Y~X:Z
#> X=~x2                                                         
#> X=~x3                                                         
#> Z=~z2                                                         
#> Z=~z3                                                         
#> Y=~y2                                                         
#> Y=~y3                                                         
#> x1~1                                                          
#> x2~1                                                          
#> x3~1                                                          
#> z1~1                                                          
#> z2~1                                                          
#> z3~1                                                          
#> y1~1                                                          
#> y2~1                                                          
#> y3~1                                                          
#> x1~~x1                                                        
#> x2~~x2                                                        
#> x3~~x3                                                        
#> z1~~z1                                                        
#> z2~~z2                                                        
#> z3~~z3                                                        
#> y1~~y1                                                        
#> y2~~y2  0.000                                                 
#> y3~~y3  0.000  0.000                                          
#> X~~X    0.000  0.000 0.001                                    
#> X~~Z    0.000  0.000 0.000 0.001                              
#> Z~~Z    0.000  0.000 0.000 0.000 0.001                        
#> Y~~Y    0.000  0.000 0.000 0.000 0.000 0.001                  
#> Y~X     0.000  0.000 0.000 0.000 0.000 0.000 0.001            
#> Y~Z     0.000  0.000 0.000 0.000 0.000 0.000 0.000 0.001      
#> Y~X:Z   0.000  0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.001
#> 
#> $vcov.free
#>        X=~x2 X=~x3 Z=~z2 Z=~z3 Y=~y2 Y=~y3  x1~1  x2~1  x3~1  z1~1  z2~1  z3~1
#> X=~x2  0.000                                                                  
#> X=~x3  0.000 0.000                                                            
#> Z=~z2  0.000 0.000 0.000                                                      
#> Z=~z3  0.000 0.000 0.000 0.000                                                
#> Y=~y2  0.000 0.000 0.000 0.000 0.000                                          
#> Y=~y3  0.000 0.000 0.000 0.000 0.000 0.000                                    
#> x1~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000                              
#> x2~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000                        
#> x3~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000                  
#> z1~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.001            
#> z2~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000      
#> z3~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y1~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y2~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y3~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> x1~~x1 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> x2~~x2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> x3~~x3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> z1~~z1 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> z2~~z2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> z3~~z3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y1~~y1 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y2~~y2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y3~~y3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> X~~X   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> X~~Z   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Z~~Z   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Y~~Y   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Y~X    0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Y~Z    0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Y~X:Z  0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#>         y1~1  y2~1  y3~1 x1~~x1 x2~~x2 x3~~x3 z1~~z1 z2~~z2 z3~~z3 y1~~y1
#> X=~x2                                                                    
#> X=~x3                                                                    
#> Z=~z2                                                                    
#> Z=~z3                                                                    
#> Y=~y2                                                                    
#> Y=~y3                                                                    
#> x1~1                                                                     
#> x2~1                                                                     
#> x3~1                                                                     
#> z1~1                                                                     
#> z2~1                                                                     
#> z3~1                                                                     
#> y1~1   0.001                                                             
#> y2~1   0.001 0.001                                                       
#> y3~1   0.001 0.001 0.001                                                 
#> x1~~x1 0.000 0.000 0.000  0.000                                          
#> x2~~x2 0.000 0.000 0.000  0.000  0.000                                   
#> x3~~x3 0.000 0.000 0.000  0.000  0.000  0.000                            
#> z1~~z1 0.000 0.000 0.000  0.000  0.000  0.000  0.000                     
#> z2~~z2 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000              
#> z3~~z3 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000       
#> y1~~y1 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> y2~~y2 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> y3~~y3 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> X~~X   0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> X~~Z   0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Z~~Z   0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Y~~Y   0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Y~X    0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Y~Z    0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Y~X:Z  0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#>        y2~~y2 y3~~y3  X~~X  X~~Z  Z~~Z  Y~~Y   Y~X   Y~Z Y~X:Z
#> X=~x2                                                         
#> X=~x3                                                         
#> Z=~z2                                                         
#> Z=~z3                                                         
#> Y=~y2                                                         
#> Y=~y3                                                         
#> x1~1                                                          
#> x2~1                                                          
#> x3~1                                                          
#> z1~1                                                          
#> z2~1                                                          
#> z3~1                                                          
#> y1~1                                                          
#> y2~1                                                          
#> y3~1                                                          
#> x1~~x1                                                        
#> x2~~x2                                                        
#> x3~~x3                                                        
#> z1~~z1                                                        
#> z2~~z2                                                        
#> z3~~z3                                                        
#> y1~~y1                                                        
#> y2~~y2  0.000                                                 
#> y3~~y3  0.000  0.000                                          
#> X~~X    0.000  0.000 0.000                                    
#> X~~Z    0.000  0.000 0.000 0.001                              
#> Z~~Z    0.000  0.000 0.000 0.000 0.000                        
#> Y~~Y    0.000  0.000 0.000 0.000 0.000 0.001                  
#> Y~X     0.000  0.000 0.000 0.000 0.000 0.000 0.001            
#> Y~Z     0.000  0.000 0.000 0.000 0.000 0.000 0.000 0.001      
#> Y~X:Z   0.000  0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.001
#> 
#> $information
#>            X=~x2     X=~x3     Z=~z2     Z=~z3     Y=~y2     Y=~y3      x1~1
#> X=~x2   8998.852                                                            
#> X=~x3  -2648.535  8221.895                                                  
#> Z=~z2     60.899    70.492  9243.514                                        
#> Z=~z3     71.250    77.495 -3030.946  8771.725                              
#> Y=~y2   -165.568  -191.302  -157.571  -170.515 23807.686                    
#> Y=~y3   -172.959  -201.809  -168.409  -181.131 -8384.418 21002.412          
#> x1~1      15.422    40.312    89.581    97.296  -120.102  -132.167  8025.860
#> x2~1      15.497    31.833    70.865    76.968   -95.007  -104.552 -3521.253
#> x3~1      13.495    39.339    79.484    86.330  -106.561  -117.266 -3949.449
#> z1~1      84.662    93.902    23.124    27.207  -125.723  -132.396    53.111
#> z2~1      71.570    79.381    26.175    23.007  -106.281  -111.922    44.898
#> z3~1      78.367    86.920    21.256    32.270  -116.371  -122.547    49.164
#> y1~1     -24.527   -26.024   -23.821   -28.091  -592.021  -629.759  -172.645
#> y2~1     -20.247   -21.483   -19.665   -23.189  1373.746  -519.899  -142.519
#> y3~1     -21.528   -22.842   -20.909   -24.656  -519.820  1205.643  -151.532
#> x1~~x1 -1289.576 -1452.130    21.448    29.983    -4.666    -5.236   -82.268
#> x2~~x2  2481.752  -928.011    24.920     9.637   -11.518   -16.419    86.836
#> x3~~x3 -1069.178  2558.825    19.183    25.452    16.570    22.705   -18.427
#> z1~~z1    19.569    20.462 -1321.785 -1444.868    16.705    15.748    42.183
#> z2~~z2    32.442    21.371  2852.200 -1036.280    -4.214    -0.955    17.883
#> z3~~z3    13.467    23.860 -1115.977  2972.201   -16.191   -19.233    25.782
#> y1~~y1     0.700     1.958     8.264     6.885 -1490.249 -1622.403     0.368
#> y2~~y2     0.889    -6.787   -12.296    -9.484  3045.583 -1171.553   -14.594
#> y3~~y3   -15.850   -16.328    -9.013   -11.551 -1249.710  2846.425     8.134
#> X~~X    1455.342  1633.422   -62.079   -64.990    71.421    74.799     8.835
#> X~~Z    -112.320  -127.600   111.884   130.238    32.863    35.785   -42.447
#> Z~~Z     -26.076   -28.177  1158.540  1266.644    48.202    51.019   -20.069
#> Y~~Y     -35.920   -52.213   -31.474   -34.565   549.480   589.038   -13.244
#> Y~X     -268.710  -299.251   -42.853   -50.215   434.920   458.045   -71.040
#> Y~Z      -59.850   -62.507  -273.539  -282.497   394.609   428.978  -424.871
#> Y~X:Z   -284.457  -320.283  -316.768  -345.446   417.270   436.678   -67.213
#>             x2~1      x3~1      z1~1      z2~1      z3~1      y1~1      y2~1
#> X=~x2                                                                       
#> X=~x3                                                                       
#> Z=~z2                                                                       
#> Z=~z3                                                                       
#> Y=~y2                                                                       
#> Y=~y3                                                                       
#> x1~1                                                                        
#> x2~1    9492.835                                                            
#> x3~1   -3124.247  8600.555                                                  
#> z1~1      42.015    47.128  7696.723                                        
#> z2~1      35.518    39.840 -3648.757  9449.721                              
#> z3~1      38.893    43.626 -3995.150 -3377.343  8925.979                    
#> y1~1    -136.572  -153.180  -144.423  -122.090  -133.680  7702.503          
#> y2~1    -112.741  -126.451  -119.222  -100.786  -110.353 -3986.012  9667.997
#> y3~1    -119.870  -134.447  -126.761  -107.159  -117.332 -4238.073 -3498.551
#> x1~~x1   -53.596   -59.614    40.077    33.880    37.098    -8.043    -6.639
#> x2~~x2    46.286    77.493    18.829    15.917    17.430     9.773     8.068
#> x3~~x3   -14.429   -31.270    29.489    24.929    27.296     4.075     3.363
#> z1~~z1    33.370    37.432   -17.850   -10.699   -11.166    -1.418    -1.171
#> z2~~z2    14.147    15.869     2.290   -12.858     2.493    -2.109    -1.741
#> z3~~z3    20.395    22.875    -7.774    -6.592   -15.048     8.643     7.135
#> y1~~y1     0.292     0.328    -4.739    -4.006    -4.385   -10.791    -4.185
#> y2~~y2   -11.544   -12.947     3.205     2.709     2.967    -4.707   -20.239
#> y3~~y3     6.436     7.219    -3.343    -2.826    -3.093    -3.400    -2.727
#> X~~X       6.988     7.836   -46.918   -39.663   -43.430    -1.168    -0.964
#> X~~Z     -33.581   -37.670   -39.448   -33.348   -36.514    56.489    46.631
#> Z~~Z     -15.876   -17.806     8.013     6.772     7.447   -13.532   -11.170
#> Y~~Y     -10.474   -11.738   -12.398   -10.481   -11.468    -1.131    -0.931
#> Y~X      -56.202   -63.051  -396.704  -335.358  -367.194   -33.668   -27.793
#> Y~Z     -336.097  -376.967   -70.435   -59.543   -65.209   -38.246   -31.572
#> Y~X:Z    -53.174   -59.651   -54.617   -46.171   -50.568   129.279   106.720
#>             y3~1    x1~~x1    x2~~x2    x3~~x3    z1~~z1    z2~~z2    z3~~z3
#> X=~x2                                                                       
#> X=~x3                                                                       
#> Z=~z2                                                                       
#> Z=~z3                                                                       
#> Y=~y2                                                                       
#> Y=~y3                                                                       
#> x1~1                                                                        
#> x2~1                                                                        
#> x3~1                                                                        
#> z1~1                                                                        
#> z2~1                                                                        
#> z3~1                                                                        
#> y1~1                                                                        
#> y2~1                                                                        
#> y3~1    8512.191                                                            
#> x1~~x1    -7.057 17221.526                                                  
#> x2~~x2     8.579  3379.648 23313.755                                        
#> x3~~x3     3.578  4279.775  2503.524 18943.320                              
#> z1~~z1    -1.243    -2.823     9.331    19.888 14811.578                    
#> z2~~z2    -1.850    46.267   -34.248    40.057  3316.728 22329.663          
#> z3~~z3     7.588    19.337    57.025    23.278  3962.023  2863.636 19903.315
#> y1~~y1    -3.805    32.939     9.682     5.725    18.285    20.628    36.764
#> y2~~y2    -3.616     0.484    -9.002    31.812     1.990    10.320    -9.738
#> y3~~y3   -11.041     2.037    12.776     7.957    13.298    -7.731     0.083
#> X~~X      -1.025   169.846   -39.917   -57.287    20.541   -24.855   -46.359
#> X~~Z      49.581   -26.306   -56.316   -61.783    12.456    13.510   -26.834
#> Z~~Z     -11.876   -12.323     8.108     4.298   266.280   187.822   235.050
#> Y~~Y      -0.976    86.883    34.446   109.295    82.821    55.522    67.141
#> Y~X      -29.549  -137.614   -85.545  -119.295    35.587   -16.411    -1.356
#> Y~Z      -33.567    -0.092    18.188    -2.681  -133.565   -43.273  -146.314
#> Y~X:Z    113.471  -137.399   -87.476  -102.943   -92.408  -117.809  -148.849
#>           y1~~y1    y2~~y2    y3~~y3      X~~X      X~~Z      Z~~Z      Y~~Y
#> X=~x2                                                                       
#> X=~x3                                                                       
#> Z=~z2                                                                       
#> Z=~z3                                                                       
#> Y=~y2                                                                       
#> Y=~y3                                                                       
#> x1~1                                                                        
#> x2~1                                                                        
#> x3~1                                                                        
#> z1~1                                                                        
#> z2~1                                                                        
#> z3~1                                                                        
#> y1~1                                                                        
#> y2~1                                                                        
#> y3~1                                                                        
#> x1~~x1                                                                      
#> x2~~x2                                                                      
#> x3~~x3                                                                      
#> z1~~z1                                                                      
#> z2~~z2                                                                      
#> z3~~z3                                                                      
#> y1~~y1 14839.859                                                            
#> y2~~y2  3729.485 23375.076                                                  
#> y3~~y3  4419.356  3351.964 18124.417                                        
#> X~~X      -9.649    -2.385     3.513  4772.078                              
#> X~~Z       0.485     1.753    -2.664  -298.262  1831.334                    
#> Z~~Z      -5.178     3.446     5.141     2.238    46.145  3587.883          
#> Y~~Y     171.565    77.546    60.861   -20.432    -1.326     8.380   757.973
#> Y~X       18.064   -10.912    11.536   178.243    36.901    -1.274    47.318
#> Y~Z      -12.680    40.122    -6.605    19.858    64.112    87.843    47.378
#> Y~X:Z    -31.073    16.345    56.629   169.599    55.910   110.466   104.647
#>              Y~X       Y~Z     Y~X:Z
#> X=~x2                               
#> X=~x3                               
#> Z=~z2                               
#> Z=~z3                               
#> Y=~y2                               
#> Y=~y3                               
#> x1~1                                
#> x2~1                                
#> x3~1                                
#> z1~1                                
#> z2~1                                
#> z3~1                                
#> y1~1                                
#> y2~1                                
#> y3~1                                
#> x1~~x1                              
#> x2~~x2                              
#> x3~~x3                              
#> z1~~z1                              
#> z2~~z2                              
#> z3~~z3                              
#> y1~~y1                              
#> y2~~y2                              
#> y3~~y3                              
#> X~~X                                
#> X~~Z                                
#> Z~~Z                                
#> Y~~Y                                
#> Y~X     1548.184                    
#> Y~Z      319.901  1581.283          
#> Y~X:Z    -51.319   -26.945  1518.431
#> 
#> $coefficients.all
#>  X=~x2  X=~x3  Z=~z2  Z=~z3  Y=~y2  Y=~y3   x1~1   x2~1   x3~1   z1~1   z2~1 
#>  0.804  0.915  0.810  0.881  0.798  0.899  1.022  1.215  0.919  1.011  1.205 
#>   z3~1   y1~1   y2~1   y3~1 x1~~x1 x2~~x2 x3~~x3 z1~~z1 z2~~z2 z3~~z3 y1~~y1 
#>  0.915  1.036  1.220  0.953  0.160  0.163  0.165  0.166  0.160  0.158  0.160 
#> y2~~y2 y3~~y3   X~~X   X~~Z   Z~~Z   Y~~Y    Y~X    Y~Z  Y~X:Z 
#>  0.154  0.164  0.972  0.199  1.017  0.984  0.672  0.569  0.713 
#> 
#> $coefficients.free
#>  X=~x2  X=~x3  Z=~z2  Z=~z3  Y=~y2  Y=~y3   x1~1   x2~1   x3~1   z1~1   z2~1 
#>  0.804  0.915  0.810  0.881  0.798  0.899  1.022  1.215  0.919  1.011  1.205 
#>   z3~1   y1~1   y2~1   y3~1 x1~~x1 x2~~x2 x3~~x3 z1~~z1 z2~~z2 z3~~z3 y1~~y1 
#>  0.915  1.036  1.220  0.953  0.160  0.163  0.165  0.166  0.160  0.158  0.160 
#> y2~~y2 y3~~y3   X~~X   X~~Z   Z~~Z   Y~~Y    Y~X    Y~Z  Y~X:Z 
#>  0.154  0.164  0.972  0.199  1.017  0.984  0.672  0.569  0.713 
#> 
#> $partable
#>    lhs op rhs label   est std.error z.value p.value ci.lower ci.upper
#> 1    X =~  x1       1.000        NA      NA      NA       NA       NA
#> 2    X =~  x2       0.804     0.012  64.315       0    0.779    0.828
#> 3    X =~  x3       0.915     0.013  67.942       0    0.888    0.941
#> 4    Z =~  z1       1.000        NA      NA      NA       NA       NA
#> 5    Z =~  z2       0.810     0.012  65.093       0    0.786    0.835
#> 6    Z =~  z3       0.881     0.013  67.605       0    0.855    0.906
#> 7    Y =~  y1       1.000        NA      NA      NA       NA       NA
#> 8    Y =~  y2       0.798     0.007 107.547       0    0.784    0.813
#> 9    Y =~  y3       0.899     0.008 112.579       0    0.884    0.915
#> 10   Y  ~   X       0.672     0.031  21.710       0    0.611    0.733
#> 11   Y  ~   Z       0.569     0.030  19.051       0    0.511    0.628
#> 12   Y  ~ X:Z       0.713     0.028  25.720       0    0.659    0.767
#> 13  x1 ~1           1.022     0.021  47.658       0    0.980    1.064
#> 14  x2 ~1           1.215     0.018  67.193       0    1.180    1.251
#> 15  x3 ~1           0.919     0.020  45.920       0    0.880    0.958
#> 16  z1 ~1           1.011     0.024  41.692       0    0.963    1.058
#> 17  z2 ~1           1.205     0.020  59.447       0    1.165    1.245
#> 18  z3 ~1           0.915     0.022  42.179       0    0.872    0.957
#> 19  y1 ~1           1.036     0.032  32.284       0    0.973    1.099
#> 20  y2 ~1           1.220     0.026  46.665       0    1.168    1.271
#> 21  y3 ~1           0.953     0.029  32.685       0    0.896    1.010
#> 22  x1 ~~  x1       0.160     0.008  19.299       0    0.144    0.177
#> 23  x2 ~~  x2       0.163     0.007  23.882       0    0.150    0.176
#> 24  x3 ~~  x3       0.165     0.008  21.229       0    0.150    0.180
#> 25  z1 ~~  z1       0.166     0.009  18.472       0    0.149    0.184
#> 26  z2 ~~  z2       0.160     0.007  22.670       0    0.146    0.173
#> 27  z3 ~~  z3       0.158     0.008  20.795       0    0.143    0.173
#> 28  y1 ~~  y1       0.160     0.009  18.010       0    0.142    0.177
#> 29  y2 ~~  y2       0.154     0.007  22.681       0    0.141    0.168
#> 30  y3 ~~  y3       0.164     0.008  20.680       0    0.148    0.179
#> 31   X ~~   X       0.972     0.033  29.853       0    0.908    1.036
#> 32   X ~~   Z       0.199     0.024   8.320       0    0.152    0.245
#> 33   Z ~~   Z       1.017     0.038  26.945       0    0.943    1.091
#> 34   Y ~~   Y       0.984     0.038  25.996       0    0.910    1.058
#> 
#> $partable.input
#>    lhs op rhs mod
#> 1    X =~  x1    
#> 2    X =~  x2    
#> 3    X =~  x3    
#> 4    Z =~  z1    
#> 5    Z =~  z2    
#> 6    Z =~  z3    
#> 7    Y =~  y1    
#> 8    Y =~  y2    
#> 9    Y =~  y3    
#> 10   Y  ~   X    
#> 11   Y  ~   Z    
#> 12   Y  ~ X:Z    
#> 
#> $loglik
#> [1] -17494.06
#> 
#> $iterations
#> [1] 45
#> 
#> $convergence
#> [1] TRUE
#> 
#> $lambda
#>        X     Z     Y
#> x1 1.000 0.000 0.000
#> x2 0.804 0.000 0.000
#> x3 0.915 0.000 0.000
#> z1 0.000 1.000 0.000
#> z2 0.000 0.810 0.000
#> z3 0.000 0.881 0.000
#> y1 0.000 0.000 1.000
#> y2 0.000 0.000 0.798
#> y3 0.000 0.000 0.899
#> 
#> $tau
#>       ~1
#> x1 1.022
#> x2 1.215
#> x3 0.919
#> z1 1.011
#> z2 1.205
#> z3 0.915
#> y1 1.036
#> y2 1.220
#> y3 0.953
#> 
#> $theta
#>       x1    x2    x3    z1    z2    z3    y1    y2    y3
#> x1 0.160                                                
#> x2 0.000 0.163                                          
#> x3 0.000 0.000 0.165                                    
#> z1 0.000 0.000 0.000 0.166                              
#> z2 0.000 0.000 0.000 0.000 0.160                        
#> z3 0.000 0.000 0.000 0.000 0.000 0.158                  
#> y1 0.000 0.000 0.000 0.000 0.000 0.000 0.160            
#> y2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.154      
#> y3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.164
#> 
#> $gamma.xi
#>       X     Z
#> Y 0.672 0.569
#> 
#> $gamma.eta
#>       Y
#> Y 0.000
#> 
#> $omega.xi.xi
#>         X     Z
#> Y~X 0.000 0.713
#> Y~Z 0.000 0.000
#> 
#> $omega.eta.xi
#>         Y
#> Y~X 0.000
#> Y~Z 0.000
#> 
#> $phi
#>       X     Z
#> X 0.972      
#> Z 0.199 1.017
#> 
#> $psi
#>       Y
#> Y 0.984
#> 
#> $alpha
#>      ~1
#> Y 0.000
#> 
#> $beta0
#>      ~1
#> X 0.000
#> Z 0.000
#> 
#> $cov.ov
#>       x1    x2    x3    z1    z2    z3    y1    y2    y3
#> x1 1.132                                                
#> x2 0.781 0.791                                          
#> x3 0.889 0.714 0.978                                    
#> z1 0.199 0.160 0.182 1.184                              
#> z2 0.161 0.129 0.147 0.824 0.827                        
#> z3 0.175 0.141 0.160 0.896 0.726 0.948                  
#> y1 0.766 0.616 0.701 0.712 0.577 0.628 2.586            
#> y2 0.612 0.492 0.559 0.569 0.461 0.501 1.937 1.701      
#> y3 0.689 0.554 0.630 0.641 0.519 0.564 2.182 1.742 2.126
#> 
#> $cov.lv
#>         X     Z   X:Z     Y
#> X   0.972                  
#> Z   0.199 1.017            
#> X:Z 0.000 0.000 1.028      
#> Y   0.766 0.712 0.733 2.427
#> 
#> $cov.all
#>         X     Z   X:Z     Y    x1    x2    x3    z1    z2    z3    y1    y2
#> X   0.972                                                                  
#> Z   0.199 1.017                                                            
#> X:Z 0.000 0.000 1.028                                                      
#> Y   0.766 0.712 0.733 2.427                                                
#> x1  0.972 0.199 0.000 0.766 1.132                                          
#> x2  0.781 0.160 0.000 0.616 0.781 0.791                                    
#> x3  0.889 0.182 0.000 0.701 0.889 0.714 0.978                              
#> z1  0.199 1.017 0.000 0.712 0.199 0.160 0.182 1.184                        
#> z2  0.161 0.824 0.000 0.577 0.161 0.129 0.147 0.824 0.827                  
#> z3  0.175 0.896 0.000 0.628 0.175 0.141 0.160 0.896 0.726 0.948            
#> y1  0.766 0.712 0.733 2.427 0.766 0.616 0.701 0.712 0.577 0.628 2.586      
#> y2  0.612 0.569 0.585 1.937 0.612 0.492 0.559 0.569 0.461 0.501 1.937 1.701
#> y3  0.689 0.641 0.659 2.182 0.689 0.554 0.630 0.641 0.519 0.564 2.182 1.742
#>        y3
#> X        
#> Z        
#> X:Z      
#> Y        
#> x1       
#> x2       
#> x3       
#> z1       
#> z2       
#> z3       
#> y1       
#> y2       
#> y3  2.126
#> 
#> $cor.ov
#>       x1    x2    x3    z1    z2    z3    y1    y2    y3
#> x1 1.000                                                
#> x2 0.826 1.000                                          
#> x3 0.845 0.812 1.000                                    
#> z1 0.172 0.165 0.169 1.000                              
#> z2 0.166 0.160 0.164 0.833 1.000                        
#> z3 0.169 0.162 0.166 0.846 0.820 1.000                  
#> y1 0.448 0.431 0.441 0.407 0.395 0.401 1.000            
#> y2 0.441 0.424 0.434 0.401 0.388 0.395 0.924 1.000      
#> y3 0.444 0.427 0.437 0.404 0.391 0.398 0.931 0.916 1.000
#> 
#> $cor.lv
#>         X     Z   X:Z     Y
#> X   1.000                  
#> Z   0.200 1.000            
#> X:Z 0.000 0.000 1.000      
#> Y   0.499 0.453 0.464 1.000
#> 
#> $cor.all
#>         X     Z   X:Z     Y    x1    x2    x3    z1    z2    z3    y1    y2
#> X   1.000                                                                  
#> Z   0.200 1.000                                                            
#> X:Z 0.000 0.000 1.000                                                      
#> Y   0.499 0.453 0.464 1.000                                                
#> x1  0.927 0.185 0.000 0.462 1.000                                          
#> x2  0.891 0.178 0.000 0.445 0.826 1.000                                    
#> x3  0.912 0.182 0.000 0.455 0.845 0.812 1.000                              
#> z1  0.185 0.927 0.000 0.420 0.172 0.165 0.169 1.000                        
#> z2  0.179 0.898 0.000 0.407 0.166 0.160 0.164 0.833 1.000                  
#> z3  0.182 0.913 0.000 0.414 0.169 0.162 0.166 0.846 0.820 1.000            
#> y1  0.483 0.439 0.450 0.969 0.448 0.431 0.441 0.407 0.395 0.401 1.000      
#> y2  0.476 0.432 0.443 0.954 0.441 0.424 0.434 0.401 0.388 0.395 0.924 1.000
#> y3  0.479 0.436 0.446 0.961 0.444 0.427 0.437 0.404 0.391 0.398 0.931 0.916
#>        y3
#> X        
#> Z        
#> X:Z      
#> Y        
#> x1       
#> x2       
#> x3       
#> z1       
#> z2       
#> z3       
#> y1       
#> y2       
#> y3  1.000
#> 
#> $mean.lv
#>        ~1
#> X   0.000
#> Z   0.000
#> X:Z 0.199
#> Y   0.142
#> 
#> $mean.ov
#>       ~1
#> x1 1.022
#> x2 1.215
#> x3 0.919
#> z1 1.011
#> z2 1.205
#> z3 0.915
#> y1 1.178
#> y2 1.333
#> y3 1.080
#> 
#> $mean.all
#>        ~1
#> X   0.000
#> Z   0.000
#> X:Z 0.199
#> Y   0.142
#> x1  1.022
#> x2  1.215
#> x3  0.919
#> z1  1.011
#> z2  1.205
#> z3  0.915
#> y1  1.178
#> y2  1.333
#> y3  1.080
#> 
#> $r2.all
#>     Y    x1    x2    x3    z1    z2    z3    y1    y2    y3 
#> 0.595 0.858 0.794 0.831 0.859 0.807 0.833 0.938 0.909 0.923 
#> 
#> $r2.lv
#>     Y 
#> 0.595 
#> 
#> $r2.ov
#>    x1    x2    x3    z1    z2    z3    y1    y2    y3 
#> 0.858 0.794 0.831 0.859 0.807 0.833 0.938 0.909 0.923 
#> 
#> $res.all
#>     Y    x1    x2    x3    z1    z2    z3    y1    y2    y3 
#> 0.405 0.142 0.206 0.169 0.141 0.193 0.167 0.062 0.091 0.077 
#> 
#> $res.lv
#>     Y 
#> 0.405 
#> 
#> $res.ov
#>    x1    x2    x3    z1    z2    z3    y1    y2    y3 
#> 0.142 0.206 0.169 0.141 0.193 0.167 0.062 0.091 0.077 
#> 
#>  
# }