Define or disable the color theme used by modsem

All arguments are optional; omitted ones fall back to the defaults below. Pass active = FALSE to turn highlighting off (and reset the palette).

Usage

set_modsem_colors(
  positive = "green3",
  negative = positive,
  true = "green3",
  false = "red",
  nan = "purple",
  na = "purple",
  inf = "purple",
  string = "cyan",
  active = TRUE
)

Arguments

positive

color of positive numbers.

negative

color of negative numbers.

true

color of TRUE.

false

color of FALSE.

nan

color of NaN.

na

color of NA.

inf

color of -Inf and Inf.

string

color of quoted strings.

active

Should color-theme be activated/deactived?

Value

TRUE if colors are active afterwards, otherwise FALSE.

Examples

set_modsem_colors(positive = "red3", 
                  negative = "red3",
                  true = "darkgreen", 
                  false = "red3", 
                  na = "purple",
                  string = "darkgreen")
#> [1] TRUE

m1 <- "
# Outer Model
  X =~ x1 + x2 + x3
  Z =~ z1 + z2 + z3
  Y =~ y1 + y2 + y3
# Inner Model
  Y ~ X + Z + X:Z
"

est <- modsem(m1, data = oneInt)
colorize_output(summary(est))
#> Estimating baseline model (H0)
#> modsem (version 1.0.11, approach = dblcent):
#> 
#> Interaction Model Fit Measures (H1):
#>   Loglikelihood                              -26807.61 
#>   Akaike (AIC)                                53735.22 
#>   Bayesian (BIC)                              54071.28 
#>   Chi-square                                    122.92 
#>   Degrees of Freedom                               111 
#>   P-value (Chi-square)                           0.207 
#>   RMSEA                                          0.007 
#>   CFI                                            1.000 
#>   SRMR                                           0.008 
#> 
#> Fit Measures for Baseline Model (H0):
#>   Loglikelihood                              -27137.74 
#>   Akaike (AIC)                                54393.48 
#>   Bayesian (BIC)                              54723.93 
#>   Chi-square                                    783.18 
#>   Degrees of Freedom                               112 
#>   P-value (Chi-square)                           0.000 
#>   RMSEA                                          0.055 
#>   CFI                                            0.987 
#>   SRMR                                           0.141 
#> 
#> Comparative Fit to H0 (LRT test):
#>   Chi-square diff                              660.257 
#>   Degrees of freedom diff                            1 
#>   P-value (LRT)                                  0.000 
#> 
#> R-Squared Interaction Model (H1):
#>   Y                                              0.602 
#> R-Squared Baseline Model (H0):
#>   Y                                              0.397 
#> R-Squared Change (H1 - H0):
#>   Y                                              0.204 
#> 
#> lavaan 0.6-19 ended normally after 159 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        60
#> 
#>   Number of observations                          2000
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                               122.924
#>   Degrees of freedom                               111
#>   P-value (Chi-square)                           0.207
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   X =~                                                
#>     x1                1.000                           
#>     x2                0.804    0.013   63.612    0.000
#>     x3                0.916    0.014   67.144    0.000
#>   Z =~                                                
#>     z1                1.000                           
#>     z2                0.812    0.013   64.763    0.000
#>     z3                0.882    0.013   67.014    0.000
#>   Y =~                                                
#>     y1                1.000                           
#>     y2                0.798    0.007  107.428    0.000
#>     y3                0.899    0.008  112.453    0.000
#>   XZ =~                                               
#>     x1z1              1.000                           
#>     x2z1              0.805    0.013   60.636    0.000
#>     x3z1              0.877    0.014   62.680    0.000
#>     x1z2              0.793    0.013   59.343    0.000
#>     x2z2              0.646    0.015   43.672    0.000
#>     x3z2              0.706    0.016   44.292    0.000
#>     x1z3              0.887    0.014   63.700    0.000
#>     x2z3              0.716    0.016   45.645    0.000
#>     x3z3              0.781    0.017   45.339    0.000
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   Y ~                                                 
#>     X                 0.675    0.027   25.379    0.000
#>     Z                 0.561    0.026   21.606    0.000
#>     XZ                0.702    0.027   26.360    0.000
#> 
#> Covariances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>  .x1z1 ~~                                             
#>    .x1z2              0.115    0.008   14.802    0.000
#>    .x1z3              0.114    0.008   13.947    0.000
#>    .x2z1              0.125    0.008   16.095    0.000
#>    .x3z1              0.140    0.009   16.135    0.000
#>  .x1z2 ~~                                             
#>    .x1z3              0.103    0.007   14.675    0.000
#>    .x2z2              0.128    0.006   20.850    0.000
#>    .x3z2              0.146    0.007   21.243    0.000
#>  .x1z3 ~~                                             
#>    .x2z3              0.116    0.007   17.818    0.000
#>    .x3z3              0.135    0.007   18.335    0.000
#>  .x2z1 ~~                                             
#>    .x2z2              0.135    0.006   20.905    0.000
#>    .x2z3              0.145    0.007   21.145    0.000
#>    .x3z1              0.114    0.007   16.058    0.000
#>  .x2z2 ~~                                             
#>    .x2z3              0.117    0.006   20.419    0.000
#>    .x3z2              0.116    0.006   20.586    0.000
#>  .x2z3 ~~                                             
#>    .x3z3              0.109    0.006   18.059    0.000
#>  .x3z1 ~~                                             
#>    .x3z2              0.138    0.007   19.331    0.000
#>    .x3z3              0.158    0.008   20.269    0.000
#>  .x3z2 ~~                                             
#>    .x3z3              0.131    0.007   19.958    0.000
#>   X ~~                                                
#>     Z                 0.201    0.024    8.271    0.000
#>     XZ                0.016    0.025    0.628    0.530
#>   Z ~~                                                
#>     XZ                0.062    0.025    2.449    0.014
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .x1                0.160    0.009   17.871    0.000
#>    .x2                0.162    0.007   22.969    0.000
#>    .x3                0.163    0.008   20.161    0.000
#>    .z1                0.168    0.009   18.143    0.000
#>    .z2                0.158    0.007   22.264    0.000
#>    .z3                0.158    0.008   20.389    0.000
#>    .y1                0.159    0.009   17.896    0.000
#>    .y2                0.154    0.007   22.640    0.000
#>    .y3                0.164    0.008   20.698    0.000
#>    .x1z1              0.311    0.014   22.227    0.000
#>    .x2z1              0.292    0.011   27.287    0.000
#>    .x3z1              0.327    0.012   26.275    0.000
#>    .x1z2              0.290    0.011   26.910    0.000
#>    .x2z2              0.239    0.008   29.770    0.000
#>    .x3z2              0.270    0.009   29.117    0.000
#>    .x1z3              0.272    0.012   23.586    0.000
#>    .x2z3              0.245    0.009   27.979    0.000
#>    .x3z3              0.297    0.011   28.154    0.000
#>     X                 0.981    0.036   26.895    0.000
#>     Z                 1.016    0.038   26.856    0.000
#>    .Y                 0.990    0.038   25.926    0.000
#>     XZ                1.045    0.044   24.004    0.000
#> 
#>  
colorize_output(est) # same as colorize_output(print(est))
#> modsem (version 1.0.11, approach = dblcent):
#> lavaan 0.6-19 ended normally after 159 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        60
#> 
#>   Number of observations                          2000
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                               122.924
#>   Degrees of freedom                               111
#>   P-value (Chi-square)                           0.207
#>  
colorize_output(modsem_inspect(est, what = "coef"))
#> $lambda
#>          X     Z     Y    XZ
#> x1   1.000 0.000 0.000 0.000
#> x2   0.804 0.000 0.000 0.000
#> x3   0.916 0.000 0.000 0.000
#> z1   0.000 1.000 0.000 0.000
#> z2   0.000 0.812 0.000 0.000
#> z3   0.000 0.882 0.000 0.000
#> y1   0.000 0.000 1.000 0.000
#> y2   0.000 0.000 0.798 0.000
#> y3   0.000 0.000 0.899 0.000
#> x1z1 0.000 0.000 0.000 1.000
#> x2z1 0.000 0.000 0.000 0.805
#> x3z1 0.000 0.000 0.000 0.877
#> x1z2 0.000 0.000 0.000 0.793
#> x2z2 0.000 0.000 0.000 0.646
#> x3z2 0.000 0.000 0.000 0.706
#> x1z3 0.000 0.000 0.000 0.887
#> x2z3 0.000 0.000 0.000 0.716
#> x3z3 0.000 0.000 0.000 0.781
#> 
#> $theta
#>         x1    x2    x3    z1    z2    z3    y1    y2    y3  x1z1  x2z1  x3z1
#> x1   0.160                                                                  
#> x2   0.000 0.162                                                            
#> x3   0.000 0.000 0.163                                                      
#> z1   0.000 0.000 0.000 0.168                                                
#> z2   0.000 0.000 0.000 0.000 0.158                                          
#> z3   0.000 0.000 0.000 0.000 0.000 0.158                                    
#> y1   0.000 0.000 0.000 0.000 0.000 0.000 0.159                              
#> y2   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.154                        
#> y3   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.164                  
#> x1z1 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.311            
#> x2z1 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.125 0.292      
#> x3z1 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.140 0.114 0.327
#> x1z2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.115 0.000 0.000
#> x2z2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.135 0.000
#> x3z2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.138
#> x1z3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.114 0.000 0.000
#> x2z3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.145 0.000
#> x3z3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.158
#>       x1z2  x2z2  x3z2  x1z3  x2z3  x3z3
#> x1                                      
#> x2                                      
#> x3                                      
#> z1                                      
#> z2                                      
#> z3                                      
#> y1                                      
#> y2                                      
#> y3                                      
#> x1z1                                    
#> x2z1                                    
#> x3z1                                    
#> x1z2 0.290                              
#> x2z2 0.128 0.239                        
#> x3z2 0.146 0.116 0.270                  
#> x1z3 0.103 0.000 0.000 0.272            
#> x2z3 0.000 0.117 0.000 0.116 0.245      
#> x3z3 0.000 0.000 0.131 0.135 0.109 0.297
#> 
#> $psi
#>        X     Z     Y    XZ
#> X  0.981                  
#> Z  0.201 1.016            
#> Y  0.000 0.000 0.990      
#> XZ 0.016 0.062 0.000 1.045
#> 
#> $beta
#>        X     Z Y    XZ
#> X  0.000 0.000 0 0.000
#> Z  0.000 0.000 0 0.000
#> Y  0.675 0.561 0 0.702
#> XZ 0.000 0.000 0 0.000
#> 
#>  

# \dontrun{
colorize_output(split = TRUE, {
  # Get live (uncolored) output
  # And print colored output at the end of execution

  est_lms <- modsem(m1, data = oneInt, method = "lms")
  summary(est_lms)
}) 
#> 
#> modsem (version 1.0.11):
#> 
#>   Estimator                                         LMS
#>   Optimization method                        EMA-NLMINB
#>   Number of observations                           2000
#>   Number of iterations                               45
#>   Loglikelihood                               -14687.69
#>   Akaike (AIC)                                 29437.39
#>   Bayesian (BIC)                               29611.01
#>  
#> Numerical Integration:
#>   Points of integration (per dim)                    24
#>   Dimensions                                          1
#>   Total points of integration                        24
#>  
#> Fit Measures for Baseline Model (H0):
#>   Loglikelihood                               -17831.87
#>   Akaike (AIC)                                 35723.75
#>   Bayesian (BIC)                               35891.78
#>   Chi-square                                      17.52
#>   Degrees of Freedom (Chi-square)                    24
#>   P-value (Chi-square)                            0.826
#>   RMSEA                                           0.000
#>  
#> Comparative Fit to H0 (LRT test):
#>   Loglikelihood change                          3144.18
#>   Difference test (D)                           6288.36
#>   Degrees of freedom (D)                              1
#>   P-value (D)                                     0.000
#>  
#> R-Squared Interaction Model (H1):
#>   Y                                               0.595
#> R-Squared Baseline Model (H0):
#>   Y                                               0.395
#> R-Squared Change (H1 - H0):
#>   Y                                               0.199
#> 
#> Parameter Estimates:
#>   Coefficients                           unstandardized
#>   Information                                  observed
#>   Standard errors                              standard
#>  
#> Latent Variables:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>   X =~          
#>     x1              1.000                             
#>     x2              0.804      0.012    64.32    0.000
#>     x3              0.915      0.013    67.94    0.000
#>   Z =~          
#>     z1              1.000                             
#>     z2              0.810      0.012    65.09    0.000
#>     z3              0.881      0.013    67.60    0.000
#>   Y =~          
#>     y1              1.000                             
#>     y2              0.798      0.007   107.55    0.000
#>     y3              0.899      0.008   112.58    0.000
#> 
#> Regressions:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>   Y ~           
#>     X               0.672      0.031    21.71    0.000
#>     Z               0.569      0.030    19.05    0.000
#>     X:Z             0.713      0.028    25.72    0.000
#> 
#> Intercepts:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>    .x1              1.022      0.021    47.66    0.000
#>    .x2              1.215      0.018    67.19    0.000
#>    .x3              0.919      0.020    45.92    0.000
#>    .z1              1.010      0.024    41.69    0.000
#>    .z2              1.205      0.020    59.44    0.000
#>    .z3              0.915      0.022    42.18    0.000
#>    .y1              1.036      0.032    32.28    0.000
#>    .y2              1.220      0.026    46.66    0.000
#>    .y3              0.953      0.029    32.68    0.000
#> 
#> Covariances:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>   X ~~          
#>     Z               0.199      0.024     8.32    0.000
#> 
#> Variances:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>    .x1              0.160      0.008    19.30    0.000
#>    .x2              0.163      0.007    23.88    0.000
#>    .x3              0.165      0.008    21.23    0.000
#>    .z1              0.166      0.009    18.47    0.000
#>    .z2              0.160      0.007    22.67    0.000
#>    .z3              0.158      0.008    20.80    0.000
#>    .y1              0.160      0.009    18.01    0.000
#>    .y2              0.154      0.007    22.68    0.000
#>    .y3              0.164      0.008    20.68    0.000
#>     X               0.972      0.033    29.85    0.000
#>     Z               1.017      0.038    26.95    0.000
#>    .Y               0.984      0.038    26.00    0.000
#> 
#> 
#> modsem (version 1.0.11):
#> 
#>   Estimator                                         LMS
#>   Optimization method                        EMA-NLMINB
#>   Number of observations                           2000
#>   Number of iterations                               45
#>   Loglikelihood                               -14687.69
#>   Akaike (AIC)                                 29437.39
#>   Bayesian (BIC)                               29611.01
#>  
#> Numerical Integration:
#>   Points of integration (per dim)                    24
#>   Dimensions                                          1
#>   Total points of integration                        24
#>  
#> Fit Measures for Baseline Model (H0):
#>   Loglikelihood                               -17831.87
#>   Akaike (AIC)                                 35723.75
#>   Bayesian (BIC)                               35891.78
#>   Chi-square                                      17.52
#>   Degrees of Freedom (Chi-square)                    24
#>   P-value (Chi-square)                            0.826
#>   RMSEA                                           0.000
#>  
#> Comparative Fit to H0 (LRT test):
#>   Loglikelihood change                          3144.18
#>   Difference test (D)                           6288.36
#>   Degrees of freedom (D)                              1
#>   P-value (D)                                     0.000
#>  
#> R-Squared Interaction Model (H1):
#>   Y                                               0.595
#> R-Squared Baseline Model (H0):
#>   Y                                               0.395
#> R-Squared Change (H1 - H0):
#>   Y                                               0.199
#> 
#> Parameter Estimates:
#>   Coefficients                           unstandardized
#>   Information                                  observed
#>   Standard errors                              standard
#>  
#> Latent Variables:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>   X =~          
#>     x1              1.000                             
#>     x2              0.804      0.012    64.32    0.000
#>     x3              0.915      0.013    67.94    0.000
#>   Z =~          
#>     z1              1.000                             
#>     z2              0.810      0.012    65.09    0.000
#>     z3              0.881      0.013    67.60    0.000
#>   Y =~          
#>     y1              1.000                             
#>     y2              0.798      0.007   107.55    0.000
#>     y3              0.899      0.008   112.58    0.000
#> 
#> Regressions:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>   Y ~           
#>     X               0.672      0.031    21.71    0.000
#>     Z               0.569      0.030    19.05    0.000
#>     X:Z             0.713      0.028    25.72    0.000
#> 
#> Intercepts:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>    .x1              1.022      0.021    47.66    0.000
#>    .x2              1.215      0.018    67.19    0.000
#>    .x3              0.919      0.020    45.92    0.000
#>    .z1              1.010      0.024    41.69    0.000
#>    .z2              1.205      0.020    59.44    0.000
#>    .z3              0.915      0.022    42.18    0.000
#>    .y1              1.036      0.032    32.28    0.000
#>    .y2              1.220      0.026    46.66    0.000
#>    .y3              0.953      0.029    32.68    0.000
#> 
#> Covariances:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>   X ~~          
#>     Z               0.199      0.024     8.32    0.000
#> 
#> Variances:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>    .x1              0.160      0.008    19.30    0.000
#>    .x2              0.163      0.007    23.88    0.000
#>    .x3              0.165      0.008    21.23    0.000
#>    .z1              0.166      0.009    18.47    0.000
#>    .z2              0.160      0.007    22.67    0.000
#>    .z3              0.158      0.008    20.80    0.000
#>    .y1              0.160      0.009    18.01    0.000
#>    .y2              0.154      0.007    22.68    0.000
#>    .y3              0.164      0.008    20.68    0.000
#>     X               0.972      0.033    29.85    0.000
#>     Z               1.017      0.038    26.95    0.000
#>    .Y               0.984      0.038    26.00    0.000
#> 
#>  
                
colorize_output(modsem_inspect(est_lms))
#> $N
#> [1] 2000
#> 
#> $vcov.all
#>        X=~x2 X=~x3 Z=~z2 Z=~z3 Y=~y2 Y=~y3  x1~1  x2~1  x3~1  z1~1  z2~1  z3~1
#> X=~x2  0.000                                                                  
#> X=~x3  0.000 0.000                                                            
#> Z=~z2  0.000 0.000 0.000                                                      
#> Z=~z3  0.000 0.000 0.000 0.000                                                
#> Y=~y2  0.000 0.000 0.000 0.000 0.000                                          
#> Y=~y3  0.000 0.000 0.000 0.000 0.000 0.000                                    
#> x1~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000                              
#> x2~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000                        
#> x3~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000                  
#> z1~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.001            
#> z2~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000      
#> z3~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y1~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y2~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y3~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> x1~~x1 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> x2~~x2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> x3~~x3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> z1~~z1 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> z2~~z2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> z3~~z3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y1~~y1 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y2~~y2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y3~~y3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> X~~X   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> X~~Z   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Z~~Z   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Y~~Y   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Y~X    0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Y~Z    0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Y~X:Z  0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#>         y1~1  y2~1  y3~1 x1~~x1 x2~~x2 x3~~x3 z1~~z1 z2~~z2 z3~~z3 y1~~y1
#> X=~x2                                                                    
#> X=~x3                                                                    
#> Z=~z2                                                                    
#> Z=~z3                                                                    
#> Y=~y2                                                                    
#> Y=~y3                                                                    
#> x1~1                                                                     
#> x2~1                                                                     
#> x3~1                                                                     
#> z1~1                                                                     
#> z2~1                                                                     
#> z3~1                                                                     
#> y1~1   0.001                                                             
#> y2~1   0.001 0.001                                                       
#> y3~1   0.001 0.001 0.001                                                 
#> x1~~x1 0.000 0.000 0.000  0.000                                          
#> x2~~x2 0.000 0.000 0.000  0.000  0.000                                   
#> x3~~x3 0.000 0.000 0.000  0.000  0.000  0.000                            
#> z1~~z1 0.000 0.000 0.000  0.000  0.000  0.000  0.000                     
#> z2~~z2 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000              
#> z3~~z3 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000       
#> y1~~y1 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> y2~~y2 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> y3~~y3 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> X~~X   0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> X~~Z   0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Z~~Z   0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Y~~Y   0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Y~X    0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Y~Z    0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Y~X:Z  0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#>        y2~~y2 y3~~y3  X~~X  X~~Z  Z~~Z  Y~~Y   Y~X   Y~Z Y~X:Z
#> X=~x2                                                         
#> X=~x3                                                         
#> Z=~z2                                                         
#> Z=~z3                                                         
#> Y=~y2                                                         
#> Y=~y3                                                         
#> x1~1                                                          
#> x2~1                                                          
#> x3~1                                                          
#> z1~1                                                          
#> z2~1                                                          
#> z3~1                                                          
#> y1~1                                                          
#> y2~1                                                          
#> y3~1                                                          
#> x1~~x1                                                        
#> x2~~x2                                                        
#> x3~~x3                                                        
#> z1~~z1                                                        
#> z2~~z2                                                        
#> z3~~z3                                                        
#> y1~~y1                                                        
#> y2~~y2  0.000                                                 
#> y3~~y3  0.000  0.000                                          
#> X~~X    0.000  0.000 0.001                                    
#> X~~Z    0.000  0.000 0.000 0.001                              
#> Z~~Z    0.000  0.000 0.000 0.000 0.001                        
#> Y~~Y    0.000  0.000 0.000 0.000 0.000 0.001                  
#> Y~X     0.000  0.000 0.000 0.000 0.000 0.000 0.001            
#> Y~Z     0.000  0.000 0.000 0.000 0.000 0.000 0.000 0.001      
#> Y~X:Z   0.000  0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.001
#> 
#> $vcov.free
#>        X=~x2 X=~x3 Z=~z2 Z=~z3 Y=~y2 Y=~y3  x1~1  x2~1  x3~1  z1~1  z2~1  z3~1
#> X=~x2  0.000                                                                  
#> X=~x3  0.000 0.000                                                            
#> Z=~z2  0.000 0.000 0.000                                                      
#> Z=~z3  0.000 0.000 0.000 0.000                                                
#> Y=~y2  0.000 0.000 0.000 0.000 0.000                                          
#> Y=~y3  0.000 0.000 0.000 0.000 0.000 0.000                                    
#> x1~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000                              
#> x2~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000                        
#> x3~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000                  
#> z1~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.001            
#> z2~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000      
#> z3~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y1~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y2~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y3~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> x1~~x1 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> x2~~x2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> x3~~x3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> z1~~z1 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> z2~~z2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> z3~~z3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y1~~y1 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y2~~y2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y3~~y3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> X~~X   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> X~~Z   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Z~~Z   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Y~~Y   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Y~X    0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Y~Z    0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Y~X:Z  0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#>         y1~1  y2~1  y3~1 x1~~x1 x2~~x2 x3~~x3 z1~~z1 z2~~z2 z3~~z3 y1~~y1
#> X=~x2                                                                    
#> X=~x3                                                                    
#> Z=~z2                                                                    
#> Z=~z3                                                                    
#> Y=~y2                                                                    
#> Y=~y3                                                                    
#> x1~1                                                                     
#> x2~1                                                                     
#> x3~1                                                                     
#> z1~1                                                                     
#> z2~1                                                                     
#> z3~1                                                                     
#> y1~1   0.001                                                             
#> y2~1   0.001 0.001                                                       
#> y3~1   0.001 0.001 0.001                                                 
#> x1~~x1 0.000 0.000 0.000  0.000                                          
#> x2~~x2 0.000 0.000 0.000  0.000  0.000                                   
#> x3~~x3 0.000 0.000 0.000  0.000  0.000  0.000                            
#> z1~~z1 0.000 0.000 0.000  0.000  0.000  0.000  0.000                     
#> z2~~z2 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000              
#> z3~~z3 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000       
#> y1~~y1 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> y2~~y2 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> y3~~y3 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> X~~X   0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> X~~Z   0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Z~~Z   0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Y~~Y   0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Y~X    0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Y~Z    0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Y~X:Z  0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#>        y2~~y2 y3~~y3  X~~X  X~~Z  Z~~Z  Y~~Y   Y~X   Y~Z Y~X:Z
#> X=~x2                                                         
#> X=~x3                                                         
#> Z=~z2                                                         
#> Z=~z3                                                         
#> Y=~y2                                                         
#> Y=~y3                                                         
#> x1~1                                                          
#> x2~1                                                          
#> x3~1                                                          
#> z1~1                                                          
#> z2~1                                                          
#> z3~1                                                          
#> y1~1                                                          
#> y2~1                                                          
#> y3~1                                                          
#> x1~~x1                                                        
#> x2~~x2                                                        
#> x3~~x3                                                        
#> z1~~z1                                                        
#> z2~~z2                                                        
#> z3~~z3                                                        
#> y1~~y1                                                        
#> y2~~y2  0.000                                                 
#> y3~~y3  0.000  0.000                                          
#> X~~X    0.000  0.000 0.000                                    
#> X~~Z    0.000  0.000 0.000 0.001                              
#> Z~~Z    0.000  0.000 0.000 0.000 0.000                        
#> Y~~Y    0.000  0.000 0.000 0.000 0.000 0.001                  
#> Y~X     0.000  0.000 0.000 0.000 0.000 0.000 0.001            
#> Y~Z     0.000  0.000 0.000 0.000 0.000 0.000 0.000 0.001      
#> Y~X:Z   0.000  0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.001
#> 
#> $information
#>            X=~x2     X=~x3     Z=~z2     Z=~z3     Y=~y2     Y=~y3      x1~1
#> X=~x2   8999.190                                                            
#> X=~x3  -2648.545  8222.029                                                  
#> Z=~z2     60.902    70.493  9243.388                                        
#> Z=~z3     71.252    77.496 -3030.926  8771.586                              
#> Y=~y2   -165.574  -191.306  -157.611  -170.558 23807.613                    
#> Y=~y3   -172.967  -201.814  -168.453  -181.178 -8384.538 21002.070          
#> x1~1      15.508    40.408    89.611    97.328  -120.125  -132.193  8025.819
#> x2~1      15.276    31.909    70.886    76.991   -95.023  -104.569 -3521.347
#> x3~1      13.570    39.137    79.506    86.354  -106.575  -117.282 -3949.451
#> z1~1      84.651    93.888    22.866    26.925  -125.738  -132.413    53.106
#> z2~1      71.563    79.371    26.884    22.769  -106.298  -111.940    44.895
#> z3~1      78.359    86.909    21.018    32.943  -116.389  -122.567    49.161
#> y1~1     -24.526   -26.022   -23.832   -28.103  -592.265  -630.022  -172.639
#> y2~1     -20.248   -21.482   -19.674   -23.200  1374.411  -520.140  -142.522
#> y3~1     -21.528   -22.842   -20.919   -24.668  -520.059  1206.216  -151.534
#> x1~~x1 -1289.439 -1451.910    21.448    29.980    -4.641    -5.209   -82.410
#> x2~~x2  2481.964  -928.039    24.922     9.639   -11.517   -16.419    86.859
#> x3~~x3 -1069.089  2558.714    19.187    25.454    16.577    22.711   -18.403
#> z1~~z1    19.569    20.458 -1321.791 -1444.858    16.715    15.756    42.196
#> z2~~z2    32.441    21.370  2852.130 -1036.274    -4.204    -0.944    17.893
#> z3~~z3    13.466    23.856 -1115.966  2972.124   -16.178   -19.220    25.798
#> y1~~y1     0.686     1.939     8.258     6.876 -1490.684 -1622.857     0.367
#> y2~~y2     0.887    -6.790   -12.295    -9.484  3045.959 -1171.731   -14.595
#> y3~~y3   -15.848   -16.324    -9.007   -11.546 -1249.652  2846.127     8.139
#> X~~X    1455.309  1633.339   -62.086   -64.995    71.415    74.789     8.816
#> X~~Z    -112.348  -127.632   111.894   130.248    32.873    35.794   -42.380
#> Z~~Z     -26.074   -28.173  1158.576  1266.678    48.214    51.032   -20.106
#> Y~~Y     -35.958   -52.257   -31.482   -34.576   549.624   589.193   -13.241
#> Y~X     -268.679  -299.211   -42.902   -50.269   434.979   458.108   -71.007
#> Y~Z      -59.845   -62.499  -273.648  -282.614   394.674   429.050  -424.937
#> Y~X:Z   -284.486  -320.308  -316.772  -345.446   417.364   436.778   -67.251
#>             x2~1      x3~1      z1~1      z2~1      z3~1      y1~1      y2~1
#> X=~x2                                                                       
#> X=~x3                                                                       
#> Z=~z2                                                                       
#> Z=~z3                                                                       
#> Y=~y2                                                                       
#> Y=~y3                                                                       
#> x1~1                                                                        
#> x2~1    9492.905                                                            
#> x3~1   -3124.156  8600.468                                                  
#> z1~1      42.010    47.121  7696.689                                        
#> z2~1      35.515    39.836 -3648.671  9449.714                              
#> z3~1      38.889    43.621 -3995.041 -3377.361  8925.955                    
#> y1~1    -136.564  -153.166  -144.456  -122.121  -133.714  7702.488          
#> y2~1    -112.740  -126.446  -119.255  -100.817  -110.387 -3985.927  9668.066
#> y3~1    -119.869  -134.442  -126.796  -107.192  -117.368 -4237.977 -3498.638
#> x1~~x1   -53.570   -59.584    40.065    33.870    37.088    -8.037    -6.635
#> x2~~x2    46.164    77.509    18.823    15.913    17.426     9.778     8.073
#> x3~~x3   -14.410   -31.403    29.480    24.921    27.290     4.077     3.366
#> z1~~z1    33.380    37.441   -17.824   -10.720   -11.188    -1.412    -1.166
#> z2~~z2    14.155    15.877     2.271   -12.830     2.477    -2.104    -1.737
#> z3~~z3    20.407    22.887    -7.798    -6.613   -15.022     8.651     7.141
#> y1~~y1     0.291     0.327    -4.752    -4.017    -4.396   -10.657    -4.253
#> y2~~y2   -11.544   -12.946     3.200     2.705     2.964    -4.777   -20.063
#> y3~~y3     6.439     7.223    -3.345    -2.828    -3.095    -3.457    -2.773
#> X~~X       6.973     7.819   -46.900   -39.649   -43.414    -1.170    -0.966
#> X~~Z     -33.526   -37.609   -39.478   -33.374   -36.542    56.495    46.639
#> Z~~Z     -15.904   -17.838     8.134     6.875     7.560   -13.533   -11.172
#> Y~~Y     -10.471   -11.735   -12.443   -10.520   -11.510    -1.100    -0.905
#> Y~X      -56.174   -63.019  -396.716  -335.379  -367.215   -33.695   -27.816
#> Y~Z     -336.140  -377.005   -70.435   -59.544   -65.210   -38.192   -31.529
#> Y~X:Z    -53.202   -59.682   -54.628   -46.181   -50.579   129.297   106.741
#>             y3~1    x1~~x1    x2~~x2    x3~~x3    z1~~z1    z2~~z2    z3~~z3
#> X=~x2                                                                       
#> X=~x3                                                                       
#> Z=~z2                                                                       
#> Z=~z3                                                                       
#> Y=~y2                                                                       
#> Y=~y3                                                                       
#> x1~1                                                                        
#> x2~1                                                                        
#> x3~1                                                                        
#> z1~1                                                                        
#> z2~1                                                                        
#> z3~1                                                                        
#> y1~1                                                                        
#> y2~1                                                                        
#> y3~1    8512.173                                                            
#> x1~~x1    -7.053 17221.390                                                  
#> x2~~x2     8.584  3380.011 23314.216                                        
#> x3~~x3     3.581  4279.841  2503.248 18942.578                              
#> z1~~z1    -1.238    -2.811     9.332    19.893 14811.361                    
#> z2~~z2    -1.846    46.264   -34.239    40.061  3316.433 22329.520          
#> z3~~z3     7.595    19.344    57.035    23.290  3961.584  2863.717 19903.135
#> y1~~y1    -3.877    32.945     9.681     5.738    18.296    20.634    36.773
#> y2~~y2    -3.678     0.493    -8.990    31.813     2.002    10.327    -9.730
#> y3~~y3   -10.923     2.042    12.780     7.956    13.308    -7.730     0.091
#> X~~X      -1.027   169.622   -39.930   -57.280    20.548   -24.836   -46.353
#> X~~Z      49.588   -26.303   -56.334   -61.782    12.439    13.497   -26.853
#> Z~~Z     -11.878   -12.319     8.118     4.300   266.329   187.832   235.064
#> Y~~Y      -0.949    86.906    34.446   109.302    82.865    55.551    67.186
#> Y~X      -29.574  -137.585   -85.540  -119.272    35.606   -16.398    -1.335
#> Y~Z      -33.522    -0.083    18.193    -2.679  -133.633   -43.316  -146.369
#> Y~X:Z    113.493  -137.397   -87.489  -102.954   -92.400  -117.807  -148.859
#>           y1~~y1    y2~~y2    y3~~y3      X~~X      X~~Z      Z~~Z      Y~~Y
#> X=~x2                                                                       
#> X=~x3                                                                       
#> Z=~z2                                                                       
#> Z=~z3                                                                       
#> Y=~y2                                                                       
#> Y=~y3                                                                       
#> x1~1                                                                        
#> x2~1                                                                        
#> x3~1                                                                        
#> z1~1                                                                        
#> z2~1                                                                        
#> z3~1                                                                        
#> y1~1                                                                        
#> y2~1                                                                        
#> y3~1                                                                        
#> x1~~x1                                                                      
#> x2~~x2                                                                      
#> x3~~x3                                                                      
#> z1~~z1                                                                      
#> z2~~z2                                                                      
#> z3~~z3                                                                      
#> y1~~y1 14839.780                                                            
#> y2~~y2  3729.164 23375.700                                                  
#> y3~~y3  4418.945  3352.292 18124.253                                        
#> X~~X      -9.652    -2.386     3.523  4771.846                              
#> X~~Z       0.487     1.760    -2.659  -298.352  1831.377                    
#> Z~~Z      -5.175     3.447     5.140     2.250    46.067  3588.092          
#> Y~~Y     171.646    77.582    60.885   -20.420    -1.322     8.379   758.331
#> Y~X       18.103   -10.904    11.537   178.209    36.920    -1.265    47.432
#> Y~Z      -12.672    40.115    -6.635    19.842    64.126    87.886    47.299
#> Y~X:Z    -31.044    16.351    56.637   169.618    55.908   110.469   104.746
#>              Y~X       Y~Z     Y~X:Z
#> X=~x2                               
#> X=~x3                               
#> Z=~z2                               
#> Z=~z3                               
#> Y=~y2                               
#> Y=~y3                               
#> x1~1                                
#> x2~1                                
#> x3~1                                
#> z1~1                                
#> z2~1                                
#> z3~1                                
#> y1~1                                
#> y2~1                                
#> y3~1                                
#> x1~~x1                              
#> x2~~x2                              
#> x3~~x3                              
#> z1~~z1                              
#> z2~~z2                              
#> z3~~z3                              
#> y1~~y1                              
#> y2~~y2                              
#> y3~~y3                              
#> X~~X                                
#> X~~Z                                
#> Z~~Z                                
#> Y~~Y                                
#> Y~X     1548.428                    
#> Y~Z      319.958  1581.433          
#> Y~X:Z    -51.195   -26.969  1518.663
#> 
#> $coefficients.all
#>  X=~x2  X=~x3  Z=~z2  Z=~z3  Y=~y2  Y=~y3   x1~1   x2~1   x3~1   z1~1   z2~1 
#>  0.804  0.915  0.810  0.881  0.798  0.899  1.022  1.215  0.919  1.010  1.205 
#>   z3~1   y1~1   y2~1   y3~1 x1~~x1 x2~~x2 x3~~x3 z1~~z1 z2~~z2 z3~~z3 y1~~y1 
#>  0.915  1.036  1.220  0.953  0.160  0.163  0.165  0.166  0.160  0.158  0.160 
#> y2~~y2 y3~~y3   X~~X   X~~Z   Z~~Z   Y~~Y    Y~X    Y~Z  Y~X:Z 
#>  0.154  0.164  0.972  0.199  1.017  0.984  0.672  0.569  0.713 
#> 
#> $coefficients.free
#>  X=~x2  X=~x3  Z=~z2  Z=~z3  Y=~y2  Y=~y3   x1~1   x2~1   x3~1   z1~1   z2~1 
#>  0.804  0.915  0.810  0.881  0.798  0.899  1.022  1.215  0.919  1.010  1.205 
#>   z3~1   y1~1   y2~1   y3~1 x1~~x1 x2~~x2 x3~~x3 z1~~z1 z2~~z2 z3~~z3 y1~~y1 
#>  0.915  1.036  1.220  0.953  0.160  0.163  0.165  0.166  0.160  0.158  0.160 
#> y2~~y2 y3~~y3   X~~X   X~~Z   Z~~Z   Y~~Y    Y~X    Y~Z  Y~X:Z 
#>  0.154  0.164  0.972  0.199  1.017  0.984  0.672  0.569  0.713 
#> 
#> $partable
#>    lhs op rhs label   est std.error z.value p.value ci.lower ci.upper
#> 1    X =~  x1       1.000        NA      NA      NA       NA       NA
#> 2    X =~  x2       0.804     0.012  64.316       0    0.779    0.828
#> 3    X =~  x3       0.915     0.013  67.943       0    0.888    0.941
#> 4    Z =~  z1       1.000        NA      NA      NA       NA       NA
#> 5    Z =~  z2       0.810     0.012  65.092       0    0.786    0.835
#> 6    Z =~  z3       0.881     0.013  67.604       0    0.855    0.906
#> 7    Y =~  y1       1.000        NA      NA      NA       NA       NA
#> 8    Y =~  y2       0.798     0.007 107.545       0    0.784    0.813
#> 9    Y =~  y3       0.899     0.008 112.578       0    0.884    0.915
#> 10   Y  ~   X       0.672     0.031  21.706       0    0.611    0.732
#> 11   Y  ~   Z       0.569     0.030  19.055       0    0.511    0.628
#> 12   Y  ~ X:Z       0.713     0.028  25.720       0    0.659    0.767
#> 13  x1 ~1           1.022     0.021  47.659       0    0.980    1.064
#> 14  x2 ~1           1.215     0.018  67.195       0    1.180    1.251
#> 15  x3 ~1           0.919     0.020  45.922       0    0.880    0.958
#> 16  z1 ~1           1.010     0.024  41.689       0    0.963    1.058
#> 17  z2 ~1           1.205     0.020  59.444       0    1.165    1.245
#> 18  z3 ~1           0.915     0.022  42.177       0    0.872    0.957
#> 19  y1 ~1           1.036     0.032  32.283       0    0.973    1.099
#> 20  y2 ~1           1.220     0.026  46.665       0    1.168    1.271
#> 21  y3 ~1           0.953     0.029  32.684       0    0.896    1.010
#> 22  x1 ~~  x1       0.160     0.008  19.299       0    0.144    0.177
#> 23  x2 ~~  x2       0.163     0.007  23.882       0    0.150    0.176
#> 24  x3 ~~  x3       0.165     0.008  21.229       0    0.150    0.180
#> 25  z1 ~~  z1       0.166     0.009  18.473       0    0.149    0.184
#> 26  z2 ~~  z2       0.160     0.007  22.670       0    0.146    0.173
#> 27  z3 ~~  z3       0.158     0.008  20.795       0    0.143    0.173
#> 28  y1 ~~  y1       0.160     0.009  18.011       0    0.142    0.177
#> 29  y2 ~~  y2       0.154     0.007  22.681       0    0.141    0.168
#> 30  y3 ~~  y3       0.164     0.008  20.680       0    0.148    0.179
#> 31   X ~~   X       0.972     0.033  29.853       0    0.908    1.036
#> 32   X ~~   Z       0.199     0.024   8.322       0    0.152    0.245
#> 33   Z ~~   Z       1.017     0.038  26.945       0    0.943    1.091
#> 34   Y ~~   Y       0.984     0.038  25.999       0    0.909    1.058
#> 
#> $partable.input
#>    lhs op rhs mod
#> 1    X =~  x1    
#> 2    X =~  x2    
#> 3    X =~  x3    
#> 4    Z =~  z1    
#> 5    Z =~  z2    
#> 6    Z =~  z3    
#> 7    Y =~  y1    
#> 8    Y =~  y2    
#> 9    Y =~  y3    
#> 10   Y  ~   X    
#> 11   Y  ~   Z    
#> 12   Y  ~ X:Z    
#> 
#> $loglik
#> [1] -14687.69
#> 
#> $iterations
#> [1] 45
#> 
#> $convergence
#> [1] TRUE
#> 
#> $lambda
#>        X     Z     Y
#> x1 1.000 0.000 0.000
#> x2 0.804 0.000 0.000
#> x3 0.915 0.000 0.000
#> z1 0.000 1.000 0.000
#> z2 0.000 0.810 0.000
#> z3 0.000 0.881 0.000
#> y1 0.000 0.000 1.000
#> y2 0.000 0.000 0.798
#> y3 0.000 0.000 0.899
#> 
#> $tau
#>       ~1
#> x1 1.022
#> x2 1.215
#> x3 0.919
#> z1 1.010
#> z2 1.205
#> z3 0.915
#> y1 1.036
#> y2 1.220
#> y3 0.953
#> 
#> $theta
#>       x1    x2    x3    z1    z2    z3    y1    y2    y3
#> x1 0.160                                                
#> x2 0.000 0.163                                          
#> x3 0.000 0.000 0.165                                    
#> z1 0.000 0.000 0.000 0.166                              
#> z2 0.000 0.000 0.000 0.000 0.160                        
#> z3 0.000 0.000 0.000 0.000 0.000 0.158                  
#> y1 0.000 0.000 0.000 0.000 0.000 0.000 0.160            
#> y2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.154      
#> y3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.164
#> 
#> $gamma.xi
#>       X     Z
#> Y 0.672 0.569
#> 
#> $gamma.eta
#>       Y
#> Y 0.000
#> 
#> $omega.xi.xi
#>         X     Z
#> Y~X 0.000 0.713
#> Y~Z 0.000 0.000
#> 
#> $omega.eta.xi
#>         Y
#> Y~X 0.000
#> Y~Z 0.000
#> 
#> $phi
#>       X     Z
#> X 0.972      
#> Z 0.199 1.017
#> 
#> $psi
#>       Y
#> Y 0.984
#> 
#> $alpha
#>      ~1
#> Y 0.000
#> 
#> $beta0
#>      ~1
#> X 0.000
#> Z 0.000
#> 
#> $cov.ov
#>       x1    x2    x3    z1    z2    z3    y1    y2    y3
#> x1 1.132                                                
#> x2 0.781 0.791                                          
#> x3 0.889 0.714 0.978                                    
#> z1 0.199 0.160 0.182 1.184                              
#> z2 0.161 0.129 0.147 0.824 0.827                        
#> z3 0.175 0.141 0.160 0.896 0.726 0.948                  
#> y1 0.766 0.616 0.701 0.713 0.577 0.628 2.586            
#> y2 0.612 0.492 0.559 0.569 0.461 0.501 1.937 1.701      
#> y3 0.689 0.554 0.630 0.641 0.519 0.564 2.182 1.742 2.125
#> 
#> $cov.lv
#>         X     Z   X:Z     Y
#> X   0.972                  
#> Z   0.199 1.017            
#> X:Z 0.000 0.000 1.028      
#> Y   0.766 0.713 0.733 2.427
#> 
#> $cov.all
#>         X     Z   X:Z     Y    x1    x2    x3    z1    z2    z3    y1    y2
#> X   0.972                                                                  
#> Z   0.199 1.017                                                            
#> X:Z 0.000 0.000 1.028                                                      
#> Y   0.766 0.713 0.733 2.427                                                
#> x1  0.972 0.199 0.000 0.766 1.132                                          
#> x2  0.781 0.160 0.000 0.616 0.781 0.791                                    
#> x3  0.889 0.182 0.000 0.701 0.889 0.714 0.978                              
#> z1  0.199 1.017 0.000 0.713 0.199 0.160 0.182 1.184                        
#> z2  0.161 0.824 0.000 0.577 0.161 0.129 0.147 0.824 0.827                  
#> z3  0.175 0.896 0.000 0.628 0.175 0.141 0.160 0.896 0.726 0.948            
#> y1  0.766 0.713 0.733 2.427 0.766 0.616 0.701 0.713 0.577 0.628 2.586      
#> y2  0.612 0.569 0.585 1.937 0.612 0.492 0.559 0.569 0.461 0.501 1.937 1.701
#> y3  0.689 0.641 0.659 2.182 0.689 0.554 0.630 0.641 0.519 0.564 2.182 1.742
#>        y3
#> X        
#> Z        
#> X:Z      
#> Y        
#> x1       
#> x2       
#> x3       
#> z1       
#> z2       
#> z3       
#> y1       
#> y2       
#> y3  2.125
#> 
#> $cor.ov
#>       x1    x2    x3    z1    z2    z3    y1    y2    y3
#> x1 1.000                                                
#> x2 0.826 1.000                                          
#> x3 0.845 0.812 1.000                                    
#> z1 0.172 0.165 0.169 1.000                              
#> z2 0.166 0.160 0.164 0.833 1.000                        
#> z3 0.169 0.162 0.166 0.846 0.820 1.000                  
#> y1 0.448 0.431 0.440 0.407 0.395 0.401 1.000            
#> y2 0.441 0.424 0.434 0.401 0.388 0.395 0.924 1.000      
#> y3 0.444 0.427 0.437 0.404 0.391 0.398 0.931 0.916 1.000
#> 
#> $cor.lv
#>         X     Z   X:Z     Y
#> X   1.000                  
#> Z   0.200 1.000            
#> X:Z 0.000 0.000 1.000      
#> Y   0.499 0.453 0.464 1.000
#> 
#> $cor.all
#>         X     Z   X:Z     Y    x1    x2    x3    z1    z2    z3    y1    y2
#> X   1.000                                                                  
#> Z   0.200 1.000                                                            
#> X:Z 0.000 0.000 1.000                                                      
#> Y   0.499 0.453 0.464 1.000                                                
#> x1  0.927 0.185 0.000 0.462 1.000                                          
#> x2  0.891 0.178 0.000 0.444 0.826 1.000                                    
#> x3  0.912 0.182 0.000 0.455 0.845 0.812 1.000                              
#> z1  0.185 0.927 0.000 0.420 0.172 0.165 0.169 1.000                        
#> z2  0.179 0.898 0.000 0.407 0.166 0.160 0.164 0.833 1.000                  
#> z3  0.182 0.913 0.000 0.414 0.169 0.162 0.166 0.846 0.820 1.000            
#> y1  0.483 0.439 0.450 0.969 0.448 0.431 0.440 0.407 0.395 0.401 1.000      
#> y2  0.476 0.432 0.443 0.954 0.441 0.424 0.434 0.401 0.388 0.395 0.924 1.000
#> y3  0.479 0.436 0.446 0.961 0.444 0.427 0.437 0.404 0.391 0.398 0.931 0.916
#>        y3
#> X        
#> Z        
#> X:Z      
#> Y        
#> x1       
#> x2       
#> x3       
#> z1       
#> z2       
#> z3       
#> y1       
#> y2       
#> y3  1.000
#> 
#> $mean.lv
#>        ~1
#> X   0.000
#> Z   0.000
#> X:Z 0.199
#> Y   0.142
#> 
#> $mean.ov
#>       ~1
#> x1 1.022
#> x2 1.215
#> x3 0.919
#> z1 1.010
#> z2 1.205
#> z3 0.915
#> y1 1.178
#> y2 1.333
#> y3 1.080
#> 
#> $mean.all
#>        ~1
#> X   0.000
#> Z   0.000
#> X:Z 0.199
#> Y   0.142
#> x1  1.022
#> x2  1.215
#> x3  0.919
#> z1  1.010
#> z2  1.205
#> z3  0.915
#> y1  1.178
#> y2  1.333
#> y3  1.080
#> 
#> $r2.all
#>     Y    x1    x2    x3    z1    z2    z3    y1    y2    y3 
#> 0.595 0.858 0.794 0.831 0.859 0.807 0.833 0.938 0.909 0.923 
#> 
#> $r2.lv
#>     Y 
#> 0.595 
#> 
#> $r2.ov
#>    x1    x2    x3    z1    z2    z3    y1    y2    y3 
#> 0.858 0.794 0.831 0.859 0.807 0.833 0.938 0.909 0.923 
#> 
#> $res.all
#>     Y    x1    x2    x3    z1    z2    z3    y1    y2    y3 
#> 0.405 0.142 0.206 0.169 0.141 0.193 0.167 0.062 0.091 0.077 
#> 
#> $res.lv
#>     Y 
#> 0.405 
#> 
#> $res.ov
#>    x1    x2    x3    z1    z2    z3    y1    y2    y3 
#> 0.142 0.206 0.169 0.141 0.193 0.167 0.062 0.091 0.077 
#> 
#>  
# }