Define or disable the color theme used by modsem

All arguments are optional; omitted ones fall back to the defaults below. Pass active = FALSE to turn highlighting off (and reset the palette).

Usage

set_modsem_colors(
  positive = "green3",
  negative = positive,
  true = "green3",
  false = "red",
  nan = "purple",
  na = "purple",
  inf = "purple",
  string = "cyan",
  active = TRUE
)

Arguments

positive

color of positive numbers.

negative

color of negative numbers.

true

color of TRUE.

false

color of FALSE.

nan

color of NaN.

na

color of NA.

inf

color of -Inf and Inf.

string

color of quoted strings.

active

Should color-theme be activated/deactived?

Value

TRUE if colors are active afterwards, otherwise FALSE.

Examples

set_modsem_colors(positive = "red3",
                  negative = "red3",
                  true = "darkgreen",
                  false = "red3",
                  na = "purple",
                  string = "darkgreen")
#> [1] TRUE

m1 <- "
# Outer Model
  X =~ x1 + x2 + x3
  Z =~ z1 + z2 + z3
  Y =~ y1 + y2 + y3
# Inner Model
  Y ~ X + Z + X:Z
"

est <- modsem(m1, data = oneInt)
colorize_output(summary(est))
#> Estimating baseline model (H0)
#> modsem (version 1.0.14, approach = dblcent):
#> 
#> Interaction Model Fit Measures (H1):
#>   Loglikelihood                              -26807.61 
#>   Akaike (AIC)                                53735.22 
#>   Bayesian (BIC)                              54071.28 
#>   Chi-square                                    122.92 
#>   Degrees of Freedom                               111 
#>   P-value (Chi-square)                           0.207 
#>   RMSEA                                          0.007 
#>   CFI                                            1.000 
#>   SRMR                                           0.008 
#> 
#> Fit Measures for Baseline Model (H0):
#>   Loglikelihood                              -27137.74 
#>   Akaike (AIC)                                54393.48 
#>   Bayesian (BIC)                              54723.93 
#>   Chi-square                                    783.18 
#>   Degrees of Freedom                               112 
#>   P-value (Chi-square)                           0.000 
#>   RMSEA                                          0.055 
#>   CFI                                            0.987 
#>   SRMR                                           0.141 
#> 
#> Comparative Fit to H0 (LRT test):
#>   Chi-square diff                              660.257 
#>   Degrees of freedom diff                            1 
#>   P-value (LRT)                                  0.000 
#> 
#> R-Squared Interaction Model (H1):
#>   Y                                              0.602 
#> R-Squared Baseline Model (H0):
#>   Y                                              0.397 
#> R-Squared Change (H1 - H0):
#>   Y                                              0.204 
#> 
#> lavaan 0.6-20 ended normally after 158 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        60
#> 
#>   Number of observations                          2000
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                               122.924
#>   Degrees of freedom                               111
#>   P-value (Chi-square)                           0.207
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   X =~                                                
#>     x1                1.000                           
#>     x2                0.804    0.013   63.612    0.000
#>     x3                0.916    0.014   67.144    0.000
#>   Z =~                                                
#>     z1                1.000                           
#>     z2                0.812    0.013   64.763    0.000
#>     z3                0.882    0.013   67.014    0.000
#>   Y =~                                                
#>     y1                1.000                           
#>     y2                0.798    0.007  107.428    0.000
#>     y3                0.899    0.008  112.453    0.000
#>   XZ =~                                               
#>     x1z1              1.000                           
#>     x2z1              0.805    0.013   60.636    0.000
#>     x3z1              0.877    0.014   62.680    0.000
#>     x1z2              0.793    0.013   59.343    0.000
#>     x2z2              0.646    0.015   43.672    0.000
#>     x3z2              0.706    0.016   44.292    0.000
#>     x1z3              0.887    0.014   63.700    0.000
#>     x2z3              0.716    0.016   45.645    0.000
#>     x3z3              0.781    0.017   45.339    0.000
#> 
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   Y ~                                                 
#>     X                 0.675    0.027   25.379    0.000
#>     Z                 0.561    0.026   21.606    0.000
#>     XZ                0.702    0.027   26.360    0.000
#> 
#> Covariances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>  .x1z1 ~~                                             
#>    .x1z2              0.115    0.008   14.802    0.000
#>    .x1z3              0.114    0.008   13.947    0.000
#>    .x2z1              0.125    0.008   16.095    0.000
#>    .x3z1              0.140    0.009   16.135    0.000
#>  .x1z2 ~~                                             
#>    .x1z3              0.103    0.007   14.675    0.000
#>    .x2z2              0.128    0.006   20.850    0.000
#>    .x3z2              0.146    0.007   21.243    0.000
#>  .x1z3 ~~                                             
#>    .x2z3              0.116    0.007   17.818    0.000
#>    .x3z3              0.135    0.007   18.335    0.000
#>  .x2z1 ~~                                             
#>    .x2z2              0.135    0.006   20.905    0.000
#>    .x2z3              0.145    0.007   21.145    0.000
#>    .x3z1              0.114    0.007   16.058    0.000
#>  .x2z2 ~~                                             
#>    .x2z3              0.117    0.006   20.419    0.000
#>    .x3z2              0.116    0.006   20.586    0.000
#>  .x2z3 ~~                                             
#>    .x3z3              0.109    0.006   18.059    0.000
#>  .x3z1 ~~                                             
#>    .x3z2              0.138    0.007   19.331    0.000
#>    .x3z3              0.158    0.008   20.269    0.000
#>  .x3z2 ~~                                             
#>    .x3z3              0.131    0.007   19.958    0.000
#>   X ~~                                                
#>     Z                 0.201    0.024    8.271    0.000
#>     XZ                0.016    0.025    0.628    0.530
#>   Z ~~                                                
#>     XZ                0.062    0.025    2.449    0.014
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .x1                0.160    0.009   17.871    0.000
#>    .x2                0.162    0.007   22.969    0.000
#>    .x3                0.163    0.008   20.161    0.000
#>    .z1                0.168    0.009   18.143    0.000
#>    .z2                0.158    0.007   22.264    0.000
#>    .z3                0.158    0.008   20.389    0.000
#>    .y1                0.159    0.009   17.896    0.000
#>    .y2                0.154    0.007   22.640    0.000
#>    .y3                0.164    0.008   20.698    0.000
#>    .x1z1              0.311    0.014   22.227    0.000
#>    .x2z1              0.292    0.011   27.287    0.000
#>    .x3z1              0.327    0.012   26.275    0.000
#>    .x1z2              0.290    0.011   26.910    0.000
#>    .x2z2              0.239    0.008   29.770    0.000
#>    .x3z2              0.270    0.009   29.117    0.000
#>    .x1z3              0.272    0.012   23.586    0.000
#>    .x2z3              0.245    0.009   27.979    0.000
#>    .x3z3              0.297    0.011   28.154    0.000
#>     X                 0.981    0.036   26.895    0.000
#>     Z                 1.016    0.038   26.856    0.000
#>    .Y                 0.990    0.038   25.926    0.000
#>     XZ                1.045    0.044   24.004    0.000
#> 
#>  
colorize_output(est) # same as colorize_output(print(est))
#> modsem (version 1.0.14, approach = dblcent):
#> lavaan 0.6-20 ended normally after 158 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        60
#> 
#>   Number of observations                          2000
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                               122.924
#>   Degrees of freedom                               111
#>   P-value (Chi-square)                           0.207
#>  
colorize_output(modsem_inspect(est, what = "coef"))
#> $lambda
#>          X     Z     Y    XZ
#> x1   1.000 0.000 0.000 0.000
#> x2   0.804 0.000 0.000 0.000
#> x3   0.916 0.000 0.000 0.000
#> z1   0.000 1.000 0.000 0.000
#> z2   0.000 0.812 0.000 0.000
#> z3   0.000 0.882 0.000 0.000
#> y1   0.000 0.000 1.000 0.000
#> y2   0.000 0.000 0.798 0.000
#> y3   0.000 0.000 0.899 0.000
#> x1z1 0.000 0.000 0.000 1.000
#> x2z1 0.000 0.000 0.000 0.805
#> x3z1 0.000 0.000 0.000 0.877
#> x1z2 0.000 0.000 0.000 0.793
#> x2z2 0.000 0.000 0.000 0.646
#> x3z2 0.000 0.000 0.000 0.706
#> x1z3 0.000 0.000 0.000 0.887
#> x2z3 0.000 0.000 0.000 0.716
#> x3z3 0.000 0.000 0.000 0.781
#> 
#> $theta
#>         x1    x2    x3    z1    z2    z3    y1    y2    y3  x1z1  x2z1  x3z1
#> x1   0.160                                                                  
#> x2   0.000 0.162                                                            
#> x3   0.000 0.000 0.163                                                      
#> z1   0.000 0.000 0.000 0.168                                                
#> z2   0.000 0.000 0.000 0.000 0.158                                          
#> z3   0.000 0.000 0.000 0.000 0.000 0.158                                    
#> y1   0.000 0.000 0.000 0.000 0.000 0.000 0.159                              
#> y2   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.154                        
#> y3   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.164                  
#> x1z1 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.311            
#> x2z1 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.125 0.292      
#> x3z1 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.140 0.114 0.327
#> x1z2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.115 0.000 0.000
#> x2z2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.135 0.000
#> x3z2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.138
#> x1z3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.114 0.000 0.000
#> x2z3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.145 0.000
#> x3z3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.158
#>       x1z2  x2z2  x3z2  x1z3  x2z3  x3z3
#> x1                                      
#> x2                                      
#> x3                                      
#> z1                                      
#> z2                                      
#> z3                                      
#> y1                                      
#> y2                                      
#> y3                                      
#> x1z1                                    
#> x2z1                                    
#> x3z1                                    
#> x1z2 0.290                              
#> x2z2 0.128 0.239                        
#> x3z2 0.146 0.116 0.270                  
#> x1z3 0.103 0.000 0.000 0.272            
#> x2z3 0.000 0.117 0.000 0.116 0.245      
#> x3z3 0.000 0.000 0.131 0.135 0.109 0.297
#> 
#> $psi
#>        X     Z     Y    XZ
#> X  0.981                  
#> Z  0.201 1.016            
#> Y  0.000 0.000 0.990      
#> XZ 0.016 0.062 0.000 1.045
#> 
#> $beta
#>        X     Z Y    XZ
#> X  0.000 0.000 0 0.000
#> Z  0.000 0.000 0 0.000
#> Y  0.675 0.561 0 0.702
#> XZ 0.000 0.000 0 0.000
#> 
#>  

# \dontrun{
colorize_output(split = TRUE, {
  # Get live (uncolored) output
  # And print colored output at the end of execution

  est_lms <- modsem(m1, data = oneInt, method = "lms")
  summary(est_lms)
})
#> 
#> modsem (1.0.14) ended normally after 16 iterations
#> 
#>   Estimator                                        LMS
#>   Optimization method                       EMA-NLMINB
#>   Number of model parameters                        31
#>                                                       
#>   Number of observations                          2000
#>  
#> Loglikelihood and Information Criteria:
#>   Loglikelihood                              -17493.60
#>   Akaike (AIC)                                35049.21
#>   Bayesian (BIC)                              35222.83
#>  
#> Numerical Integration:
#>   Points of integration (per dim)                   24
#>   Dimensions                                         1
#>   Total points of integration                       24
#>  
#> Fit Measures for Baseline Model (H0):
#>                                               Standard
#>   Chi-square                                     17.52
#>   Degrees of Freedom (Chi-square)                   24
#>   P-value (Chi-square)                           0.826
#>   RMSEA                                          0.000
#>                                                       
#>   Loglikelihood                              -17831.87
#>   Akaike (AIC)                                35723.75
#>   Bayesian (BIC)                              35891.78
#>  
#> Comparative Fit to H0 (LRT test):
#>   Loglikelihood change                          338.27
#>   Difference test (D)                           676.54
#>   Degrees of freedom (D)                             1
#>   P-value (D)                                    0.000
#>  
#> R-Squared Interaction Model (H1):
#>   Y                                              0.598
#> R-Squared Baseline Model (H0):
#>   Y                                              0.395
#> R-Squared Change (H1 - H0):
#>   Y                                              0.203
#> 
#> Parameter Estimates:
#>   Coefficients                          unstandardized
#>   Information                                 observed
#>   Standard errors                             standard
#>  
#> Latent Variables:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>   X =~          
#>     x1              1.000                             
#>     x2              0.803      0.013   63.817    0.000
#>     x3              0.914      0.014   67.615    0.000
#>   Z =~          
#>     z1              1.000                             
#>     z2              0.810      0.012   65.088    0.000
#>     z3              0.881      0.013   67.616    0.000
#>   Y =~          
#>     y1              1.000                             
#>     y2              0.798      0.007  107.548    0.000
#>     y3              0.899      0.008  112.583    0.000
#> 
#> Regressions:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>   Y ~           
#>     X               0.672      0.031   21.650    0.000
#>     Z               0.568      0.030   18.690    0.000
#>     X:Z             0.718      0.028   25.828    0.000
#> 
#> Intercepts:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>    .x1              1.022      0.024   42.752    0.000
#>    .x2              1.215      0.020   60.825    0.000
#>    .x3              0.918      0.022   41.349    0.000
#>    .z1              1.011      0.024   41.533    0.000
#>    .z2              1.205      0.020   59.232    0.000
#>    .z3              0.915      0.022   42.022    0.000
#>    .y1              1.035      0.033   31.358    0.000
#>    .y2              1.219      0.027   45.394    0.000
#>    .y3              0.952      0.030   31.765    0.000
#> 
#> Covariances:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>   X ~~          
#>     Z               0.200      0.024    8.242    0.000
#> 
#> Variances:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>    .x1              0.158      0.009   18.170    0.000
#>    .x2              0.162      0.007   23.160    0.000
#>    .x3              0.164      0.008   20.758    0.000
#>    .z1              0.167      0.009   18.505    0.000
#>    .z2              0.160      0.007   22.680    0.000
#>    .z3              0.158      0.008   20.778    0.000
#>    .y1              0.160      0.009   18.011    0.000
#>    .y2              0.154      0.007   22.687    0.000
#>    .y3              0.164      0.008   20.682    0.000
#>     X               0.981      0.036   26.975    0.000
#>     Z               1.017      0.038   26.935    0.000
#>    .Y               0.980      0.038   25.932    0.000
#> 
#> 
#> modsem (1.0.14) ended normally after 16 iterations
#> 
#>   Estimator                                        LMS
#>   Optimization method                       EMA-NLMINB
#>   Number of model parameters                        31
#>                                                       
#>   Number of observations                          2000
#>  
#> Loglikelihood and Information Criteria:
#>   Loglikelihood                              -17493.60
#>   Akaike (AIC)                                35049.21
#>   Bayesian (BIC)                              35222.83
#>  
#> Numerical Integration:
#>   Points of integration (per dim)                   24
#>   Dimensions                                         1
#>   Total points of integration                       24
#>  
#> Fit Measures for Baseline Model (H0):
#>                                               Standard
#>   Chi-square                                     17.52
#>   Degrees of Freedom (Chi-square)                   24
#>   P-value (Chi-square)                           0.826
#>   RMSEA                                          0.000
#>                                                       
#>   Loglikelihood                              -17831.87
#>   Akaike (AIC)                                35723.75
#>   Bayesian (BIC)                              35891.78
#>  
#> Comparative Fit to H0 (LRT test):
#>   Loglikelihood change                          338.27
#>   Difference test (D)                           676.54
#>   Degrees of freedom (D)                             1
#>   P-value (D)                                    0.000
#>  
#> R-Squared Interaction Model (H1):
#>   Y                                              0.598
#> R-Squared Baseline Model (H0):
#>   Y                                              0.395
#> R-Squared Change (H1 - H0):
#>   Y                                              0.203
#> 
#> Parameter Estimates:
#>   Coefficients                          unstandardized
#>   Information                                 observed
#>   Standard errors                             standard
#>  
#> Latent Variables:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>   X =~          
#>     x1              1.000                             
#>     x2              0.803      0.013   63.817    0.000
#>     x3              0.914      0.014   67.615    0.000
#>   Z =~          
#>     z1              1.000                             
#>     z2              0.810      0.012   65.088    0.000
#>     z3              0.881      0.013   67.616    0.000
#>   Y =~          
#>     y1              1.000                             
#>     y2              0.798      0.007  107.548    0.000
#>     y3              0.899      0.008  112.583    0.000
#> 
#> Regressions:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>   Y ~           
#>     X               0.672      0.031   21.650    0.000
#>     Z               0.568      0.030   18.690    0.000
#>     X:Z             0.718      0.028   25.828    0.000
#> 
#> Intercepts:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>    .x1              1.022      0.024   42.752    0.000
#>    .x2              1.215      0.020   60.825    0.000
#>    .x3              0.918      0.022   41.349    0.000
#>    .z1              1.011      0.024   41.533    0.000
#>    .z2              1.205      0.020   59.232    0.000
#>    .z3              0.915      0.022   42.022    0.000
#>    .y1              1.035      0.033   31.358    0.000
#>    .y2              1.219      0.027   45.394    0.000
#>    .y3              0.952      0.030   31.765    0.000
#> 
#> Covariances:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>   X ~~          
#>     Z               0.200      0.024    8.242    0.000
#> 
#> Variances:
#>                  Estimate  Std.Error  z.value  P(>|z|)
#>    .x1              0.158      0.009   18.170    0.000
#>    .x2              0.162      0.007   23.160    0.000
#>    .x3              0.164      0.008   20.758    0.000
#>    .z1              0.167      0.009   18.505    0.000
#>    .z2              0.160      0.007   22.680    0.000
#>    .z3              0.158      0.008   20.778    0.000
#>    .y1              0.160      0.009   18.011    0.000
#>    .y2              0.154      0.007   22.687    0.000
#>    .y3              0.164      0.008   20.682    0.000
#>     X               0.981      0.036   26.975    0.000
#>     Z               1.017      0.038   26.935    0.000
#>    .Y               0.980      0.038   25.932    0.000
#> 
#>  

colorize_output(modsem_inspect(est_lms))
#> $N
#> [1] 2000
#> 
#> $vcov.all
#>        X=~x2 X=~x3 Z=~z2 Z=~z3 Y=~y2 Y=~y3  x1~1  x2~1  x3~1  z1~1  z2~1  z3~1
#> X=~x2  0.000                                                                  
#> X=~x3  0.000 0.000                                                            
#> Z=~z2  0.000 0.000 0.000                                                      
#> Z=~z3  0.000 0.000 0.000 0.000                                                
#> Y=~y2  0.000 0.000 0.000 0.000 0.000                                          
#> Y=~y3  0.000 0.000 0.000 0.000 0.000 0.000                                    
#> x1~1   0.000 0.000 0.000 0.000 0.000 0.000 0.001                              
#> x2~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000                        
#> x3~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000                  
#> z1~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.001            
#> z2~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000      
#> z3~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y1~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y2~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y3~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> x1~~x1 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> x2~~x2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> x3~~x3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> z1~~z1 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> z2~~z2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> z3~~z3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y1~~y1 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y2~~y2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y3~~y3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> X~~X   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> X~~Z   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Z~~Z   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Y~~Y   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Y~X    0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Y~Z    0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Y~X:Z  0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#>         y1~1  y2~1  y3~1 x1~~x1 x2~~x2 x3~~x3 z1~~z1 z2~~z2 z3~~z3 y1~~y1
#> X=~x2                                                                    
#> X=~x3                                                                    
#> Z=~z2                                                                    
#> Z=~z3                                                                    
#> Y=~y2                                                                    
#> Y=~y3                                                                    
#> x1~1                                                                     
#> x2~1                                                                     
#> x3~1                                                                     
#> z1~1                                                                     
#> z2~1                                                                     
#> z3~1                                                                     
#> y1~1   0.001                                                             
#> y2~1   0.001 0.001                                                       
#> y3~1   0.001 0.001 0.001                                                 
#> x1~~x1 0.000 0.000 0.000  0.000                                          
#> x2~~x2 0.000 0.000 0.000  0.000  0.000                                   
#> x3~~x3 0.000 0.000 0.000  0.000  0.000  0.000                            
#> z1~~z1 0.000 0.000 0.000  0.000  0.000  0.000  0.000                     
#> z2~~z2 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000              
#> z3~~z3 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000       
#> y1~~y1 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> y2~~y2 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> y3~~y3 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> X~~X   0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> X~~Z   0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Z~~Z   0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Y~~Y   0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Y~X    0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Y~Z    0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Y~X:Z  0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#>        y2~~y2 y3~~y3  X~~X  X~~Z  Z~~Z  Y~~Y   Y~X   Y~Z Y~X:Z
#> X=~x2                                                         
#> X=~x3                                                         
#> Z=~z2                                                         
#> Z=~z3                                                         
#> Y=~y2                                                         
#> Y=~y3                                                         
#> x1~1                                                          
#> x2~1                                                          
#> x3~1                                                          
#> z1~1                                                          
#> z2~1                                                          
#> z3~1                                                          
#> y1~1                                                          
#> y2~1                                                          
#> y3~1                                                          
#> x1~~x1                                                        
#> x2~~x2                                                        
#> x3~~x3                                                        
#> z1~~z1                                                        
#> z2~~z2                                                        
#> z3~~z3                                                        
#> y1~~y1                                                        
#> y2~~y2  0.000                                                 
#> y3~~y3  0.000  0.000                                          
#> X~~X    0.000  0.000 0.001                                    
#> X~~Z    0.000  0.000 0.000 0.001                              
#> Z~~Z    0.000  0.000 0.000 0.000 0.001                        
#> Y~~Y    0.000  0.000 0.000 0.000 0.000 0.001                  
#> Y~X     0.000  0.000 0.000 0.000 0.000 0.000 0.001            
#> Y~Z     0.000  0.000 0.000 0.000 0.000 0.000 0.000 0.001      
#> Y~X:Z   0.000  0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.001
#> 
#> $vcov.free
#>        X=~x2 X=~x3 Z=~z2 Z=~z3 Y=~y2 Y=~y3  x1~1  x2~1  x3~1  z1~1  z2~1  z3~1
#> X=~x2  0.000                                                                  
#> X=~x3  0.000 0.000                                                            
#> Z=~z2  0.000 0.000 0.000                                                      
#> Z=~z3  0.000 0.000 0.000 0.000                                                
#> Y=~y2  0.000 0.000 0.000 0.000 0.000                                          
#> Y=~y3  0.000 0.000 0.000 0.000 0.000 0.000                                    
#> x1~1   0.000 0.000 0.000 0.000 0.000 0.000 0.001                              
#> x2~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000                        
#> x3~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000                  
#> z1~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.001            
#> z2~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000      
#> z3~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y1~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y2~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y3~1   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> x1~~x1 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> x2~~x2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> x3~~x3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> z1~~z1 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> z2~~z2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> z3~~z3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y1~~y1 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y2~~y2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> y3~~y3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> X~~X   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> X~~Z   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Z~~Z   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Y~~Y   0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Y~X    0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Y~Z    0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#> Y~X:Z  0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000
#>         y1~1  y2~1  y3~1 x1~~x1 x2~~x2 x3~~x3 z1~~z1 z2~~z2 z3~~z3 y1~~y1
#> X=~x2                                                                    
#> X=~x3                                                                    
#> Z=~z2                                                                    
#> Z=~z3                                                                    
#> Y=~y2                                                                    
#> Y=~y3                                                                    
#> x1~1                                                                     
#> x2~1                                                                     
#> x3~1                                                                     
#> z1~1                                                                     
#> z2~1                                                                     
#> z3~1                                                                     
#> y1~1   0.001                                                             
#> y2~1   0.001 0.001                                                       
#> y3~1   0.001 0.001 0.001                                                 
#> x1~~x1 0.000 0.000 0.000  0.000                                          
#> x2~~x2 0.000 0.000 0.000  0.000  0.000                                   
#> x3~~x3 0.000 0.000 0.000  0.000  0.000  0.000                            
#> z1~~z1 0.000 0.000 0.000  0.000  0.000  0.000  0.000                     
#> z2~~z2 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000              
#> z3~~z3 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000       
#> y1~~y1 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> y2~~y2 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> y3~~y3 0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> X~~X   0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> X~~Z   0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Z~~Z   0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Y~~Y   0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Y~X    0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Y~Z    0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#> Y~X:Z  0.000 0.000 0.000  0.000  0.000  0.000  0.000  0.000  0.000  0.000
#>        y2~~y2 y3~~y3  X~~X  X~~Z  Z~~Z  Y~~Y   Y~X   Y~Z Y~X:Z
#> X=~x2                                                         
#> X=~x3                                                         
#> Z=~z2                                                         
#> Z=~z3                                                         
#> Y=~y2                                                         
#> Y=~y3                                                         
#> x1~1                                                          
#> x2~1                                                          
#> x3~1                                                          
#> z1~1                                                          
#> z2~1                                                          
#> z3~1                                                          
#> y1~1                                                          
#> y2~1                                                          
#> y3~1                                                          
#> x1~~x1                                                        
#> x2~~x2                                                        
#> x3~~x3                                                        
#> z1~~z1                                                        
#> z2~~z2                                                        
#> z3~~z3                                                        
#> y1~~y1                                                        
#> y2~~y2  0.000                                                 
#> y3~~y3  0.000  0.000                                          
#> X~~X    0.000  0.000 0.000                                    
#> X~~Z    0.000  0.000 0.000 0.001                              
#> Z~~Z    0.000  0.000 0.000 0.000 0.000                        
#> Y~~Y    0.000  0.000 0.000 0.000 0.000 0.001                  
#> Y~X     0.000  0.000 0.000 0.000 0.000 0.000 0.001            
#> Y~Z     0.000  0.000 0.000 0.000 0.000 0.000 0.000 0.001      
#> Y~X:Z   0.000  0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.001
#> 
#> $information
#>            X=~x2     X=~x3     Z=~z2     Z=~z3     Y=~y2     Y=~y3      x1~1
#> X=~x2   8941.557                                                            
#> X=~x3  -2765.565  8146.575                                                  
#> Z=~z2     63.762    74.728  9240.182                                        
#> Z=~z3     72.757    81.968 -3033.092  8777.388                              
#> Y=~y2   -169.977  -194.596  -158.987  -172.450 23811.626                    
#> Y=~y3   -177.016  -205.778  -170.036  -183.316 -8383.417 21007.771          
#> x1~1      29.515    30.611    92.210    97.653  -122.137  -132.626  8012.626
#> x2~1      35.932    23.990    72.266    76.532   -95.720  -103.941 -3634.713
#> x3~1      25.929    39.508    81.009    85.790  -107.300  -116.515 -4074.443
#> z1~1      90.696    98.955    22.576    27.134  -126.628  -133.510    53.666
#> z2~1      76.777    83.768    25.691    22.978  -107.195  -113.020    45.430
#> z3~1      84.227    91.897    20.822    32.245  -117.597  -123.988    49.838
#> y1~1     -23.955   -27.010   -24.174   -28.696  -597.341  -635.313  -175.994
#> y2~1     -19.770   -22.291   -19.951   -23.683  1386.778  -524.353  -145.248
#> y3~1     -21.022   -23.703   -21.215   -25.183  -524.411  1217.328  -154.448
#> x1~~x1 -1396.172 -1483.521    17.293    27.838   -10.518   -11.357   -10.652
#> x2~~x2  2711.936  -914.338    25.917    15.856    -6.362   -14.559   -13.818
#> x3~~x3 -1033.516  2732.963    14.496    17.648    14.538    22.525    -3.615
#> z1~~z1    12.157    13.080 -1311.440 -1440.761    14.589    13.785    38.901
#> z2~~z2    33.224    27.807  2848.729 -1039.845    -2.987     0.230    12.449
#> z3~~z3    24.348    33.527 -1122.527  2976.240   -14.459   -17.574    34.502
#> y1~~y1     1.243     0.793     7.222     5.924 -1490.681 -1623.709     2.761
#> y2~~y2     3.531    -5.621   -12.342    -9.434  3046.498 -1169.239   -11.132
#> y3~~y3   -16.508   -12.652    -8.495   -10.897 -1249.382  2844.100     2.552
#> X~~X    1137.425  1267.806   -51.024   -57.981    62.962    66.074    -7.882
#> X~~Z    -118.655  -133.049   111.992   130.683    32.165    34.613   -51.608
#> Z~~Z     -25.936   -29.600  1157.869  1268.248    48.933    52.027   -17.095
#> Y~~Y     -30.000   -43.101   -32.848   -35.161   550.940   591.073   -13.199
#> Y~X     -290.121  -314.920   -40.246   -49.027   435.398   458.864   -72.179
#> Y~Z      -64.496   -69.495  -275.948  -285.861   396.782   432.159  -418.710
#> Y~X:Z   -292.094  -336.978  -317.981  -348.693   419.220   438.918   -77.387
#>             x2~1      x3~1      z1~1      z2~1      z3~1      y1~1      y2~1
#> X=~x2                                                                       
#> X=~x3                                                                       
#> Z=~z2                                                                       
#> Z=~z3                                                                       
#> Y=~y2                                                                       
#> Y=~y3                                                                       
#> x1~1                                                                        
#> x2~1    9491.163                                                            
#> x3~1   -3193.179  8580.908                                                  
#> z1~1      42.058    47.146  7693.672                                        
#> z2~1      35.604    39.911 -3641.256  9448.834                              
#> z3~1      39.058    43.784 -3994.584 -3381.545  8933.233                    
#> y1~1    -137.928  -154.615  -144.356  -122.202  -134.060  7704.226          
#> y2~1    -113.832  -127.604  -119.137  -100.853  -110.640 -3985.348  9668.386
#> y3~1    -121.043  -135.686  -126.683  -107.242  -117.648 -4237.792 -3497.452
#> x1~~x1    13.092    14.675    49.593    41.982    46.055    -6.783    -5.598
#> x2~~x2   -42.098   -12.140    13.141    11.124    12.204     3.964     3.271
#> x3~~x3    -2.833   -25.580    31.592    26.744    29.339     5.854     4.831
#> z1~~z1    30.487    34.175   -18.031    -9.649   -10.046    -1.821    -1.503
#> z2~~z2     9.757    10.937     2.844   -13.593     3.006    -2.267    -1.871
#> z3~~z3    27.040    30.311    -9.792    -8.313   -18.263     9.398     7.757
#> y1~~y1     2.164     2.425    -4.812    -4.073    -4.467   -10.780    -4.265
#> y2~~y2    -8.724    -9.779     2.605     2.205     2.420    -4.318   -19.498
#> y3~~y3     2.000     2.242    -3.263    -2.762    -3.029    -3.881    -3.123
#> X~~X      -6.178    -6.925   -31.322   -26.514   -29.099    -2.217    -1.829
#> X~~Z     -40.445   -45.338   -38.575   -32.655   -35.824    55.780    46.035
#> Z~~Z     -13.397   -15.018     7.934     6.715     7.397   -13.142   -10.846
#> Y~~Y     -10.344   -11.595   -13.794   -11.678   -12.801    -1.112    -0.916
#> Y~X      -56.568   -63.411  -399.221  -337.953  -370.747   -33.777   -27.877
#> Y~Z     -328.147  -367.846   -71.459   -60.492   -66.376   -37.056   -30.582
#> Y~X:Z    -60.649   -67.986   -51.627   -43.703   -47.961   129.411   106.803
#>             y3~1    x1~~x1    x2~~x2    x3~~x3    z1~~z1    z2~~z2    z3~~z3
#> X=~x2                                                                       
#> X=~x3                                                                       
#> Z=~z2                                                                       
#> Z=~z3                                                                       
#> Y=~y2                                                                       
#> Y=~y3                                                                       
#> x1~1                                                                        
#> x2~1                                                                        
#> x3~1                                                                        
#> z1~1                                                                        
#> z2~1                                                                        
#> z3~1                                                                        
#> y1~1                                                                        
#> y2~1                                                                        
#> y3~1    8513.069                                                            
#> x1~~x1    -5.953 16046.235                                                  
#> x2~~x2     3.479  3562.272 22501.454                                        
#> x3~~x3     5.137  4023.126  2496.422 18384.334                              
#> z1~~z1    -1.596     2.214    14.399    21.140 14803.658                    
#> z2~~z2    -1.989    51.943   -22.975    23.568  3284.905 22325.613          
#> z3~~z3     8.249    13.123    42.398    18.640  3963.828  2882.810 19935.245
#> y1~~y1    -3.895    20.118    10.857    16.981    18.868    20.465    35.752
#> y2~~y2    -3.278     6.802   -18.912    29.468     3.031    10.420   -10.326
#> y3~~y3   -11.265     8.638    21.475    -5.541    12.899    -6.965     0.405
#> X~~X      -1.944   235.669   140.954   217.284   -10.141   -15.301    -7.097
#> X~~Z      48.951   -23.280   -50.189   -62.849    12.802    13.370   -27.530
#> Z~~Z     -11.533   -13.487     4.567     8.662   264.485   188.050   236.721
#> Y~~Y      -0.958    87.017    35.662    97.443    85.693    57.281    64.060
#> Y~X      -29.642  -170.841   -57.264  -129.191    33.324   -16.986     6.728
#> Y~Z      -32.518   -22.964    16.033     4.563  -137.925   -41.688  -143.932
#> Y~X:Z    113.569  -105.379  -101.057   -73.944  -101.688  -117.098  -141.540
#>           y1~~y1    y2~~y2    y3~~y3      X~~X      X~~Z      Z~~Z      Y~~Y
#> X=~x2                                                                       
#> X=~x3                                                                       
#> Z=~z2                                                                       
#> Z=~z3                                                                       
#> Y=~y2                                                                       
#> Y=~y3                                                                       
#> x1~1                                                                        
#> x2~1                                                                        
#> x3~1                                                                        
#> z1~1                                                                        
#> z2~1                                                                        
#> z3~1                                                                        
#> y1~1                                                                        
#> y2~1                                                                        
#> y3~1                                                                        
#> x1~~x1                                                                      
#> x2~~x2                                                                      
#> x3~~x3                                                                      
#> z1~~z1                                                                      
#> z2~~z2                                                                      
#> z3~~z3                                                                      
#> y1~~y1 14844.964                                                            
#> y2~~y2  3727.680 23380.841                                                  
#> y3~~y3  4422.001  3348.001 18127.815                                        
#> X~~X      -3.245     0.709     5.325  3681.191                              
#> X~~Z       1.129     0.388    -1.829  -317.072  1821.606                    
#> Z~~Z      -4.992     4.027     4.555     0.762    44.774  3590.300          
#> Y~~Y     170.467    79.557    60.784     7.080    -0.904     8.741   760.060
#> Y~X       18.440    -9.466    11.719   121.090    36.303    -2.285    52.116
#> Y~Z      -13.145    42.544    -8.431     7.417    59.738    90.385    47.402
#> Y~X:Z    -25.812    15.037    54.488   121.813    54.481   111.367   109.073
#>              Y~X       Y~Z     Y~X:Z
#> X=~x2                               
#> X=~x3                               
#> Z=~z2                               
#> Z=~z3                               
#> Y=~y2                               
#> Y=~y3                               
#> x1~1                                
#> x2~1                                
#> x3~1                                
#> z1~1                                
#> z2~1                                
#> z3~1                                
#> y1~1                                
#> y2~1                                
#> y3~1                                
#> x1~~x1                              
#> x2~~x2                              
#> x3~~x3                              
#> z1~~z1                              
#> z2~~z2                              
#> z3~~z3                              
#> y1~~y1                              
#> y2~~y2                              
#> y3~~y3                              
#> X~~X                                
#> X~~Z                                
#> Z~~Z                                
#> Y~~Y                                
#> Y~X     1549.283                    
#> Y~Z      320.231  1596.397          
#> Y~X:Z    -61.366   -25.035  1511.460
#> 
#> $coefficients.all
#>  X=~x2  X=~x3  Z=~z2  Z=~z3  Y=~y2  Y=~y3   x1~1   x2~1   x3~1   z1~1   z2~1 
#>  0.803  0.914  0.810  0.881  0.798  0.899  1.022  1.215  0.918  1.011  1.205 
#>   z3~1   y1~1   y2~1   y3~1 x1~~x1 x2~~x2 x3~~x3 z1~~z1 z2~~z2 z3~~z3 y1~~y1 
#>  0.915  1.035  1.219  0.952  0.158  0.162  0.164  0.167  0.160  0.158  0.160 
#> y2~~y2 y3~~y3   X~~X   X~~Z   Z~~Z   Y~~Y    Y~X    Y~Z  Y~X:Z 
#>  0.154  0.164  0.981  0.200  1.017  0.980  0.672  0.568  0.718 
#> 
#> $coefficients.free
#>  X=~x2  X=~x3  Z=~z2  Z=~z3  Y=~y2  Y=~y3   x1~1   x2~1   x3~1   z1~1   z2~1 
#>  0.803  0.914  0.810  0.881  0.798  0.899  1.022  1.215  0.918  1.011  1.205 
#>   z3~1   y1~1   y2~1   y3~1 x1~~x1 x2~~x2 x3~~x3 z1~~z1 z2~~z2 z3~~z3 y1~~y1 
#>  0.915  1.035  1.219  0.952  0.158  0.162  0.164  0.167  0.160  0.158  0.160 
#> y2~~y2 y3~~y3   X~~X   X~~Z   Z~~Z   Y~~Y    Y~X    Y~Z  Y~X:Z 
#>  0.154  0.164  0.981  0.200  1.017  0.980  0.672  0.568  0.718 
#> 
#> $partable
#>    lhs op rhs label   est std.error z.value p.value ci.lower ci.upper
#> 1    X =~  x1       1.000        NA      NA      NA       NA       NA
#> 2    X =~  x2       0.803     0.013  63.817       0    0.779    0.828
#> 3    X =~  x3       0.914     0.014  67.615       0    0.887    0.940
#> 4    Z =~  z1       1.000        NA      NA      NA       NA       NA
#> 5    Z =~  z2       0.810     0.012  65.088       0    0.786    0.835
#> 6    Z =~  z3       0.881     0.013  67.616       0    0.856    0.907
#> 7    Y =~  y1       1.000        NA      NA      NA       NA       NA
#> 8    Y =~  y2       0.798     0.007 107.548       0    0.784    0.813
#> 9    Y =~  y3       0.899     0.008 112.583       0    0.884    0.915
#> 10   Y  ~   X       0.672     0.031  21.650       0    0.611    0.733
#> 11   Y  ~   Z       0.568     0.030  18.690       0    0.508    0.628
#> 12   Y  ~ X:Z       0.718     0.028  25.828       0    0.664    0.773
#> 13  x1 ~1           1.022     0.024  42.752       0    0.975    1.069
#> 14  x2 ~1           1.215     0.020  60.825       0    1.175    1.254
#> 15  x3 ~1           0.918     0.022  41.349       0    0.875    0.962
#> 16  z1 ~1           1.011     0.024  41.533       0    0.963    1.058
#> 17  z2 ~1           1.205     0.020  59.232       0    1.165    1.245
#> 18  z3 ~1           0.915     0.022  42.022       0    0.872    0.958
#> 19  y1 ~1           1.035     0.033  31.358       0    0.970    1.100
#> 20  y2 ~1           1.219     0.027  45.394       0    1.166    1.271
#> 21  y3 ~1           0.952     0.030  31.765       0    0.893    1.011
#> 22  x1 ~~  x1       0.158     0.009  18.170       0    0.141    0.175
#> 23  x2 ~~  x2       0.162     0.007  23.160       0    0.148    0.176
#> 24  x3 ~~  x3       0.164     0.008  20.758       0    0.149    0.180
#> 25  z1 ~~  z1       0.167     0.009  18.505       0    0.149    0.184
#> 26  z2 ~~  z2       0.160     0.007  22.680       0    0.146    0.173
#> 27  z3 ~~  z3       0.158     0.008  20.778       0    0.143    0.173
#> 28  y1 ~~  y1       0.160     0.009  18.011       0    0.142    0.177
#> 29  y2 ~~  y2       0.154     0.007  22.687       0    0.141    0.168
#> 30  y3 ~~  y3       0.164     0.008  20.682       0    0.148    0.179
#> 31   X ~~   X       0.981     0.036  26.975       0    0.910    1.052
#> 32   X ~~   Z       0.200     0.024   8.242       0    0.152    0.247
#> 33   Z ~~   Z       1.017     0.038  26.935       0    0.943    1.091
#> 34   Y ~~   Y       0.980     0.038  25.932       0    0.906    1.054
#> 
#> $partable.input
#>    lhs op rhs mod
#> 1    X =~  x1    
#> 2    X =~  x2    
#> 3    X =~  x3    
#> 4    Z =~  z1    
#> 5    Z =~  z2    
#> 6    Z =~  z3    
#> 7    Y =~  y1    
#> 8    Y =~  y2    
#> 9    Y =~  y3    
#> 10   Y  ~   X    
#> 11   Y  ~   Z    
#> 12   Y  ~ X:Z    
#> 
#> $loglik
#> [1] -17493.6
#> 
#> $iterations
#> [1] 16
#> 
#> $convergence
#> [1] TRUE
#> 
#> $lambda
#>        X     Z     Y
#> x1 1.000 0.000 0.000
#> x2 0.803 0.000 0.000
#> x3 0.914 0.000 0.000
#> z1 0.000 1.000 0.000
#> z2 0.000 0.810 0.000
#> z3 0.000 0.881 0.000
#> y1 0.000 0.000 1.000
#> y2 0.000 0.000 0.798
#> y3 0.000 0.000 0.899
#> 
#> $tau
#>       ~1
#> x1 1.022
#> x2 1.215
#> x3 0.918
#> z1 1.011
#> z2 1.205
#> z3 0.915
#> y1 1.035
#> y2 1.219
#> y3 0.952
#> 
#> $theta
#>       x1    x2    x3    z1    z2    z3    y1    y2    y3
#> x1 0.158                                                
#> x2 0.000 0.162                                          
#> x3 0.000 0.000 0.164                                    
#> z1 0.000 0.000 0.000 0.167                              
#> z2 0.000 0.000 0.000 0.000 0.160                        
#> z3 0.000 0.000 0.000 0.000 0.000 0.158                  
#> y1 0.000 0.000 0.000 0.000 0.000 0.000 0.160            
#> y2 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.154      
#> y3 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.000 0.164
#> 
#> $gamma.xi
#>       X     Z
#> Y 0.672 0.568
#> 
#> $gamma.eta
#>       Y
#> Y 0.000
#> 
#> $omega.xi.xi
#>         X     Z
#> Y~X 0.000 0.718
#> Y~Z 0.000 0.000
#> 
#> $omega.eta.xi
#>         Y
#> Y~X 0.000
#> Y~Z 0.000
#> 
#> $phi
#>       X     Z
#> X 0.981      
#> Z 0.200 1.017
#> 
#> $psi
#>       Y
#> Y 0.980
#> 
#> $alpha
#>      ~1
#> Y 0.000
#> 
#> $beta0
#>      ~1
#> X 0.000
#> Z 0.000
#> 
#> $cov.ov
#>       x1    x2    x3    z1    z2    z3    y1    y2    y3
#> x1 1.139                                                
#> x2 0.788 0.795                                          
#> x3 0.897 0.720 0.984                                    
#> z1 0.200 0.161 0.183 1.184                              
#> z2 0.162 0.130 0.148 0.824 0.828                        
#> z3 0.176 0.142 0.161 0.896 0.726 0.948                  
#> y1 0.773 0.621 0.706 0.712 0.577 0.628 2.599            
#> y2 0.617 0.496 0.564 0.569 0.461 0.501 1.947 1.709      
#> y3 0.695 0.558 0.635 0.640 0.519 0.564 2.194 1.751 2.136
#> 
#> $cov.lv
#>         X     Z   X:Z     Y
#> X   0.981                  
#> Z   0.200 1.017            
#> X:Z 0.000 0.000 1.038      
#> Y   0.773 0.712 0.745 2.439
#> 
#> $cov.all
#>         X     Z   X:Z     Y    x1    x2    x3    z1    z2    z3    y1    y2
#> X   0.981                                                                  
#> Z   0.200 1.017                                                            
#> X:Z 0.000 0.000 1.038                                                      
#> Y   0.773 0.712 0.745 2.439                                                
#> x1  0.981 0.200 0.000 0.773 1.139                                          
#> x2  0.788 0.161 0.000 0.621 0.788 0.795                                    
#> x3  0.897 0.183 0.000 0.706 0.897 0.720 0.984                              
#> z1  0.200 1.017 0.000 0.712 0.200 0.161 0.183 1.184                        
#> z2  0.162 0.824 0.000 0.577 0.162 0.130 0.148 0.824 0.828                  
#> z3  0.176 0.896 0.000 0.628 0.176 0.142 0.161 0.896 0.726 0.948            
#> y1  0.773 0.712 0.745 2.439 0.773 0.621 0.706 0.712 0.577 0.628 2.599      
#> y2  0.617 0.569 0.595 1.947 0.617 0.496 0.564 0.569 0.461 0.501 1.947 1.709
#> y3  0.695 0.640 0.670 2.194 0.695 0.558 0.635 0.640 0.519 0.564 2.194 1.751
#>        y3
#> X        
#> Z        
#> X:Z      
#> Y        
#> x1       
#> x2       
#> x3       
#> z1       
#> z2       
#> z3       
#> y1       
#> y2       
#> y3  2.136
#> 
#> $cor.ov
#>       x1    x2    x3    z1    z2    z3    y1    y2    y3
#> x1 1.000                                                
#> x2 0.828 1.000                                          
#> x3 0.847 0.814 1.000                                    
#> z1 0.172 0.166 0.169 1.000                              
#> z2 0.167 0.160 0.164 0.833 1.000                        
#> z3 0.170 0.163 0.167 0.846 0.820 1.000                  
#> y1 0.449 0.432 0.442 0.406 0.394 0.400 1.000            
#> y2 0.442 0.425 0.435 0.400 0.387 0.394 0.924 1.000      
#> y3 0.446 0.428 0.438 0.403 0.390 0.397 0.931 0.917 1.000
#> 
#> $cor.lv
#>         X     Z   X:Z     Y
#> X   1.000                  
#> Z   0.200 1.000            
#> X:Z 0.000 0.000 1.000      
#> Y   0.500 0.452 0.468 1.000
#> 
#> $cor.all
#>         X     Z   X:Z     Y    x1    x2    x3    z1    z2    z3    y1    y2
#> X   1.000                                                                  
#> Z   0.200 1.000                                                            
#> X:Z 0.000 0.000 1.000                                                      
#> Y   0.500 0.452 0.468 1.000                                                
#> x1  0.928 0.186 0.000 0.464 1.000                                          
#> x2  0.892 0.179 0.000 0.446 0.828 1.000                                    
#> x3  0.913 0.183 0.000 0.456 0.847 0.814 1.000                              
#> z1  0.185 0.927 0.000 0.419 0.172 0.166 0.169 1.000                        
#> z2  0.180 0.898 0.000 0.406 0.167 0.160 0.164 0.833 1.000                  
#> z3  0.183 0.913 0.000 0.413 0.170 0.163 0.167 0.846 0.820 1.000            
#> y1  0.484 0.438 0.454 0.969 0.449 0.432 0.442 0.406 0.394 0.400 1.000      
#> y2  0.477 0.431 0.447 0.954 0.442 0.425 0.435 0.400 0.387 0.394 0.924 1.000
#> y3  0.480 0.434 0.450 0.961 0.446 0.428 0.438 0.403 0.390 0.397 0.931 0.917
#>        y3
#> X        
#> Z        
#> X:Z      
#> Y        
#> x1       
#> x2       
#> x3       
#> z1       
#> z2       
#> z3       
#> y1       
#> y2       
#> y3  1.000
#> 
#> $mean.lv
#>        ~1
#> X   0.000
#> Z   0.000
#> X:Z 0.200
#> Y   0.144
#> 
#> $mean.ov
#>       ~1
#> x1 1.022
#> x2 1.215
#> x3 0.918
#> z1 1.011
#> z2 1.205
#> z3 0.915
#> y1 1.178
#> y2 1.333
#> y3 1.081
#> 
#> $mean.all
#>        ~1
#> X   0.000
#> Z   0.000
#> X:Z 0.200
#> Y   0.144
#> x1  1.022
#> x2  1.215
#> x3  0.918
#> z1  1.011
#> z2  1.205
#> z3  0.915
#> y1  1.178
#> y2  1.333
#> y3  1.081
#> 
#> $r2.all
#>     Y    x1    x2    x3    z1    z2    z3    y1    y2    y3 
#> 0.598 0.861 0.796 0.833 0.859 0.807 0.833 0.939 0.910 0.923 
#> 
#> $r2.lv
#>     Y 
#> 0.598 
#> 
#> $r2.ov
#>    x1    x2    x3    z1    z2    z3    y1    y2    y3 
#> 0.861 0.796 0.833 0.859 0.807 0.833 0.939 0.910 0.923 
#> 
#> $res.all
#>     Y    x1    x2    x3    z1    z2    z3    y1    y2    y3 
#> 0.402 0.139 0.204 0.167 0.141 0.193 0.167 0.061 0.090 0.077 
#> 
#> $res.lv
#>     Y 
#> 0.402 
#> 
#> $res.ov
#>    x1    x2    x3    z1    z2    z3    y1    y2    y3 
#> 0.139 0.204 0.167 0.141 0.193 0.167 0.061 0.090 0.077 
#> 
#>  
# }